% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_abstract_prediction_model.R
\name{abstract_prediction_model}
\alias{abstract_prediction_model}
\title{An abstract data wrapper for TI predictions}
\usage{
abstract_prediction_model(cell_ids, cell_info = NULL, ...)
}
\arguments{
\item{cell_ids}{The ids of the cells.}

\item{cell_info}{Optional meta-information pertaining the cells.}

\item{...}{Extra information to be stored in the wrapper.}
}
\description{
An abstract data wrapper for TI predictions
}
