% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_container_ti_method.R
\name{create_container_ti_method}
\alias{create_container_ti_method}
\title{Creating a TI method from a docker repository}
\usage{
create_container_ti_method(docker_repository, run_environment = NULL, ...)
}
\arguments{
\item{docker_repository}{The id of the docker repository.}

\item{run_environment}{In which environment to run the method, can be \code{"docker"} or \code{"singularity"}.}

\item{...}{Default parameters for the method.}
}
\description{
Supports both docker and singularity as a backend.
}
