% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_docker.R
\name{create_docker_ti_method}
\alias{create_docker_ti_method}
\title{Create a TI method from a docker(hub) container}
\usage{
create_docker_ti_method(container, name = NULL, input = NULL,
  output = NULL, par_set = NULL, parameters = NULL, definition = NULL,
  pull = TRUE, docker_client = stevedore::docker_client())
}
\arguments{
\item{container}{The name of the container. Required}

\item{name}{The name of the method. Optional, will be extracted from the container definition.yml file}

\item{input}{Input data, such as expression, counts or prior_information}

\item{output}{Output generated by the method}

\item{par_set}{The parameter set as created by \link[ParamHelpers:makeParamSet]{ParamHelpers::makeParamSet}}

\item{parameters}{A named list of parameters, containing the type of parameter and additional parameters such as lower and upper bounds, default value and/or possible values. Optional, will be extracted from the container definition.yml file}

\item{definition}{A list containing the name, container, input, output and parameters of a method. Usually used for when these values are loaded from a yaml or json file}

\item{pull}{Whether to pull the container first}

\item{docker_client}{Optional, a \code{\link[stevedore:docker_client]{stevedore::docker_client()}}}
}
\description{
Create a TI method from a docker(hub) container
}
