% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_create_ti_method.R
\name{create_ti_method}
\alias{create_ti_method}
\title{Create a TI method wrapper}
\usage{
create_ti_method(name, short_name, package_loaded, package_required, par_set,
  run_fun, plot_fun)
}
\arguments{
\item{name}{The name of the TI method}

\item{short_name}{A short name for the method, max 8 characters}

\item{package_loaded}{The packages that need to be loaded before executing the method}

\item{package_required}{The packages that need to be installed before executing the method}

\item{par_set}{A bunch of parameters created by \code{\link[ParamHelpers:makeParamSet]{ParamHelpers::makeParamSet()}}}

\item{run_fun}{A function to run the TI, needs to have 'counts' as its first param.}

\item{plot_fun}{A function to plot the results of a TI, needs to have 'prediction' as its first param.
of \code{run_fun} with those described in \code{par_set}.}
}
\description{
Create a TI method wrapper
}
