% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_angle.R
\name{description_angle}
\alias{description_angle}
\title{Inferring trajectories with Angle}
\usage{
description_angle(dimred = "pca")
}
\arguments{
\item{dimred}{A character vector specifying which dimensionality reduction method to use.
See \code{\link{list_dimred_methods}} for the list of available dimensionality reduction methods.}
}
\description{
Arguments passed to this function will be used as default parameters for the method.
}
