% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_error.R
\name{description_error}
\alias{description_error}
\title{Inferring trajectories with Control: error}
\usage{
description_error(dummy_param = 0.5)
}
\arguments{
\item{dummy_param}{This parameter does not do anything.}
}
\description{
This control method will always produce an error.
}
