% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_matcher.R
\name{description_matcher}
\alias{description_matcher}
\title{Inferring trajectories with MATCHER}
\usage{
description_matcher(quantiles = 50L, method = "linear")
}
\description{
Arguments passed to this function will be used as default parameters for the method.
}
