% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_mfa.R
\name{description_mfa}
\alias{description_mfa}
\title{Inferring trajectories with mfa}
\usage{
description_mfa(b = 2L, iter = 2000L, thin = 1L, pc_initialise = 1L,
  prop_collapse = 0, scale_input = TRUE, zero_inflation = FALSE)
}
\description{
Arguments passed to this function will be used as default parameters for the method.
}
