% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_ouija.R
\name{description_ouija}
\alias{description_ouija}
\title{Inferring trajectories with ouija}
\usage{
description_ouija(iter = 100, response_type = "switch",
  inference_type = "hmc", normalise_expression = TRUE)
}
\description{
Arguments passed to this function will be used as default parameters for the method.
}
