% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_slicer.R
\name{description_slicer}
\alias{description_slicer}
\title{Inferring trajectories with SLICER}
\usage{
description_slicer(kmin = 10L, m = 2L, min_branch_len = 5,
  min_representative_percentage = 0.8, max_same_milestone_distance = 0.1)
}
\description{
Arguments passed to this function will be used as default parameters for the method.
}
