% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynmethods-package.R
\docType{package}
\name{dynmethods}
\alias{dynmethods}
\alias{dynmethods-package}
\title{Wrappers for trajectory inference}
\description{
Wrappers for trajectory inference
}
\author{
\strong{Maintainer}: Robrecht Cannoodt \email{rcannood@gmail.com}

Authors:
\itemize{
  \item Wouter Saelens \email{wouter.saelens@ugent.be}
  \item Helena Todorov \email{helena.todorov@ugent.be}
}

}
