% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_execute_method.R
\name{execute_method}
\alias{execute_method}
\title{Run a method on a set of tasks with a set of parameters}
\usage{
execute_method(tasks, method, parameters, give_priors = NULL, mc_cores = 1)
}
\arguments{
\item{tasks}{The tasks on which to evaluate.}

\item{method}{The method to evaluate.}

\item{parameters}{The parameters to evaluate with.}

\item{give_priors}{All the priors a method is allowed to receive. Must be a subset of: \code{"start_milestones"},
\code{"start_cells"}, \code{"end_milestones"}, \code{"end_cells"}, \code{"grouping_assignment"} and \code{"grouping_network"}}

\item{mc_cores}{The number of cores to use, allowing to parallellise the different tasks}
}
\description{
Run a method on a set of tasks with a set of parameters
}
