% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_infer_trajectory.R
\name{execute_method_internal}
\alias{execute_method_internal}
\title{Internal method for executing a method}
\usage{
execute_method_internal(method, arglist, setseed_detection_file)
}
\arguments{
\item{method}{One or more methods. Can be a description as created by the ti_... functions, a character vector containing the methods to execute, or a dynguidelines dataframe}

\item{arglist}{The arguments to apply to the method}

\item{setseed_detection_file}{A file to which will be written if a method
uses the set.seed function.}
}
\description{
If you're reading this, you're supposed to be using \code{infer_trajectory} instead.
}
