% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{method_versions}
\alias{method_versions}
\title{Version numbers of the method repos}
\format{An object of class \code{character} of length 59.}
\usage{
method_versions
}
\description{
Version numbers of the method repos
}
\keyword{datasets}
