% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{methods_containerised}
\alias{methods_containerised}
\title{Metadata on the different containerised TI methods}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 44 rows and 5 columns.}
\usage{
methods_containerised
}
\description{
Metadata on the different containerised TI methods
}
\keyword{datasets}
