% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{methods_info}
\alias{methods_info}
\title{Metadata on the all TI methods, even those not wrapped within dynmethods}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 80 rows and 48 columns.}
\usage{
methods_info
}
\description{
Metadata on the all TI methods, even those not wrapped within dynmethods
}
\keyword{datasets}
