% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_monocle.R
\name{ti_monocle_ddrtree}
\alias{ti_monocle_ddrtree}
\alias{ti_monocle_ica}
\title{Inferring trajectories with Monocle}
\usage{
ti_monocle_ddrtree(reduction_method = "DDRTree", max_components = 2L,
  norm_method = "vstExprs", auto_param_selection = TRUE)

ti_monocle_ica(reduction_method = "ICA", max_components = 2L,
  norm_method = "vstExprs")
}
\description{
Inferring trajectories with Monocle
}
