% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{repo_digests}
\alias{repo_digests}
\title{Digests of the method repos}
\format{An object of class \code{character} of length 57.}
\usage{
repo_digests
}
\description{
Digests of the method repos
}
\keyword{datasets}
