% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_angle.R
\name{run_angle}
\alias{run_angle}
\title{Infer trajectory with \code{angle}}
\usage{
run_angle(expression, dimred)
}
\arguments{
\item{expression}{An M-by-N expression matrix, with M the number of cells and N the number of genes.}

\item{dimred}{A character vector specifying which dimensionality reduction method to use.
See \code{\link{list_dimred_methods}} for the list of available dimensionality reduction methods.}
}
\description{
Infer trajectory with \code{angle}
}
