% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_calista.R
\name{ti_calista}
\alias{ti_calista}
\title{CALISTA}
\usage{
ti_calista(runs = 50L, max_iter = 100L)
}
\arguments{
\item{runs}{Number of independent runs of greedy algorithm. Domain: U(20, 100).
Default: 50. Format: integer.}

\item{max_iter}{Number of iterations in greedy algorithm. Domain: U(20, 400).
Default: 100. Format: integer.}
}
\value{
A TI method wrapper to be used together with
\code{\link[dynwrap:infer_trajectories]{infer_trajectory}}
}
\description{
Will generate a trajectory using \href{https://doi.org/10.1101/257550}{CALISTA}.

This method was wrapped inside a
\href{https://github.com/dynverse/ti_calista}{container}.
The original code of this method is available
\href{https://github.com/CABSEL/CALISTA}{here}.
}
\references{
Gao, N.P., Hartmann, T., Fang, T., Gunawan, R., 2018. CALISTA:
Clustering and Lineage Inference in Single-Cell Transcriptional Analysis.
}
\keyword{method}
