% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_cellrouter.R
\name{ti_cellrouter}
\alias{ti_cellrouter}
\title{Inferring a trajectory inference using CellRouter}
\usage{
ti_cellrouter(ndim_pca = 20L, ndim_tsne = 11L, max_iter = 1000L,
  cluster_method = "graph.clustering", k_clustering = 20L,
  ndim_pca_clustering = 20L, k_knn = 10L, ndim_pca_knn = 20L,
  sim_type = "jaccard", distance_method_paths = "graph",
  ranks = "rank", num_cells = 3L, neighs = 3L, perplexity = 30L)
}
\arguments{
\item{ndim_pca}{Number of principal components to compute. Domain: U(2, 100).
Default: 20. Format: integer.}

\item{ndim_tsne}{Number of tsne dimensions to compute. Domain: U(2, 100).
Default: 11. Format: integer.}

\item{max_iter}{Maximal number of tsne iterations. Domain: e^U(4.61, 11.51).
Default: 1000. Format: integer.}

\item{cluster_method}{Method to use for clustering. Domain: {graph.clustering,
model.clustering}. Default: graph.clustering. Format: character.}

\item{k_clustering}{Number of nearest neighbors to build a k-nearest neighbors
graph for clustering. Domain: U(2, 1000). Default: 20. Format: integer.}

\item{ndim_pca_clustering}{Number of PCA dimensions used for k-nearest
neighbors graph for clustering. Domain: U(2, 100). Default: 20. Format:
integer.}

\item{k_knn}{Number of nearest neighbors to build a k-nearest neighbors graph
for knn. Domain: e^U(0.69, 6.91). Default: 10. Format: integer.}

\item{ndim_pca_knn}{Number of PCA dimensions used for knn. Domain: U(2, 100).
Default: 20. Format: integer.}

\item{sim_type}{Similarity type for knn. Domain: {jaccard}. Default: jaccard.
Format: character.}

\item{distance_method_paths}{Distance method for paths. Domain: {euclidean,
maximum, manhattan, canberra, binary, graph}. Default: graph. Format:
character.}

\item{ranks}{How to rank the paths. Domain: {path_cost, path_flow, rank,
length}. Default: rank. Format: character.}

\item{num_cells}{Trajectories should contain at least num.cells. Domain: U(3,
100). Default: 3. Format: integer.}

\item{neighs}{The size of the neighborhood in kNN graph used to smoothen
kinetic profiles. Domain: U(2, 100). Default: 3. Format: integer.}

\item{perplexity}{Perplexity parameter for tsne. Domain: U(5, 100). Default:
30. Format: numeric.}
}
\value{
A TI method wrapper to be used together with
\code{\link[dynwrap:infer_trajectories]{infer_trajectory}}
}
\description{
Will generate a trajectory using
\href{https://doi.org/10.1038/s41467-018-03214-y}{CellRouter}.

This method was wrapped inside a
\href{https://github.com/dynverse/ti_cellrouter}{container}.
The original code of this method is available
\href{https://github.com/edroaldo/cellrouter}{here}.
}
\references{
Lummertz da Rocha, E., Rowe, R.G., Lundin, V., Malleshaiah, M.,
Jha, D.K., Rambo, C.R., Li, H., North, T.E., Collins, J.J., Daley, G.Q., 2018.
Reconstruction of complex single-cell trajectories using CellRouter. Nature
Communications 9.
}
