% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_container.R
\name{ti_celltrails}
\alias{ti_celltrails}
\title{Inferring a trajectory inference using \href{https://doi.org/10.1016/j.celrep.2018.05.002}{CellTrails}}
\usage{
ti_celltrails(threshold_dl = 2L, threshold_cov = 0.05, threshold_ff = 1L,
  frac = 100L, min_size = 0.01, min_feat = 5L, max_pval = 1e-04,
  min_fc = 2L, l = 10L)
}
\arguments{
\item{threshold_dl}{Minimum number of samples; if value < 1 it is interpreted as fraction, otherwise as absolute sample count \cr
integer; default: 2L; possible values between 0 and 100}

\item{threshold_cov}{Minimum coefficient of variation; numeric value between 0 and 1 \cr
numeric; default: 0.05; possible values between 0 and 1}

\item{threshold_ff}{A Z-score cutoff \cr
numeric; default: 1L; possible values between 0 and 5}

\item{frac}{Fraction or number (if frac > 1) of eigengaps used to perform linear fit. \cr
numeric; default: 100L; possible values between 1 and 1000}

\item{min_size}{The initial cluster dedrogram is cut at an height such that the minimum cluster size is at least min_size; if min_size < 1 than the fraction of total samples is used, otherwise it is used as absoulte count \cr
numeric; default: 0.01; possible values between 0.001 and 1}

\item{min_feat}{Minimum number of differentially expressed features between siblings. If this number is not reached, two neighboring clusters (siblings) in the pruned dendrogram get joined. \cr
integer; default: 5L; possible values between 1 and 100}

\item{max_pval}{Maximum P-value for differential expression computation. \cr
numeric; default: 1e-04; possible values between 1e-07 and 1}

\item{min_fc}{Mimimum fold-change for differential expression computation \cr
numeric; default: 2L; possible values between 1 and 5}

\item{l}{Neighborhood size \cr
integer; default: 10L; possible values between 1 and 20}
}
\value{
The trajectory model
}
\description{
Will generate a trajectory using \href{https://doi.org/10.1016/j.celrep.2018.05.002}{CellTrails}. This method was wrapped inside a \href{https://github.com/dynverse/dynmethods/tree/master/containers/celltrails}{container}.
}
\details{
The original code of this method is available \href{https://github.com/dcellwanger/CellTrails}{here}.

The method is described in: \href{https://doi.org/10.1016/j.celrep.2018.05.002}{Ellwanger, D.C., Scheibinger, M., Dumont, R.A., Barr-Gillespie, P.G., Heller, S., 2018. Transcriptional Dynamics of Hair-Bundle Morphogenesis Revealed with CellTrails. Cell Reports 23, 2901–2914.e14.}
}
