% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_comp1.R
\name{ti_comp1}
\alias{ti_comp1}
\title{Inferring a trajectory inference using Component 1}
\usage{
ti_comp1(dimred = "pca", ndim = 2L, component = 1L)
}
\arguments{
\item{dimred}{discrete; Which dimensionality reduction method to use. (default:
\code{"pca"}; values: {\code{"pca"}, \code{"mds"}, \code{"tsne"}, \code{"ica"}, \code{"lle"},
\code{"landmark_mds"}, \code{"mds_sammon"}, \code{"mds_isomds"}, \code{"mds_smacof"}, \code{"umap"},
\code{"dm_diffusionMap"}})}

\item{ndim}{integer; (default: \code{2L}; range: from \code{2L} to \code{30L})}

\item{component}{integer; (default: \code{1L}; range: from \code{1L} to \code{10L})}
}
\value{
A TI method wrapper to be used together with
\code{\link[dynwrap:infer_trajectories]{infer_trajectory}}
}
\description{
Will generate a trajectory using Component 1.

This method was wrapped inside a
\href{https://github.com/dynverse/ti_comp1}{container}.
}
