% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_elpigraph.R
\name{ti_elpigraph}
\alias{ti_elpigraph}
\title{ElPiGraph}
\usage{
ti_elpigraph(
  topology = "tree",
  NumNodes = 50L,
  NumEdges = 100000L,
  InitNodes = 2L,
  Mu = 0.1,
  Lambda = 0.01,
  MaxNumberOfIterations = 10L,
  eps = 0.01,
  CenterData = FALSE
)
}
\arguments{
\item{topology}{The kind of topology to detect. Domain: {cycle, tree, linear}.
Default: tree. Format: character.}

\item{NumNodes}{The number of nodes of the principal graph. Domain: U(2, 1000).
Default: 50. Format: integer.}

\item{NumEdges}{The maximum number of edges. Domain: e^U(0.69, 11.51). Default:
100000. Format: integer.}

\item{InitNodes}{Number of points to include in the initial graph. Domain:
e^U(0.69, 6.91). Default: 2. Format: integer.}

\item{Mu}{Controls the elastic energy. Domain: e^U(-6.91, 0.00). Default: 0.1.
Format: numeric.}

\item{Lambda}{Controls the elastic energy. Domain: e^U(-6.91, 0.00). Default:
0.01. Format: numeric.}

\item{MaxNumberOfIterations}{Maximum number of steps to embed the nodes.
Domain: e^U(0.00, 6.91). Default: 10. Format: integer.}

\item{eps}{Minimal relative change in the position of the nodes to stop
embedment. Domain: e^U(-6.91, 0.00). Default: 0.01. Format: numeric.}

\item{CenterData}{Should data and initial node positions be centered?. Default:
FALSE. Format: logical.}
}
\value{
A TI method wrapper to be used together with
\code{\link[dynwrap:infer_trajectories]{infer_trajectory}}
}
\description{
Will generate a trajectory using
\href{https://github.com/Albluca/ElPiGraph.R}{ElPiGraph}.

This method was wrapped inside a
\href{https://github.com/dynverse/ti_elpigraph}{container}.
The original code of this method is available
\href{https://github.com/Albluca/ElPiGraph.R}{here}.
}
\keyword{method}
