% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_elpilinear.R
\name{ti_elpilinear}
\alias{ti_elpilinear}
\title{Inferring a trajectory inference using ElPiGraph linear}
\usage{
ti_elpilinear(topology = "linear", NumNodes = 50L, NumEdges = 100000L,
  InitNodes = 2L, Mu = 0.1, Lambda = 0.01, MaxNumberOfIterations = 10L,
  eps = 0.01, CenterData = FALSE, run_environment = NULL)
}
\arguments{
\item{topology}{discrete; The kind of topology to detect (default: \code{"linear"};
values: {\code{"linear"}})}

\item{NumNodes}{integer; The number of nodes of the principal graph (default:
\code{50L}; range: from \code{2L} to \code{1000L})}

\item{NumEdges}{integer; The maximum number of edges (default: \code{100000L};
range: from \code{2L} to \code{100000L})}

\item{InitNodes}{integer; Number of points to include in the initial graph
(default: \code{2L}; range: from \code{2L} to \code{1000L})}

\item{Mu}{numeric; Controls the elastic energy (default: \code{0.1}; range: from
\code{0.001} to \code{1L})}

\item{Lambda}{numeric; Controls the elastic energy (default: \code{0.01}; range:
from \code{0.001} to \code{1L})}

\item{MaxNumberOfIterations}{integer; Maximum number of steps to embed the
nodes (default: \code{10L}; range: from \code{1L} to \code{1000L})}

\item{eps}{numeric; Minimal relative change in the position of the nodes to
stop embedment (default: \code{0.01}; range: from \code{0.001} to \code{1L})}

\item{CenterData}{logical; Should data and initial node positions be centered?}

\item{run_environment}{In which environment to run the method, can be \code{"docker"} or \code{"singularity"}.}
}
\value{
A TI method wrapper to be used together with
\code{\link[dynwrap:infer_trajectories]{infer_trajectory}}
}
\description{
Will generate a trajectory using \href{https://github.com/Albluca/ElPiGraph.R}{ElPiGraphlinear}.

This method was wrapped inside a
\href{https://github.com/dynverse/dynmethods/tree/master/containers/elpigraph}{container}.
The original code of this method is available
\href{https://github.com/Albluca/ElPiGraph.R}{here}.
}
