% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_embeddr.R
\name{ti_embeddr}
\alias{ti_embeddr}
\title{Inferring trajectories with Embeddr}
\usage{
ti_embeddr(kernel = "nn", metric = "correlation", nn_pct = 1, eps = 1,
  t = 1, symmetrize = "mean", measure_type = "unorm", p = 2L,
  thresh = 0.001, maxit = 10L, stretch = 2, smoother = "smooth.spline")
}
\description{
Arguments passed to this function will be used as default parameters for the method.
}
