% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_controls.R, R/ti_container.R
\name{ti_error}
\alias{ti_error}
\alias{ti_error}
\title{Inferring a trajectory inference using Control: error}
\usage{

  ti_error(run_environment = ifelse(is.null(getOption("dynwrap_run_environment")),
  "local", getOption("dynwrap_run_environment")), dummy_param = 0.5)


  ti_error(run_environment = ifelse(is.null(getOption("dynwrap_run_environment")),
  "local", getOption("dynwrap_run_environment")), dummy_param = 0.5)
}
\arguments{
\item{run_environment}{In which environment to run the method, can be 'local', 'docker' or 'singularity'}

\item{dummy_param}{This parameter does not do anything.}
}
\value{
The trajectory model
}
\description{
Will generate a trajectory using Control: error. This method was wrapped inside a \href{https://github.com/dynverse/dynmethods/tree/master/containers/error}{container}.
}
\details{
This methods was first wrapped inside R, see \link{ti_error}
}
