% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_container.R
\name{ti_fateid}
\alias{ti_fateid}
\title{Inferring a trajectory inference using \href{https://doi.org/10.1038/nmeth.4662}{FateID}}
\usage{
ti_fateid(alpha = 0.1, lambda_auto = TRUE, lambda = 0.1)
}
\arguments{
\item{alpha}{The elastic net mixing parameter of the ‘glmnet’ classifier. \cr
numeric; default: 0.1; possible values between 0.001 and 10}

\item{lambda_auto}{Whether to select the lambda by cross-validation \cr}

\item{lambda}{The lambda penalty of GLM. \cr
numeric; default: 0.1; possible values between 0.05 and 1}
}
\value{
The trajectory model
}
\description{
Will generate a trajectory using \href{https://doi.org/10.1038/nmeth.4662}{FateID}. This method was wrapped inside a \href{https://github.com/dynverse/dynmethods/tree/master/containers/fateid}{container}.
}
\details{
The original code of this method is available \href{https://git.embl.de/velten/STEMNET}{here}.

The method is described in: \href{https://doi.org/10.1038/nmeth.4662}{Herman, J.S., Sagar, Grün, D., 2018. FateID infers cell fate bias in multipotent progenitors from single-cell RNA-seq data. Nature Methods 15, 379–386.}
}
