% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_forks.R
\name{ti_forks}
\alias{ti_forks}
\title{Inferring a trajectory inference using FORKS}
\usage{
ti_forks(param1 = 5L, param2 = "one", param3 = 0.5, param4 = TRUE,
  run_environment = NULL)
}
\arguments{
\item{param1}{integer; Description of param1 (default: \code{5L}; range: from \code{1L}
to \code{10L})}

\item{param2}{discrete; Description of param2 (default: \code{"one"}; values:
{\code{"one"}, \code{"two"}, \code{"three"}})}

\item{param3}{numeric; Description of param3 (default: \code{0.5}; range: from \code{0}
to \code{1})}

\item{param4}{logical; Description of param4}

\item{run_environment}{In which environment to run the method, can be \code{"docker"} or \code{"singularity"}.}
}
\value{
A TI method wrapper to be used together with
\code{\link[dynwrap:infer_trajectories]{infer_trajectory}}
}
\description{
Will generate a trajectory using \href{https://doi.org/10.1101/132811}{FORKS}.

This method was wrapped inside a
\href{https://github.com/dynverse/dynmethods/tree/master/containers/forks}{container}.
The original code of this method is available
\href{https://github.com/macsharma/FORKS}{here}.
}
\references{
Sharma, M., Li, H., Sengupta, D., Prabhakar, S., Jayadeva, J.,
2017. FORKS: Finding Orderings Robustly using K-means and Steiner trees.
}
