% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_container.R
\name{ti_merlot}
\alias{ti_merlot}
\title{Inferring a trajectory inference using \href{https://doi.org/10.1101/261768}{MERLoT}}
\usage{
ti_merlot(run_environment = NULL, sigma = "local", distance = "euclidean",
  ndim = 20L, density_norm = TRUE, n_local_lower = 5L,
  n_local_upper = 7L, w_width = 0.01, n_components_to_use = 3,
  N_yk = 100, lambda_0 = 1e-10, mu_0 = 0.0025, increaseFactor_mu = 20,
  increaseFactor_lambda = 20, FixEndpoints = FALSE)
}
\arguments{
\item{run_environment}{In which environment to run the method, can be 'docker' or 'singularity'}

\item{sigma}{Diffusion scale parameter of the Gaussian kernel. A larger sigma might be necessary if the eigenvalues can not be found because of a singularity in the matrix. Must be one of:\itemize{\item A character vector: \code{"local"} (default) or \code{"global"},\item a numeric global sigma -- a global sigma will be calculated using \code{\link[destiny:find_sigmas]{destiny::find_sigmas()}}\item or a \code{\link[destiny:Sigmas-class]{destiny::Sigmas-class()}} object.} \cr
discrete; default: "local"; possible values: local, global}

\item{distance}{A \code{\link[stats:dist]{stats::dist()}} object, or a character vector specifying which distance metric to use. Allowed measures:\itemize{\item Euclidean distance (default),\item cosine distance (1-corr(c_1, c_2)), or\item rank correlation distance (1-corr(rank(c_1), rank(c_2)))} \cr
discrete; default: "euclidean"; possible values: euclidean, cosine, rankcor}

\item{ndim}{Number of eigenvectors/dimensions to return \cr
integer; default: 20L; possible values between 2 and 20}

\item{density_norm}{Logical. If TRUE, use density normalisation \cr
logical; default: TRUE; possible values: TRUE, FALSE}

\item{n_local_lower}{If sigma == 'local', the \code{n_local_lower}:\code{n_local_upper} nearest neighbor(s) determine(s) the local sigma \cr
integer; default: 5L; possible values between 2 and 20}

\item{n_local_upper}{See \code{n_local_lower} \cr
integer; default: 7L; possible values between 2 and 20}

\item{w_width}{Window width to use for deciding the branch cutoff \cr
numeric; default: 0.01; possible values between 1e-04 and 1}

\item{n_components_to_use}{Which components to use in downstream analysis \cr
integer; default: 3; possible values between 2 and 20}

\item{N_yk}{Number of nodes for the elastic principal tree \cr
integer; default: 100; possible values between 2 and 1000}

\item{lambda_0}{Principal elastic tree energy function parameter. \cr
numeric; default: 1e-10; possible values between 1e-15 and 1e-04}

\item{mu_0}{Principal elastic tree energy function parameter. \cr
numeric; default: 0.0025; possible values between 5e-04 and 0.005}

\item{increaseFactor_mu}{Factor by which the mu will be increased for the embedding \cr
numeric; default: 20; possible values between 2 and 50}

\item{increaseFactor_lambda}{Factor by which the mu will be increased for the embedding \cr
numeric; default: 20; possible values between 2 and 50}

\item{FixEndpoints}{Documentation not provided by authors \cr
logical; default: FALSE; possible values: TRUE, FALSE}
}
\value{
The trajectory model
}
\description{
Will generate a trajectory using \href{https://doi.org/10.1101/261768}{MERLoT}. This method was wrapped inside a \href{https://github.com/dynverse/dynmethods/tree/master/containers/merlot}{container}.
}
\details{
The original code of this method is available \href{https://github.com/soedinglab/merlot}{here}.

The method is described in: \href{https://doi.org/10.1101/261768}{Parra, R.G., Papadopoulos, N., Ahumada-Arranz, L., El Kholtei, J., Treutlein, B., Soeding, J., 2018. Reconstructing complex lineage trees from scRNA-seq data using MERLoT.}
}
