% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_monocle.R, R/ti_container.R
\name{ti_monocle_ddrtree}
\alias{ti_monocle_ddrtree}
\alias{ti_monocle_ddrtree}
\title{Inferring trajectories with Monocle v2}
\usage{
ti_monocle_ddrtree(docker = "stevedore" \%in\% rownames(installed.packages()),
  reduction_method = "DDRTree", max_components = 2,
  norm_method = "vstExprs", auto_param_selection = TRUE)

ti_monocle_ddrtree(docker = "stevedore" \%in\% rownames(installed.packages()),
  reduction_method = "DDRTree", max_components = 2,
  norm_method = "vstExprs", auto_param_selection = TRUE)
}
\arguments{
\item{docker}{Whether to use the docker container or the R wrapper}

\item{reduction_method}{A character string specifying the algorithm to use for dimensionality reduction.}

\item{max_components}{the dimensionality of the reduced space}

\item{norm_method}{Determines how to transform expression values prior to reducing dimensionality}

\item{auto_param_selection}{when this argument is set to TRUE (default), it will automatically calculate the proper value for the ncenter (number of centroids) parameters which will be passed into DDRTree call.}
}
\value{
The trajectory model
}
\description{
Inferring trajectories with Monocle v2

Will generate a trajectory using \href{https://doi.org/10.1038/nmeth.4402}{Monocle DDRTree}. This method was wrapped inside a \href{https://github.com/dynverse/dynmethods/tree/master/containers/monocle_ddrtree}{container}.
}
\details{
This methods was first wrapped inside R, see \link{ti_monocle_ddrtree}

The original code of this method is available \href{https://github.com/cole-trapnell-lab/monocle-release}{here}.

The method is described in: \href{https://doi.org/10.1038/nmeth.4402}{Qiu, X., Mao, Q., Tang, Y., Wang, L., Chawla, R., Pliner, H.A., Trapnell, C., 2017. Reversed graph embedding resolves complex single-cell trajectories. Nature Methods 14, 979–982.}
}
\seealso{
\code{\link[monocle:reduceDimension]{monocle::reduceDimension()}}, \code{\link[monocle:orderCells]{monocle::orderCells()}}
}
