% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_monocle_ica.R
\name{ti_monocle_ica}
\alias{ti_monocle_ica}
\title{Inferring a trajectory inference using Monocle ICA}
\usage{
ti_monocle_ica(reduction_method = "ICA", max_components = 2L,
  norm_method = "vstExprs", run_environment = NULL)
}
\arguments{
\item{reduction_method}{discrete; A character string specifying the algorithm
to use for dimensionality reduction. (default: \code{"ICA"}; values: {\code{"ICA"}})}

\item{max_components}{integer; The dimensionality of the reduced space
(default: \code{2L}; range: from \code{2L} to \code{20L})}

\item{norm_method}{discrete; Determines how to transform expression values
prior to reducing dimensionality (default: \code{"vstExprs"}; values: {\code{"vstExprs"},
\code{"log"}, \code{"none"}})}

\item{run_environment}{In which environment to run the method, can be \code{"docker"} or \code{"singularity"}.}
}
\value{
A TI method wrapper to be used together with
\code{\link[dynwrap:infer_trajectories]{infer_trajectory}}
}
\description{
Will generate a trajectory using \href{https://doi.org/10.1038/nmeth.4402}{Monocle ICA}.

This method was wrapped inside a
\href{https://github.com/dynverse/dynmethods/tree/master/containers/monocle_ica}{container}.
The original code of this method is available
\href{https://github.com/cole-trapnell-lab/monocle-release}{here}.
}
\references{
Qiu, X., Mao, Q., Tang, Y., Wang, L., Chawla, R., Pliner, H.A.,
Trapnell, C., 2017. Reversed graph embedding resolves complex single-cell
trajectories. Nature Methods 14, 979–982.
}
