% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_mpath.R
\name{ti_mpath}
\alias{ti_mpath}
\title{Inferring trajectories with Mpath}
\usage{
ti_mpath(distMethod = "euclidean", method = "kmeans", numcluster = 11L,
  numcluster_null = TRUE, diversity_cut = 0.6, size_cut = 0.05)
}
\description{
Arguments passed to this function will be used as default parameters for the method.
}
