% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_mst.R
\name{ti_mst}
\alias{ti_mst}
\title{Inferring a trajectory inference using MST}
\usage{
ti_mst(dimred = "pca", ndim = 2L)
}
\arguments{
\item{dimred}{discrete; Which dimensionality reduction method to use. (default:
\code{"pca"}; values: {\code{"pca"}, \code{"mds"}, \code{"tsne"}, \code{"ica"}, \code{"lle"},
\code{"landmark_mds"}, \code{"mds_sammon"}, \code{"mds_isomds"}, \code{"mds_smacof"}, \code{"umap"},
\code{"dm_diffusionMap"}})}

\item{ndim}{integer; (default: \code{2L}; range: from \code{2L} to \code{30L})}
}
\value{
A TI method wrapper to be used together with
\code{\link[dynwrap:infer_trajectories]{infer_trajectory}}
}
\description{
Will generate a trajectory using MST.

This method was wrapped inside a
\href{https://github.com/dynverse/ti_mst}{container}.
}
