% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_ouija.R
\name{ti_ouija}
\alias{ti_ouija}
\title{Inferring trajectories with ouija}
\usage{
ti_ouija(iter = 100, response_type = "switch", inference_type = "hmc",
  normalise_expression = TRUE)
}
\description{
Inferring trajectories with ouija
}
