% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_ouijaflow.R
\name{ti_ouijaflw}
\alias{ti_ouijaflw}
\title{Inferring trajectories with ouijaflow}
\usage{
ti_ouijaflw(iter = 1000)
}
\arguments{
\item{iter}{Number of iterations of optimisation to perform}
}
\description{
Inferring trajectories with ouijaflow
}
