% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_pcreode.R
\name{ti_pcreode}
\alias{ti_pcreode}
\title{Inferring a trajectory inference using pCreode}
\usage{
ti_pcreode(n_pca_components = 3L, num_runs = 10L,
  config = dynwrap::container_config())
}
\arguments{
\item{n_pca_components}{integer; (default: \code{3L}; range: from \code{2L} to \code{10L})}

\item{num_runs}{integer; (default: \code{10L}; range: from \code{10L} to \code{100L})}

\item{config}{A container config. See \code{\link[=container_config]{container_config()}} for more information.}
}
\value{
A TI method wrapper to be used together with
\code{\link[dynwrap:infer_trajectories]{infer_trajectory}}
}
\description{
Will generate a trajectory using
\href{https://doi.org/10.1016/j.cels.2017.10.012}{pCreode}.

This method was wrapped inside a
\href{https://github.com/dynverse/dynmethods/tree/master/containers/pcreode}{container}.
The original code of this method is available
\href{https://github.com/KenLauLab/pCreode}{here}.
}
\references{
Herring, C.A., Banerjee, A., McKinley, E.T., Simmons, A.J., Ping,
J., Roland, J.T., Franklin, J.L., Liu, Q., Gerdes, M.J., Coffey, R.J., Lau,
K.S., 2018. Unsupervised Trajectory Analysis of Single-Cell RNA-Seq and Imaging
Data Reveals Alternative Tuft Cell Origins in the Gut. Cell Systems 6,
37–51.e9.
}
