% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_periodpc.R
\name{ti_periodpc}
\alias{ti_periodpc}
\title{Inferring a trajectory inference using Periodic PrinCurve}
\usage{
ti_periodpc(ndim = 3L, maxit = 10L)
}
\arguments{
\item{ndim}{Parameter; , Domain: U(2, 10), Default: 3, Format: integer.}

\item{maxit}{Parameter; , Domain: U(0, 100), Default: 10, Format: integer.}
}
\value{
A TI method wrapper to be used together with
\code{\link[dynwrap:infer_trajectories]{infer_trajectory}}
}
\description{
Will generate a trajectory using Periodic PrinCurve.

This method was wrapped inside a
\href{https://github.com/dynverse/ti_periodpc}{container}.
}
