% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_projected_dpt.R
\name{ti_projected_dpt}
\alias{ti_projected_dpt}
\title{Inferring a trajectory inference using Projected DPT}
\usage{
ti_projected_dpt(sigma = "local", distance = "euclidean", ndim = 20L,
  density_norm = TRUE, n_local = c(5L, 7L), w_width = 0.1)
}
\arguments{
\item{sigma}{Parameter; Diffusion scale parameter of the Gaussian kernel. A
larger sigma might be necessary if the eigenvalues can not be found because of
a singularity in the matrix. Must be one of:\itemize{\item A character vector:
\code{"local"} (default) or \code{"global"},\item a numeric global sigma -- a
global sigma will be calculated using \code{destiny::find_sigmas()}\item or a
\code{\link[destiny:Sigmas-class]{destiny::Sigmas-class()}} object.}., Domain:
{local, global}, Default: local, Format: character.}

\item{distance}{Parameter; A \code{stats::dist()} object, or a character vector
specifying which distance metric to use. Allowed measures:\itemize{\item
Euclidean distance (default),\item cosine distance (1-corr(c_1, c_2)), or\item
rank correlation distance (1-corr(rank(c_1), rank(c_2)))}., Domain: {euclidean,
cosine, rankcor}, Default: euclidean, Format: character.}

\item{ndim}{Parameter; Number of eigenvectors/dimensions to return., Domain:
U(3, 100), Default: 20, Format: integer.}

\item{density_norm}{Parameter; Logical. If TRUE, use density normalisation.,
Default: TRUE, Format: logical.}

\item{n_local}{Parameter; If sigma == 'local', the \code{n_local} nearest
neighbor(s) determine(s) the local sigma., Domain: ( U(2, 20), U(2, 20) ),
Default: (5, 7), Format: range.}

\item{w_width}{Parameter; Window width to use for deciding the branch cutoff.,
Domain: e^U(-9.21, 0.00), Default: 0.1, Format: numeric.}
}
\value{
A TI method wrapper to be used together with
\code{\link[dynwrap:infer_trajectories]{infer_trajectory}}
}
\description{
Will generate a trajectory using \href{https://bioconductor.org/packages/release/bioc/html/destiny.html}{Projected DPT}.

This method was wrapped inside a
\href{https://github.com/dynverse/ti_projected_dpt}{container}.
The original code of this method is available
\href{https://bioconductor.org/packages/release/bioc/html/destiny.html}{here}.
}
