% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_projected_gng.R
\name{ti_projected_gng}
\alias{ti_projected_gng}
\title{Inferring a trajectory inference using Projected GNG}
\usage{
ti_projected_gng(dimred = "pca", ndim = 5L, max_iter = 15000,
  max_nodes = 8L, apply_mst = TRUE, run_environment = NULL)
}
\arguments{
\item{dimred}{discrete; Which dimensionality reduction method to use. (default:
\code{"pca"}; values: {\code{"pca"}, \code{"mds"}, \code{"tsne"}, \code{"ica"}, \code{"lle"},
\code{"landmark_mds"}, \code{"mds_sammon"}, \code{"mds_isomds"}, \code{"mds_smacof"}, \code{"umap"},
\code{"dm_diffusionMap"}})}

\item{ndim}{integer; The number of dimensions (default: \code{5L}; range: from \code{2L}
to \code{10L})}

\item{max_iter}{numeric; The max number of iterations (default: \code{15000}; range:
from \code{25} to \code{1e+06})}

\item{max_nodes}{integer; The maximum number of nodes (default: \code{8L}; range:
from \code{2L} to \code{30L})}

\item{apply_mst}{logical; If true, an MST post-processing of the GNG is
performed.}

\item{run_environment}{In which environment to run the method, can be \code{"docker"} or \code{"singularity"}.}
}
\value{
A TI method wrapper to be used together with
\code{\link[dynwrap:infer_trajectories]{infer_trajectory}}
}
\description{
Will generate a trajectory using \href{https://github.com/rcannood/gng}{Projected GNG}.

This method was wrapped inside a
\href{https://github.com/dynverse/dynmethods/tree/master/containers/projected_gng}{container}.
The original code of this method is available
\href{https://github.com/rcannood/gng}{here}.
}
