% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_projected_slingshot.R
\name{ti_projected_slingshot}
\alias{ti_projected_slingshot}
\title{Inferring a trajectory inference using Projected Slingshot}
\usage{
ti_projected_slingshot(shrink = 1, reweight = TRUE, reassign = TRUE,
  thresh = 0.001, maxit = 10L, stretch = 2,
  smoother = "smooth.spline", shrink.method = "cosine",
  container_type = NULL)
}
\arguments{
\item{shrink}{numeric; Logical or numeric between 0 and 1, determines whether
and how much to shrink branching lineages toward their average prior to the
split. (default: \code{1}; range: from \code{0} to \code{1})}

\item{reweight}{logical; Logical, whether to allow cells shared between
lineages to be reweighted during curve-fitting. If \code{TRUE}, cells shared
between lineages will be weighted by: distance to nearest curve / distance to
curve.}

\item{reassign}{logical; Logical, whether to reassign cells to lineages at each
iteration. If TRUE, cells will be added to a lineage when their projection
distance to the curve is less than the median distance for all cells currently
assigned to the lineage. Additionally, shared cells will be removed from a
lineage if their projection distance to the curve is above the 90th percentile
and their weight along the curve is less than 0.1.}

\item{thresh}{numeric; Numeric, determines the convergence criterion. Percent
change in the total distance from cells to their projections along curves must
be less than thresh. Default is 0.001, similar to principal.curve. (default:
\code{0.001}; range: from \code{1e-05} to \code{1e+05})}

\item{maxit}{integer; Numeric, maximum number of iterations, see
principal.curve. (default: \code{10L}; range: from \code{0L} to \code{50L})}

\item{stretch}{numeric; Numeric factor by which curves can be extrapolated
beyond endpoints. Default is 2, see principal.curve. (default: \code{2}; range: from
\code{0} to \code{5})}

\item{smoother}{discrete; Choice of scatter plot smoother. Same as
principal.curve, but "lowess" option is replaced with "loess" for additional
flexibility. (default: \code{"smooth.spline"}; values: {\code{"smooth.spline"},
\code{"loess"}, \code{"periodic.lowess"}})}

\item{shrink.method}{discrete; Character denoting how to determine the
appropriate amount of shrinkage for a branching lineage. Accepted values are
the same as for \code{kernel} in \code{\link[=density]{density()}} (default is \code{"cosine"}), as
well as \code{"tricube"} and \code{"density"}. See 'Details' for more.
(default: \code{"cosine"}; values: {\code{"cosine"}, \code{"tricube"}, \code{"density"}})}

\item{container_type}{In which environment to run the method, can be \code{"docker"} or \code{"singularity"}.
By default, docker will be used.}
}
\value{
A TI method wrapper to be used together with
\code{\link[dynwrap:infer_trajectories]{infer_trajectory}}
}
\description{
Will generate a trajectory using \href{https://github.com/kstreet13/slingshot}{Projected Slingshot}.

This method was wrapped inside a
\href{https://github.com/dynverse/dynmethods/tree/master/containers/projected_slingshot}{container}.
The original code of this method is available
\href{https://github.com/kstreet13/slingshot}{here}.
}
