% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_pseudogp.R, R/ti_container.R
\name{ti_pseudogp}
\alias{ti_pseudogp}
\alias{ti_pseudogp}
\title{Inferring trajectories with pseudogp}
\usage{
ti_pseudogp(docker = "stevedore" \%in\% rownames(installed.packages()),
  smoothing_alpha = 10, smoothing_beta = 3, pseudotime_mean = 0.5,
  pseudotime_var = 1, chains = 3, iter = 100, dimreds = c(TRUE, TRUE,
  FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE),
  initialise_from = "random")

ti_pseudogp(docker = "stevedore" \%in\% rownames(installed.packages()),
  smoothing_alpha = 10, smoothing_beta = 3, pseudotime_mean = 0.5,
  pseudotime_var = 1, chains = 3, iter = 100, dimreds = c(TRUE, TRUE,
  FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE),
  initialise_from = "random")
}
\arguments{
\item{docker}{Whether to use the docker container or the R wrapper}

\item{smoothing_alpha}{The hyperparameter for the Gamma distribution that controls arc-length}

\item{smoothing_beta}{The hyperparameter for the Gamma distribution that controls arc-length}

\item{pseudotime_mean}{The mean of the constrained normal prior on the pseudotimes}

\item{pseudotime_var}{The variance of the constrained normal prior on the pseudotimes}

\item{chains}{The number of chains for the MCMC trace}

\item{iter}{The number of iterations for the MCMC trace}

\item{dimreds}{A character vector specifying which dimensionality reduction methods to use.
See \code{\link[dyndimred:dimred]{dyndimred::dimred()}} for the list of available dimensionality reduction methods.}

\item{initialise_from}{How to initialise the MCMC chain. One of "random" (stan decides),
"principal_curve", or "pca" (the first component of PCA rescaled is taken to be the pseudotimes).
Note: if multiple representations are provided, \code{pseudogp} will take the principal curve or
pca from the first rather than combining them. If a particular representation is required, it is
up to the user to re-order them.}
}
\value{
The trajectory model
}
\description{
Inferring trajectories with pseudogp

Will generate a trajectory using \href{https://doi.org/10.1371/journal.pcbi.1005212}{pseudogp}. This method was wrapped inside a \href{https://github.com/dynverse/dynmethods/tree/master/containers/pseudogp}{container}.
}
\details{
This methods was first wrapped inside R, see \link{ti_pseudogp}

The original code of this method is available \href{https://github.com/kieranrcampbell/pseudogp}{here}.

The method is described in: \href{https://doi.org/10.1371/journal.pcbi.1005212}{Campbell, K.R., Yau, C., 2016. Order Under Uncertainty: Robust Differential Expression Analysis Using Probabilistic Models for Pseudotime Inference. PLOS Computational Biology 12, e1005212.}
}
\seealso{
\code{\link[pseudogp:fitPseudotime]{pseudogp::fitPseudotime()}}
}
