% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_random.R
\name{ti_random}
\alias{ti_random}
\title{Inferring a trajectory inference using Random}
\usage{
ti_random(dummy_param = 0.5)
}
\arguments{
\item{dummy_param}{numeric; Dummy parameter (default: \code{0.5}; range: from \code{0} to
\code{1})}
}
\value{
A TI method wrapper to be used together with
\code{\link[dynwrap:infer_trajectories]{infer_trajectory}}
}
\description{
Will generate a trajectory using Random.

This method was wrapped inside a
\href{https://github.com/dynverse/ti_random}{container}.
}
