% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_recat.R
\name{ti_recat}
\alias{ti_recat}
\title{Inferring trajectories with reCAT}
\usage{
ti_recat(TSPFold = 2L, beginNum = 10L, endNum = 15L, step_size = 2L,
  base_cycle_range_start = 6L, base_cycle_range_end = 9L, max_num = 300L,
  clustMethod = "GMM")
}
\description{
Inferring trajectories with reCAT
}
