% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_scorpius.R
\name{ti_scorpius}
\alias{ti_scorpius}
\title{Inferring a trajectory inference using SCORPIUS}
\usage{
ti_scorpius(distance_method = "spearman", ndim = 3L, k = 4L,
  thresh = 0.001, maxit = 10L, stretch = 0,
  smoother = "smooth_spline", sparse = TRUE, run_environment = NULL)
}
\arguments{
\item{distance_method}{discrete; A character string indicating which
correlationcoefficient (or covariance) is to be computed. One of "pearson",
"kendall", or "spearman". (default: \code{"spearman"}; values: {\code{"spearman"},
\code{"pearson"}, \code{"kendall"}})}

\item{ndim}{integer; The number of dimensions in the new space. (default: \code{3L};
range: from \code{2L} to \code{20L})}

\item{k}{integer; The number of clusters to cluster the data into. (default:
\code{4L}; range: from \code{1L} to \code{20L})}

\item{thresh}{numeric; \code{\link[princurve]{principal_curve}} parameter:
convergence threshhold on shortest distances to the curve (default: \code{0.001};
range: from \code{1e-05} to \code{1e+05})}

\item{maxit}{integer; \code{\link[princurve]{principal_curve}} parameter:
maximum number of iterations (default: \code{10L}; range: from \code{0L} to \code{50L})}

\item{stretch}{numeric; \code{\link[princurve]{principal_curve}} parameter: a
factor by which the curve can be extrapolated when points are projected
(default: \code{0}; range: from \code{0} to \code{5})}

\item{smoother}{discrete; \code{\link[princurve]{principal_curve}} parameter:
choice of smoother (default: \code{"smooth_spline"}; values: {\code{"smooth_spline"},
\code{"lowess"}, \code{"periodic_lowess"}})}

\item{sparse}{logical; Whether or not to use sparse MDS dimensionality
reduction,for datasets with large amounts of cells.}

\item{run_environment}{In which environment to run the method, can be \code{"docker"} or \code{"singularity"}.}
}
\value{
A TI method wrapper to be used together with
\code{\link[dynwrap:infer_trajectories]{infer_trajectory}}
}
\description{
Will generate a trajectory using \href{https://doi.org/10.1101/079509}{SCORPIUS}.

This method was wrapped inside a
\href{https://github.com/dynverse/dynmethods/tree/master/containers/scorpius}{container}.
The original code of this method is available
\href{https://github.com/rcannood/SCORPIUS}{here}.
}
\references{
Cannoodt, R., Saelens, W., Sichien, D., Tavernier, S., Janssens,
S., Guilliams, M., Lambrecht, B.N., De Preter, K., Saeys, Y., 2016. SCORPIUS
improves trajectory inference and identifies novel modules in dendritic cell
development.
}
