% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_container.R
\name{ti_scorpius_sparse}
\alias{ti_scorpius_sparse}
\title{Inferring a trajectory inference using \href{https://doi.org/10.1101/079509}{SCORPIUS sparse}}
\usage{
ti_scorpius_sparse(docker = TRUE, distance_method = "spearman", ndim = 3,
  k = 4, thresh = 0.001, maxit = 10, stretch = 0,
  smoother = "smooth.spline", sparse = TRUE)
}
\arguments{
\item{docker}{Whether to use the docker container or the R wrapper}
}
\value{
The trajectory model
}
\description{
Will generate a trajectory using \href{https://doi.org/10.1101/079509}{SCORPIUS sparse}. This method was wrapped inside a \href{https://github.com/dynverse/dynmethods/tree/master/containers/scorpius_sparse}{container}.
}
\details{
This methods was first wrapped inside R, see \link{ti_scorpius_sparse}

The original code of this method is available \href{https://github.com/rcannood/SCORPIUS}{here}.

The method is described in: \href{https://doi.org/10.1101/079509}{Cannoodt, R., Saelens, W., Sichien, D., Tavernier, S., Janssens, S., Guilliams, M., Lambrecht, B.N., De Preter, K., Saeys, Y., 2016. SCORPIUS improves trajectory inference and identifies novel modules in dendritic cell development.}
}
