% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_scoup.R
\name{ti_scoup}
\alias{ti_scoup}
\title{Inferring a trajectory inference using SCOUP}
\usage{
ti_scoup(ndim = 2L, max_ite1 = 100, max_ite2 = 100, alpha_min = 0.1,
  alpha_max = 100, t_min = 0.001, t_max = 2, sigma_squared_min = 0.1,
  thresh = 0.01, run_environment = NULL)
}
\arguments{
\item{ndim}{integer; Number of pca dimensions (default: \code{2L}; range: from \code{2L}
to \code{20L})}

\item{max_ite1}{numeric; Upper bound of EM iteration (without pseudo-time
optimization). The detailed explanation is described in the supplementary text.
(default is 1,000) (default: \code{100}; range: from \code{2} to \code{5000})}

\item{max_ite2}{numeric; Upper bound of EM iteration (including pseudo-time
optimization) (default is 1,000). (default: \code{100}; range: from \code{2} to \code{5e+05})}

\item{alpha_min}{numeric; Lower bound of alpha (default is 0.1) (default:
\code{0.1}; range: from \code{0.001} to \code{10})}

\item{alpha_max}{numeric; Upper bound of alpha (default is 100) (default:
\code{100}; range: from \code{1} to \code{10000})}

\item{t_min}{numeric; Lower bound of pseudo-time (default is 0.001) (default:
\code{0.001}; range: from \code{1e-05} to \code{1})}

\item{t_max}{numeric; Upper bound of pseudo-time (default is 2.0) (default:
\code{2}; range: from \code{0.1} to \code{100})}

\item{sigma_squared_min}{numeric; Lower bound of sigma squared (default is 0.1)
(default: \code{0.1}; range: from \code{0.001} to \code{10})}

\item{thresh}{numeric; Threshold (default: \code{0.01}; range: from \code{0.01} to \code{10})}

\item{run_environment}{In which environment to run the method, can be \code{"docker"} or \code{"singularity"}.}
}
\value{
A TI method wrapper to be used together with
\code{\link[dynwrap:infer_trajectories]{infer_trajectory}}
}
\description{
Will generate a trajectory using
\href{https://doi.org/10.1186/s12859-016-1109-3}{SCOUP}.

This method was wrapped inside a
\href{https://github.com/dynverse/dynmethods/tree/master/containers/scoup}{container}.
The original code of this method is available
\href{https://github.com/hmatsu1226/SCOUP}{here}.
}
\references{
Matsumoto, H., Kiryu, H., 2016. SCOUP: a probabilistic model based
on the Ornstein–Uhlenbeck process to analyze single-cell expression data during
differentiation. BMC Bioinformatics 17.
}
