% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_scuba.R
\name{ti_scuba}
\alias{ti_scuba}
\title{Inferring trajectories with SCUBA}
\usage{
ti_scuba(rigorous_gap_stats = TRUE, N_dim = 2L, low_gene_threshold = 1,
  low_gene_fraction_max = 0.7, min_split = 15L,
  min_percentage_split = 0.25)
}
\description{
Inferring trajectories with SCUBA
}
