% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_slice.R
\name{ti_slice}
\alias{ti_slice}
\title{Inferring trajectories with SLICE}
\usage{
ti_slice(distance_method = "euclidean", dimred_method = "none",
  clust.method = "max.distance", mutual = TRUE, max.distance = 0, k = 3,
  shortest.rank.percent = 10, graph.algorithm = "MST",
  graph.using.cells.clustering = FALSE, k_imc = 3,
  pct_leaf_node_cutoff = 0.5)
}
\arguments{
\item{k}{The number of cell clusters. If NULL, Gap statistic will be used to determine an optimal k.}
}
\description{
Inferring trajectories with SLICE
}
