% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_slicer.R, R/ti_container.R
\name{ti_slicer}
\alias{ti_slicer}
\alias{ti_slicer}
\title{Inferring trajectories with SLICER}
\usage{

  ti_slicer(run_environment = ifelse(is.null(getOption("dynwrap_run_environment")),
  "local", getOption("dynwrap_run_environment")), kmin = 10, m = 2)


  ti_slicer(run_environment = ifelse(is.null(getOption("dynwrap_run_environment")),
  "local", getOption("dynwrap_run_environment")), kmin = 10, m = 2)
}
\arguments{
\item{run_environment}{In which environment to run the method, can be 'local', 'docker' or 'singularity'}

\item{kmin}{Smallest value of k to try}

\item{m}{ intrinsic dimension of the data. This parameter mainly influences the visualisation of the results. The real intrinsic dimension will be calculated automaticly. }
}
\value{
The trajectory model
}
\description{
Inferring trajectories with SLICER

Will generate a trajectory using \href{https://doi.org/10.1186/s13059-016-0975-3}{SLICER}. This method was wrapped inside a \href{https://github.com/dynverse/dynmethods/tree/master/containers/slicer}{container}.
}
\details{
This methods was first wrapped inside R, see \link{ti_slicer}

The original code of this method is available \href{https://github.com/jw156605/SLICER}{here}.

The method is described in: \href{https://doi.org/10.1186/s13059-016-0975-3}{Welch, J.D., Hartemink, A.J., Prins, J.F., 2016. SLICER: inferring branched, nonlinear cellular trajectories from single cell RNA-seq data. Genome Biology 17.}
}
