% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_slingshot.R, R/ti_container.R
\name{ti_slingshot}
\alias{ti_slingshot}
\alias{ti_slingshot}
\title{Inferring trajectories with Slingshot}
\usage{
ti_slingshot(docker = dynwrap::test_docker_installation(), ndim = 3,
  nclus = 5, dimred = "pca", shrink = 1, reweight = TRUE,
  reassign = TRUE, thresh = 0.001, maxit = 10, stretch = 2,
  smoother = "smooth.spline", shrink.method = "cosine")

ti_slingshot(docker = dynwrap::test_docker_installation(), ndim = 3,
  nclus = 5, dimred = "pca", shrink = 1, reweight = TRUE,
  reassign = TRUE, thresh = 0.001, maxit = 10, stretch = 2,
  smoother = "smooth.spline", shrink.method = "cosine")
}
\arguments{
\item{docker}{Whether to use the docker container or the R wrapper}

\item{ndim}{The number of dimensions}

\item{nclus}{Number of clusters}

\item{dimred}{A character vector specifying which dimensionality reduction method to use.
See \link[dyndimred:dimred]{dyndimred::dimred} for the list of available dimensionality reduction methods.}

\item{shrink}{logical or numeric between 0 and 1, determines whether and how 
much to shrink branching lineages toward their average prior to the split.}

\item{reweight}{logical, whether to allow cells shared between lineages to be
reweighted during curve-fitting. If \code{TRUE}, cells shared between 
lineages will be iteratively reweighted based on the quantiles of their 
projection distances to each curve. See 'Details' for more.}

\item{reassign}{logical, whether to reassign cells to lineages at each 
iteration. If \code{TRUE}, cells will be added to a lineage when their 
projection distance to the curve is less than the median distance for all 
cells currently assigned to the lineage. Additionally, shared cells will be
removed from a lineage if their projection distance to the curve is above 
the 90th percentile and their weight along the curve is less than 
\code{0.1}.}

\item{thresh}{numeric, determines the convergence criterion. Percent change in the total distance from cells to their projections along curves must be less than thresh. Default is 0.001, similar to principal.curve.}

\item{maxit}{numeric, maximum number of iterations, see principal.curve.}

\item{stretch}{numeric factor by which curves can be extrapolated beyond endpoints. Default is 2, see principal.curve.}

\item{smoother}{choice of scatter plot smoother. Same as principal.curve, but "lowess" option is replaced with "loess" for additional flexibility.}

\item{shrink.method}{character denoting how to determine the appropriate 
amount of shrinkage for a branching lineage. Accepted values are the same 
as for \code{kernel} in \code{\link{density}} (default is \code{"cosine"}),
as well as \code{"tricube"} and \code{"density"}. See 'Details' for more.}
}
\value{
The trajectory model
}
\description{
Inferring trajectories with Slingshot

Will generate a trajectory using \href{https://doi.org/10.1186/s12864-018-4772-0}{Slingshot}. This method was wrapped inside a \href{https://github.com/dynverse/dynmethods/tree/master/containers/slingshot}{container}.
}
\details{
This methods was first wrapped inside R, see \link{ti_slingshot}

The original code of this method is available \href{https://github.com/kstreet13/slingshot}{here}.

The method is described in: \href{https://doi.org/10.1186/s12864-018-4772-0}{Street, K., Risso, D., Fletcher, R.B., Das, D., Ngai, J., Yosef, N., Purdom, E., Dudoit, S., 2018. Slingshot: cell lineage and pseudotime inference for single-cell transcriptomics. BMC Genomics 19.}
}
\seealso{
\code{\link[slingshot:slingshot]{slingshot::slingshot()}}
}
