% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_slingshot.R
\name{ti_slngsht}
\alias{ti_slngsht}
\title{Inferring trajectories with Slingshot}
\usage{
ti_slngsht(ndim = 3L, nclus = 5L, dimred = "pca", shrink = 1,
  reweight = TRUE, drop.multi = TRUE, thresh = 0.001, maxit = 10L,
  stretch = 2, smoother = "smooth.spline", shrink.method = "cosine")
}
\description{
Arguments passed to this function will be used as default parameters for the method.
}
