% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_stemid.R
\name{ti_stemid}
\alias{ti_stemid}
\title{Inferring trajectories with StemID}
\usage{
ti_stemid(clustnr = 30, bootnr = 50, metric = "pearson",
  num_cluster_method = "sat", SE.method = "Tibs2001SEmax",
  SE.factor = 0.25, B.gap = 50, cln = 30, FUNcluster = "kmedoids",
  dimred_method = "tsne", outminc = 0, outlg = 2, probthr = 0.001,
  thr_lower = -10, thr_upper = -5, outdistquant = 0.95, nmode = FALSE,
  pdishuf = 2000, pthr = 1e-04, pethr = 1e-04)
}
\description{
Inferring trajectories with StemID
}
