% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_container.R
\name{ti_stemnet}
\alias{ti_stemnet}
\title{Inferring a trajectory inference using \href{https://doi.org/10.1038/ncb3493}{STEMNET}}
\usage{
ti_stemnet(alpha = 0.1, lambda_auto = TRUE, lambda = 0.1)
}
\arguments{
\item{alpha}{The elastic net mixing parameter of the ‘glmnet’ classifier. \cr
numeric; default: 0.1; possible values between 0.001 and 10}

\item{lambda_auto}{Whether to select the lambda by cross-validation \cr}

\item{lambda}{The lambda penalty of GLM. \cr
numeric; default: 0.1; possible values between 0.05 and 1}
}
\value{
The trajectory model
}
\description{
Will generate a trajectory using \href{https://doi.org/10.1038/ncb3493}{STEMNET}. This method was wrapped inside a \href{https://github.com/dynverse/dynmethods/tree/master/containers/stemnet}{container}.
}
\details{
The original code of this method is available \href{https://git.embl.de/velten/STEMNET}{here}.

The method is described in: \href{https://doi.org/10.1038/ncb3493}{Velten, L., Haas, S.F., Raffel, S., Blaszkiewicz, S., Islam, S., Hennig, B.P., Hirche, C., Lutz, C., Buss, E.C., Nowak, D., Boch, T., Hofmann, W.-K., Ho, A.D., Huber, W., Trumpp, A., Essers, M.A.G., Steinmetz, L.M., 2017. Human haematopoietic stem cell lineage commitment is a continuous process. Nature Cell Biology 19, 271–281.}
}
