% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_container.R
\name{ti_wanderlust}
\alias{ti_wanderlust}
\title{Inferring a trajectory inference using \href{https://doi.org/10.1016/j.cell.2014.04.005}{wanderlust}}
\usage{
ti_wanderlust(branch = FALSE, epsilon = 1L, k = 15L, knn = 15L,
  n_diffusion_components = 2L, n_pca_components = 15L, normalise = TRUE,
  num_waypoints = 250L)
}
\arguments{
\item{branch}{Whether to allow a single bifurcation within the trajectory (wishbone versus wanderlust) \cr}

\item{epsilon}{Epsilon \cr
numeric; default: 1L; possible values between 0.1 and 10}

\item{k}{K parameter \cr
integer; default: 15L; possible values between 2 and 100}

\item{knn}{K-nearest neighbours for diffusion \cr
integer; default: 15L; possible values between 2 and 100}

\item{n_diffusion_components}{Number of diffusion components \cr
integer; default: 2L; possible values between 2 and 20}

\item{n_pca_components}{Number of pca components \cr
integer; default: 15L; possible values between 2 and 30}

\item{normalise}{\cr}

\item{num_waypoints}{Number of waypoints \cr
integer; default: 250L; possible values between 2 and 500}
}
\value{
The trajectory model
}
\description{
Will generate a trajectory using \href{https://doi.org/10.1016/j.cell.2014.04.005}{wanderlust}. This method was wrapped inside a \href{https://github.com/dynverse/dynmethods/tree/master/containers/wanderlust}{container}.
}
\details{
The original code of this method is available \href{https://www.c2b2.columbia.edu/danapeerlab/html/wanderlust.html}{here}.

The method is described in: \href{https://doi.org/10.1016/j.cell.2014.04.005}{Bendall, S.C., Davis, K.L., Amir, E.D., Tadmor, M.D., Simonds, E.F., Chen, T.J., Shenfeld, D.K., Nolan, G.P., Pe’er, D., 2014. Single-Cell Trajectory Detection Uncovers Progression and Regulatory Coordination in Human B Cell Development. Cell 157, 714–725.}
}
