% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_waterfall.R
\name{ti_waterfall}
\alias{ti_waterfall}
\title{Inferring trajectories with Waterfall}
\usage{
ti_waterfall(num_clusters = 10)
}
\arguments{
\item{num_clusters}{Number of clusters}
}
\description{
Inferring trajectories with Waterfall
}
