% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_waterfall.R
\name{ti_waterfll}
\alias{ti_waterfll}
\title{Inferring trajectories with Waterfall}
\usage{
ti_waterfll(num_clusters = 10L)
}
\description{
Arguments passed to this function will be used as default parameters for the method.
}
