% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_container.R
\name{ti_wishbone}
\alias{ti_wishbone}
\title{Inferring a trajectory inference using \href{https://doi.org/10.1038/nbt.3569}{Wishbone}}
\usage{
ti_wishbone(run_environment = NULL, normalise = TRUE, knn = 25L,
  n_diffusion_components = 3L, n_pca_components = 30L, k = 25L,
  num_waypoints = 250L, epsilon = 1L, branch = TRUE)
}
\arguments{
\item{run_environment}{In which environment to run the method, can be 'docker' or 'singularity'}

\item{normalise}{\cr}

\item{knn}{K-nearest neighbours for diffusion \cr
integer; default: 25L; possible values between 15 and 100}

\item{n_diffusion_components}{Number of diffusion components \cr
integer; default: 3L; possible values between 3 and 20}

\item{n_pca_components}{Number of pca components \cr
integer; default: 30L; possible values between 15 and 100}

\item{k}{K parameter \cr
integer; default: 25L; possible values between 15 and 100}

\item{num_waypoints}{Number of waypoints \cr
integer; default: 250L; possible values between 100 and 500}

\item{epsilon}{Epsilon \cr
numeric; default: 1L; possible values between 0.1 and 5}

\item{branch}{Whether to allow a single bifurcation within the trajectory (wishbone versus wanderlust) \cr}
}
\value{
The trajectory model
}
\description{
Will generate a trajectory using \href{https://doi.org/10.1038/nbt.3569}{Wishbone}. This method was wrapped inside a \href{https://github.com/dynverse/dynmethods/tree/master/containers/wishbone}{container}.
}
\details{
The original code of this method is available \href{https://github.com/ManuSetty/wishbone}{here}.

The method is described in: \href{https://doi.org/10.1038/nbt.3569}{Setty, M., Tadmor, M.D., Reich-Zeliger, S., Angel, O., Salame, T.M., Kathail, P., Choi, K., Bendall, S., Friedman, N., Pe’er, D., 2016. Wishbone identifies bifurcating developmental trajectories from single-cell data. Nature Biotechnology 34, 637–645.}
}
