% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_wot.R
\name{ti_wot}
\alias{ti_wot}
\title{Inferring a trajectory inference using Waddington-OT}
\usage{
ti_wot(param1 = 5L, param2 = "one", param3 = 0.5, param4 = TRUE,
  run_environment = NULL)
}
\arguments{
\item{param1}{integer; Description of param1 (default: \code{5L}; range: from \code{1L}
to \code{10L})}

\item{param2}{discrete; Description of param2 (default: \code{"one"}; values:
{\code{"one"}, \code{"two"}, \code{"three"}})}

\item{param3}{numeric; Description of param3 (default: \code{0.5}; range: from \code{0}
to \code{1})}

\item{param4}{logical; Description of param4}

\item{run_environment}{In which environment to run the method, can be \code{"docker"} or \code{"singularity"}.}
}
\value{
A TI method wrapper to be used together with
\code{\link[dynwrap:infer_trajectories]{infer_trajectory}}
}
\description{
Will generate a trajectory using
\href{https://doi.org/101010101/1101010101}{Waddington-OT}.

This method was wrapped inside a
\href{https://github.com/dynverse/dynmethods/tree/master/containers/wot}{container}.
The original code of this method is available
\href{https://github.com/dynverse/template}{here}.
}
