% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_wishbone.R
\name{description_wishbone}
\alias{description_wishbone}
\alias{description_wndrlst}
\title{Inferring trajectories with Wanderlust/Wishbone}
\usage{
description_wishbone(knn = 15L, n_diffusion_components = 2L,
  n_pca_components = 15L, branch = TRUE, k = 15L, num_waypoints = 250L,
  normalize = TRUE, epsilon = 1, method_name = "wishbone")

description_wndrlst(knn = 15L, n_diffusion_components = 2L,
  n_pca_components = 15L, branch = FALSE, k = 15L, num_waypoints = 250L,
  normalize = TRUE, epsilon = 1, method_name = "wndrlst")
}
\description{
Arguments passed to this function will be used as default parameters for the method.
}
