% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_prediction_model.R
\name{wrap_prediction_model}
\alias{wrap_prediction_model}
\title{Wrapping TI predictions}
\usage{
wrap_prediction_model(cell_ids, milestone_ids, milestone_network,
  milestone_percentages = NULL, progressions = NULL, ...)
}
\arguments{
\item{cell_ids}{the ids of the cells in the trajectory and counts}

\item{milestone_ids}{the ids of the milestones in the trajectory}

\item{milestone_network}{the network of the milestones}

\item{milestone_percentages}{what percentage of milestone is each cell}

\item{progressions}{what progression does a cell have}

\item{...}{extra information to be stored in the wrapper}
}
\description{
Wrapping TI predictions
}
