% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_prediction_model.R
\name{wrap_prediction_model_linear}
\alias{wrap_prediction_model_linear}
\title{Wrap a linear TI prediction}
\usage{
wrap_prediction_model_linear(cell_ids, pseudotimes, ...)
}
\arguments{
\item{cell_ids}{the ids of the cells in the trajectory and counts}

\item{pseudotimes}{The pseudotimes of the \code{cell_ids}}

\item{...}{extra information to be stored in the wrapper}
}
\description{
This function will generate the milestone_network and progressions.
}
