% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_celltree.R
\name{abstract_celltree_description}
\alias{abstract_celltree_description}
\alias{description_ctmaptpx}
\alias{description_ctgibbs}
\alias{description_ctvem}
\title{Inferring trajectories with \code{celltree}}
\usage{
abstract_celltree_description(method, num_topics_lower, num_topics_upper,
  num_topics, tot_iter, tolerance, sd_filter, absolute_width,
  width_scale_factor, outlier_tolerance_factor, rooting_method)

description_ctmaptpx(num_topics_lower = 2, num_topics_upper = 15,
  tot_iter = 1e+06, tolerance = 0.05, sd_filter = 0.5,
  absolute_width = 0, width_scale_factor = 1.5,
  outlier_tolerance_factor = 0.1, rooting_method = "null")

description_ctgibbs(num_topics = 4L, tot_iter = 200, tolerance = 1e-05,
  sd_filter = 0.5, absolute_width = 0, width_scale_factor = 1.5,
  outlier_tolerance_factor = 0.1, rooting_method = "null")

description_ctvem(num_topics = 4L, tot_iter = 1e+06, tolerance = 1e-05,
  sd_filter = 0.5, absolute_width = 0, width_scale_factor = 1.5,
  outlier_tolerance_factor = 0.1, rooting_method = "null")
}
\arguments{
\item{num_topics_lower}{The lower bound of topics to be fitted in the model.}

\item{num_topics_upper}{The upper bound of topics to be fitted in the model.}

\item{num_topics}{The number of topics to fit in the model.}

\item{tot_iter}{Number of iterations of the LDA inference.}

\item{tolerance}{Tolerance values of the LDA inference.}

\item{sd_filter}{Standard-deviation threshold below which genes should be removed from the data.}

\item{absolute_width}{Distance threshold below which a cell vertex is considered to be attached to a backbone vertex (see paper for more details).
By default, this threshold is computed dynamically, based on the distance distribution for each branch.}

\item{width_scale_factor}{A scaling factor for the dynamically-computed distance threshold (ignored if absolute_width is provided).
Higher values will result in less branches in the backbone tree, while lower values might lead to a large number of backbone branches.}

\item{rooting_method}{Method used to root the backbone tree. Must be one of: ‘null’, ‘longest.path’, ‘center.start.group’ or ‘average.start.group’.
‘longest.path' picks one end of the longest shortest-path between two vertices.
'center.start.group’ picks the vertex in the starting group with lowest mean-square-distance to the others.
‘average.start.group’ creates a new artificial vertex, as the average of all cells in the starting group.
‘null’ picks the best method based on the type of grouping and start group information available.}

\item{outlier_distance_factor}{Proportion of vertices, out of the total number of vertices divided by the total number of branches,
that can be left at the end of the backbone tree-building algorithm.}
}
\description{
Arguments passed to this function will be used as default parameters for the method.
}
