% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_install_github_tagged_version.R
\name{install_github_tagged_version}
\alias{install_github_tagged_version}
\title{Install remotes using git tags and R versions}
\usage{
install_github_tagged_version(remotes, is_interactive = interactive())
}
\description{
Install remotes using git tags and R versions
}
