% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{methods}
\alias{methods}
\title{Metadata on the different TI methods}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 47 rows and 26 columns.}
\usage{
methods
}
\description{
Metadata on the different TI methods
}
\keyword{datasets}
