% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_monocle.R
\name{abstract_monocle_description}
\alias{abstract_monocle_description}
\alias{description_mnclddr}
\alias{description_mnclica}
\title{Inferring trajectories with Monocle}
\usage{
abstract_monocle_description(short_name)

description_mnclddr(reduction_method = "DDRTree", max_components = 2L,
  norm_method = "vstExprs", auto_param_selection = TRUE)

description_mnclica(reduction_method = "ICA", max_components = 2L,
  norm_method = "vstExprs")
}
\description{
Arguments passed to this function will be used as default parameters for the method.
}
