% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_monocle.R
\name{ti_mnclddr}
\alias{ti_mnclddr}
\alias{ti_mnclica}
\title{Inferring trajectories with Monocle}
\usage{
ti_mnclddr(reduction_method = "DDRTree", max_components = 2L,
  norm_method = "vstExprs", auto_param_selection = TRUE)

ti_mnclica(reduction_method = "ICA", max_components = 2L,
  norm_method = "vstExprs")
}
\description{
Arguments passed to this function will be used as default parameters for the method.
}
