% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_paga.R
\name{abstract_paga_description}
\alias{abstract_paga_description}
\alias{description_paga}
\title{Inferring trajectories with PAGA}
\usage{
abstract_paga_description(method)

description_paga(n_neighbours = 30L, n_comps = 50L, resolution = 2.5)
}
\description{
Arguments passed to this function will be used as default parameters for the method.
}
