% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_angle.R
\name{ti_angle}
\alias{ti_angle}
\title{Inferring a trajectory inference using Angle}
\usage{
ti_angle(dimred = "pca")
}
\arguments{
\item{dimred}{discrete; Which dimensionality reduction method to use. (default:
\code{"pca"}; values: {\code{"pca"}, \code{"mds"}, \code{"tsne"}, \code{"ica"}, \code{"lle"},
\code{"landmark_mds"}, \code{"mds_sammon"}, \code{"mds_isomds"}, \code{"mds_smacof"}, \code{"umap"},
\code{"dm_diffusionMap"}})}
}
\value{
A TI method wrapper to be used together with
\code{\link[dynwrap:infer_trajectories]{infer_trajectory}}
}
\description{
Will generate a trajectory using Angle.

This method was wrapped inside a
\href{https://github.com/dynverse/ti_angle}{container}.
}
