% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_angle.R, R/ti_container.R
\name{ti_angle}
\alias{ti_angle}
\alias{ti_angle}
\title{Inferring trajectories with Angle}
\usage{
ti_angle(docker = dynwrap::test_docker_installation(), dimred = "pca")

ti_angle(docker = dynwrap::test_docker_installation(), dimred = "pca")
}
\arguments{
\item{docker}{Whether to use the docker container or the R wrapper}

\item{dimred}{A character vector specifying which dimensionality reduction method to use.
See \code{\link[dyndimred:dimred]{dyndimred::dimred()}} for the list of available dimensionality reduction methods.}
}
\value{
The trajectory model
}
\description{
Inferring trajectories with Angle

Will generate a trajectory using Angle. This method was wrapped inside a \href{https://github.com/dynverse/dynmethods/tree/master/containers/angle}{container}.
}
\details{
This methods was first wrapped inside R, see \link{ti_angle}
}
