% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_angle.R, R/ti_container.R
\name{ti_angle}
\alias{ti_angle}
\alias{ti_angle}
\title{Inferring trajectories with Angle}
\usage{

  ti_angle(run_environment = ifelse(is.null(getOption("dynwrap_run_environment")),
  "local", getOption("dynwrap_run_environment")), dimred = "pca")


  ti_angle(run_environment = ifelse(is.null(getOption("dynwrap_run_environment")),
  "local", getOption("dynwrap_run_environment")), dimred = "pca")
}
\arguments{
\item{run_environment}{In which environment to run the method, can be 'local', 'docker' or 'singularity'}

\item{dimred}{A character vector specifying which dimensionality reduction method to use.
See \link[dyndimred:dimred]{dyndimred::dimred} for the list of available dimensionality reduction methods.}
}
\value{
The trajectory model
}
\description{
Inferring trajectories with Angle

Will generate a trajectory using Angle. This method was wrapped inside a \href{https://github.com/dynverse/dynmethods/tree/master/containers/angle}{container}.
}
\details{
This methods was first wrapped inside R, see \link{ti_angle}
}
