% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_angle.R
\name{ti_angle}
\alias{ti_angle}
\title{Inferring a trajectory inference using Angle}
\usage{
ti_angle(dimred = "pca")
}
\arguments{
\item{dimred}{Which dimensionality reduction method to use. Domain: {pca, mds,
tsne, ica, lle, landmark_mds, mds_sammon, mds_isomds, mds_smacof, umap,
dm_diffusionMap, dm_destiny}. Default: pca. Format: character.}
}
\value{
A TI method wrapper to be used together with
\code{\link[dynwrap:infer_trajectories]{infer_trajectory}}
}
\description{
Will generate a trajectory using Angle.

This method was wrapped inside a
\href{https://github.com/dynverse/ti_angle}{container}.
}
