% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_calista.R
\name{ti_calista}
\alias{ti_calista}
\title{Inferring a trajectory inference using CALISTA}
\usage{
ti_calista(runs = 50L, max_iter = 100L,
  config = dynwrap::container_config())
}
\arguments{
\item{runs}{integer; Number of independent runs of greedy algorithm (default:
\code{50L}; range: from \code{20L} to \code{100L})}

\item{max_iter}{integer; Number of iterations in greedy algorithm (default:
\code{100L}; range: from \code{20L} to \code{400L})}

\item{config}{A container config. See \code{\link[=container_config]{container_config()}} for more information.}
}
\value{
A TI method wrapper to be used together with
\code{\link[dynwrap:infer_trajectories]{infer_trajectory}}
}
\description{
Will generate a trajectory using \href{https://doi.org/10.1101/257550}{CALISTA}.

This method was wrapped inside a
\href{https://github.com/dynverse/ti_calista}{container}.
The original code of this method is available
\href{https://github.com/CABSEL/CALISTA}{here}.
}
\references{
Papili Gao, N., Hartmann, T., Gunawan, R., 2018. CALISTA:
Clustering And Lineage Inference in Single-Cell Transcriptional Analysis.
}
