% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_cellrouter.R
\name{ti_cellrouter}
\alias{ti_cellrouter}
\title{Inferring a trajectory inference using CellRouter}
\usage{
ti_cellrouter(ndim_pca = 20L, ndim_tsne = 11L, max_iter = 1000L,
  cluster_method = "graph.clustering", k_clustering = 20L,
  ndim_pca_clustering = 20L, k_knn = 10L, ndim_pca_knn = 20L,
  sim_type = "jaccard", distance_method_paths = "graph",
  ranks = "rank", num_cells = 3L, neighs = 3L, perplexity = 30L)
}
\arguments{
\item{ndim_pca}{integer; Number of principal components to compute (default:
\code{20L}; range: from \code{2L} to \code{100L})}

\item{ndim_tsne}{integer; Number of tsne dimensions to compute (default: \code{11L};
range: from \code{2L} to \code{100L})}

\item{max_iter}{integer; Maximal number of tsne iterations (default: \code{1000L};
range: from \code{100L} to \code{100000L})}

\item{cluster_method}{discrete; Method to use for clustering (default:
\code{"graph.clustering"}; values: {\code{"graph.clustering"}, \code{"model.clustering"}})}

\item{k_clustering}{integer; Number of nearest neighbors to build a k-nearest
neighbors graph for clustering (default: \code{20L}; range: from \code{2L} to \code{1000L})}

\item{ndim_pca_clustering}{integer; Number of PCA dimensions used for k-nearest
neighbors graph for clustering (default: \code{20L}; range: from \code{2L} to \code{100L})}

\item{k_knn}{integer; Number of nearest neighbors to build a k-nearest
neighbors graph for knn (default: \code{10L}; range: from \code{2L} to \code{1000L})}

\item{ndim_pca_knn}{integer; Number of PCA dimensions used for knn (default:
\code{20L}; range: from \code{2L} to \code{100L})}

\item{sim_type}{discrete; Similarity type for knn (default: \code{"jaccard"};
values: {\code{"jaccard"}})}

\item{distance_method_paths}{discrete; Distance method for paths (default:
\code{"graph"}; values: {\code{"euclidean"}, \code{"maximum"}, \code{"manhattan"}, \code{"canberra"},
\code{"binary"}, \code{"graph"}})}

\item{ranks}{discrete; How to rank the paths (default: \code{"rank"}; values:
{\code{"path_cost"}, \code{"path_flow"}, \code{"rank"}, \code{"length"}})}

\item{num_cells}{integer; Trajectories should contain at least num.cells
(default: \code{3L}; range: from \code{3L} to \code{100L})}

\item{neighs}{integer; The size of the neighborhood in kNN graph used to
smoothen kinetic profiles (default: \code{3L}; range: from \code{2L} to \code{100L})}

\item{perplexity}{numeric; Perplexity parameter for tsne (default: \code{30L};
range: from \code{5L} to \code{100L})}
}
\value{
A TI method wrapper to be used together with
\code{\link[dynwrap:infer_trajectories]{infer_trajectory}}
}
\description{
Will generate a trajectory using
\href{https://doi.org/10.1038/s41467-018-03214-y}{CellRouter}.

This method was wrapped inside a
\href{https://github.com/dynverse/ti_cellrouter}{container}.
The original code of this method is available
\href{https://github.com/edroaldo/cellrouter}{here}.
}
\references{
Lummertz da Rocha, E., Rowe, R.G., Lundin, V., Malleshaiah, M.,
Jha, D.K., Rambo, C.R., Li, H., North, T.E., Collins, J.J., Daley, G.Q., 2018.
Reconstruction of complex single-cell trajectories using CellRouter. Nature
Communications 9.
}
