% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_celltrails.R
\name{ti_celltrails}
\alias{ti_celltrails}
\title{CellTrails}
\usage{
ti_celltrails(filter_features = FALSE, threshold_dl = 2L,
  threshold_cov = 0.05, threshold_ff = 1.7, min_expr = 0L,
  frac = 100L, min_size = 0.01, min_feat = 5L, max_pval = 1e-04,
  min_fc = 2L, l = 10L)
}
\arguments{
\item{filter_features}{Whether to include celltrails feature filtering.
Default: FALSE. Format: logical.}

\item{threshold_dl}{Minimum number of samples; if value < 1 it is interpreted
as fraction, otherwise as absolute sample count. Domain: U(0, 100). Default: 2.
Format: integer.}

\item{threshold_cov}{Minimum coefficient of variation; numeric value between 0
and 1. Domain: U(0, 1). Default: 0.05. Format: numeric.}

\item{threshold_ff}{A Z-score cutoff. Domain: U(0, 5). Default: 1.7. Format:
numeric.}

\item{min_expr}{Minimum average feature expression. Domain: U(0, 2). Default:
0. Format: numeric.}

\item{frac}{Fraction or number (if frac > 1) of eigengaps used to perform
linear fit. Domain: e^U(0.00, 6.91). Default: 100. Format: numeric.}

\item{min_size}{The initial cluster dedrogram is cut at an height such that the
minimum cluster size is at least min_size; if min_size < 1 than the fraction of
total samples is used, otherwise it is used as absoulte count. Domain:
e^U(-6.91, 0.00). Default: 0.01. Format: numeric.}

\item{min_feat}{Minimum number of differentially expressed features between
siblings. If this number is not reached, two neighboring clusters (siblings) in
the pruned dendrogram get joined. Domain: U(1, 100). Default: 5. Format:
integer.}

\item{max_pval}{Maximum P-value for differential expression computation.
Domain: e^U(-16.12, 0.00). Default: 1e-04. Format: numeric.}

\item{min_fc}{Mimimum fold-change for differential expression computation.
Domain: U(0, 5). Default: 2. Format: numeric.}

\item{l}{Neighborhood size. Domain: U(1, 50). Default: 10. Format: integer.}
}
\value{
A TI method wrapper to be used together with
\code{\link[dynwrap:infer_trajectories]{infer_trajectory}}
}
\description{
Will generate a trajectory using
\href{https://doi.org/10.1016/j.celrep.2018.05.002}{CellTrails}.

This method was wrapped inside a
\href{https://github.com/dynverse/ti_celltrails}{container}.
The original code of this method is available
\href{https://github.com/dcellwanger/CellTrails}{here}.
}
\references{
Ellwanger, D.C., Scheibinger, M., Dumont, R.A., Barr-Gillespie,
P.G., Heller, S., 2018. Transcriptional Dynamics of Hair-Bundle Morphogenesis
Revealed with CellTrails. Cell Reports 23, 2901–2914.e13.
}
\keyword{method}
