% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_celltrails.R
\name{ti_celltrails}
\alias{ti_celltrails}
\title{Inferring a trajectory inference using CellTrails}
\usage{
ti_celltrails(threshold_dl = 2L, threshold_cov = 0.05,
  threshold_ff = 1.7, min_expr = 0L, frac = 100L, min_size = 0.01,
  min_feat = 5L, max_pval = 1e-04, min_fc = 2L, l = 10L,
  config = dynwrap::container_config())
}
\arguments{
\item{threshold_dl}{integer; Minimum number of samples; if value < 1 it is
interpreted as fraction, otherwise as absolute sample count (default: \code{2L};
range: from \code{0L} to \code{100L})}

\item{threshold_cov}{numeric; Minimum coefficient of variation; numeric value
between 0 and 1 (default: \code{0.05}; range: from \code{0L} to \code{1L})}

\item{threshold_ff}{numeric; A Z-score cutoff (default: \code{1.7}; range: from \code{0L}
to \code{5L})}

\item{min_expr}{numeric; Minimum average feature expression (default: \code{0L};
range: from \code{0L} to \code{2L})}

\item{frac}{numeric; Fraction or number (if frac > 1) of eigengaps used to
perform linear fit. (default: \code{100L}; range: from \code{1L} to \code{1000L})}

\item{min_size}{numeric; The initial cluster dedrogram is cut at an height such
that the minimum cluster size is at least min_size; if min_size < 1 than the
fraction of total samples is used, otherwise it is used as absoulte count
(default: \code{0.01}; range: from \code{0.001} to \code{1L})}

\item{min_feat}{integer; Minimum number of differentially expressed features
between siblings. If this number is not reached, two neighboring clusters
(siblings) in the pruned dendrogram get joined. (default: \code{5L}; range: from
\code{1L} to \code{100L})}

\item{max_pval}{numeric; Maximum P-value for differential expression
computation. (default: \code{1e-04}; range: from \code{1e-07} to \code{1L})}

\item{min_fc}{numeric; Mimimum fold-change for differential expression
computation (default: \code{2L}; range: from \code{0L} to \code{5L})}

\item{l}{integer; Neighborhood size (default: \code{10L}; range: from \code{1L} to \code{50L})}

\item{config}{A container config. See \code{\link[=container_config]{container_config()}} for more information.}
}
\value{
A TI method wrapper to be used together with
\code{\link[dynwrap:infer_trajectories]{infer_trajectory}}
}
\description{
Will generate a trajectory using
\href{https://doi.org/10.1016/j.celrep.2018.05.002}{CellTrails}.

This method was wrapped inside a
\href{https://github.com/dynverse/ti_celltrails}{container}.
The original code of this method is available
\href{https://github.com/dcellwanger/CellTrails}{here}.
}
\references{
Ellwanger, D.C., Scheibinger, M., Dumont, R.A., Barr-Gillespie,
P.G., Heller, S., 2018. Transcriptional Dynamics of Hair-Bundle Morphogenesis
Revealed with CellTrails. Cell Reports 23, 2901–2914.e14.
}
