% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_celltree_maptpx.R
\name{ti_celltree_maptpx}
\alias{ti_celltree_maptpx}
\title{Inferring a trajectory inference using cellTree with maptpx}
\usage{
ti_celltree_maptpx(method = "maptpx", sd_filter = 0.5,
  absolute_width = 0, width_scale_factor = 1.5,
  outlier_tolerance_factor = 0.1, rooting_method = "null",
  num_topics_lower = 2, num_topics_upper = 15, tot_iter = 1e+06,
  tolerance = 0.05, run_environment = NULL)
}
\arguments{
\item{method}{discrete; (default: \code{"maptpx"}; values: {\code{"maptpx"}})}

\item{sd_filter}{numeric; (default: \code{0.5}; range: from \code{0.01} to \code{5})}

\item{absolute_width}{numeric; (default: \code{0}; range: from \code{NULL} to \code{NULL})}

\item{width_scale_factor}{numeric; (default: \code{1.5}; range: from \code{0.1} to \code{100})}

\item{outlier_tolerance_factor}{numeric; (default: \code{0.1}; range: from \code{1e-04}
to \code{1000})}

\item{rooting_method}{discrete; (default: \code{"null"}; values: {\code{"longest.path"},
\code{"center.start.group"}, \code{"average.start.group"}, \code{"null"}})}

\item{num_topics_lower}{integer; (default: \code{2}; range: from \code{2} to \code{15})}

\item{num_topics_upper}{integer; (default: \code{15}; range: from \code{2} to \code{15})}

\item{tot_iter}{numeric; (default: \code{1e+06}; range: from \code{10000} to \code{1e+07})}

\item{tolerance}{numeric; (default: \code{0.05}; range: from \code{0.001} to \code{0.5})}

\item{run_environment}{In which environment to run the method, can be \code{"docker"} or \code{"singularity"}.}
}
\value{
A TI method wrapper to be used together with
\code{\link[dynwrap:infer_trajectories]{infer_trajectory}}
}
\description{
Will generate a trajectory using \href{https://doi.org/10.1186/s12859-016-1175-6}{cellTree with maptpx}.

This method was wrapped inside a
\href{https://github.com/dynverse/dynmethods/tree/master/containers/celltree_maptpx}{container}.
The original code of this method is available
\href{https://bioconductor.org/packages/release/bioc/html/cellTree.html}{here}.
}
\references{
duVerle, D.A., Yotsukura, S., Nomura, S., Aburatani, H., Tsuda, K.,
2016. CellTree: an R/bioconductor package to infer the hierarchical structure
of cell populations from single-cell RNA-seq data. BMC Bioinformatics 17.
}
