% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_comp1.R
\name{ti_comp1}
\alias{ti_comp1}
\title{Component 1}
\usage{
ti_comp1(dimred = "pca", ndim = 2L, component = 1L)
}
\arguments{
\item{dimred}{Which dimensionality reduction method to use. Domain: {pca, mds,
tsne, ica, lle, landmark_mds, mds_sammon, mds_isomds, mds_smacof, umap,
dm_diffusionMap}. Default: pca. Format: character.}

\item{ndim}{. Domain: U(2, 30). Default: 2. Format: integer.}

\item{component}{. Domain: U(1, 10). Default: 1. Format: integer.}
}
\value{
A TI method wrapper to be used together with
\code{\link[dynwrap:infer_trajectories]{infer_trajectory}}
}
\description{
Will generate a trajectory using Component 1.

This method was wrapped inside a
\href{https://github.com/dynverse/ti_comp1}{container}.
}
\keyword{method}
