% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_comp1.R
\name{ti_comp1}
\alias{ti_comp1}
\title{Inferring a trajectory inference using Component 1}
\usage{
ti_comp1(dimred = "pca", ndim = 2, component = 1,
  container_type = NULL)
}
\arguments{
\item{dimred}{discrete; Which dimensionality reduction method to use. (default:
\code{"pca"}; values: {\code{"pca"}, \code{"mds"}, \code{"tsne"}, \code{"ica"}, \code{"lle"},
\code{"landmark_mds"}, \code{"mds_sammon"}, \code{"mds_isomds"}, \code{"mds_smacof"}, \code{"umap"},
\code{"dm_diffusionMap"}})}

\item{ndim}{integer; (default: \code{2}; range: from \code{2} to \code{30})}

\item{component}{integer; (default: \code{1}; range: from \code{1} to \code{10})}

\item{container_type}{In which environment to run the method, can be \code{"docker"} or \code{"singularity"}.
By default, docker will be used.}
}
\value{
A TI method wrapper to be used together with
\code{\link[dynwrap:infer_trajectories]{infer_trajectory}}
}
\description{
Will generate a trajectory using Component 1.

This method was wrapped inside a
\href{https://github.com/dynverse/dynmethods/tree/master/containers/comp1}{container}.
}
