% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_container.R
\name{ti_comp1}
\alias{ti_comp1}
\title{Inferring a trajectory inference using Component 1}
\usage{
ti_comp1(dimred = "pca", ndim = 2, component = 1)
}
\arguments{
\item{dimred}{\cr
discrete; default: "pca"; possible values: pca, mds, tsne, ica, lle, landmark_mds, mds_sammon, mds_isomds, mds_smacof, umap, dm_diffusionMap}

\item{ndim}{\cr
integer; default: 2; possible values between 2 and 30}

\item{component}{\cr
integer; default: 1; possible values between 1 and 10}
}
\value{
The trajectory model
}
\description{
Will generate a trajectory using Component 1. This method was wrapped inside a \href{https://github.com/dynverse/dynmethods/tree/master/containers/comp1}{container}.
}
