% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_controls.R, R/ti_container.R
\name{ti_comp1}
\alias{ti_comp1}
\alias{ti_comp1}
\title{Inferring a trajectory inference using Component 1}
\usage{

  ti_comp1(run_environment = ifelse(is.null(getOption("dynwrap_run_environment")),
  "local", getOption("dynwrap_run_environment")), dimred = "pca", ndim = 2,
  component = 1)


  ti_comp1(run_environment = ifelse(is.null(getOption("dynwrap_run_environment")),
  "local", getOption("dynwrap_run_environment")), dimred = "pca", ndim = 2,
  component = 1)
}
\arguments{
\item{run_environment}{In which environment to run the method, can be 'local', 'docker' or 'singularity'}

\item{dimred}{A character vector specifying which dimensionality reduction method to use.
See \link[dyndimred:dimred]{dyndimred::dimred} for the list of available dimensionality reduction methods.}

\item{ndim}{The number of dimensions}

\item{component}{The component to use}
}
\value{
The trajectory model
}
\description{
Will generate a trajectory using Component 1. This method was wrapped inside a \href{https://github.com/dynverse/dynmethods/tree/master/containers/comp1}{container}.
}
\details{
This methods was first wrapped inside R, see \link{ti_comp1}
}
