% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_dpt.R
\name{ti_dpt}
\alias{ti_dpt}
\title{DPT}
\usage{
ti_dpt(sigma = "local", distance = "euclidean", ndim = 20L,
  density_norm = TRUE, n_local = c(5L, 7L), w_width = 0.1)
}
\arguments{
\item{sigma}{Diffusion scale parameter of the Gaussian kernel. A larger sigma
might be necessary if the eigenvalues can not be found because of a singularity
in the matrix. Must a character vector -- \code{"local"} (default) or \code{"global"}.
Domain: {local, global}. Default: local. Format: character.}

\item{distance}{A character vector specifying which distance metric to use.
Allowed measures are the Euclidean distance (default), the cosine distance
(\code{1-corr(c_1, c_2)}), or the rank correlation distance (\code{1-corr(rank(c_1), rank(c_2))}). Domain: {euclidean, cosine, rankcor}. Default: euclidean. Format:
character.}

\item{ndim}{Number of eigenvectors/dimensions to return. Domain: U(3, 100).
Default: 20. Format: integer.}

\item{density_norm}{Logical. If TRUE, use density normalisation. Default: TRUE.
Format: logical.}

\item{n_local}{If sigma == 'local', the \code{n_local} nearest neighbor(s)
determine(s) the local sigma. Domain: ( U(2, 20), U(2, 20) ). Default: (5, 7).
Format: integer_range.}

\item{w_width}{Window width to use for deciding the branch cutoff. Domain:
e^U(-9.21, 0.00). Default: 0.1. Format: numeric.}
}
\value{
A TI method wrapper to be used together with
\code{\link[dynwrap:infer_trajectories]{infer_trajectory}}
}
\description{
Will generate a trajectory using \href{https://doi.org/10.1038/nmeth.3971}{DPT}.

This method was wrapped inside a
\href{https://github.com/dynverse/ti_dpt}{container}.
The original code of this method is available
\href{https://bioconductor.org/packages/release/bioc/html/destiny.html}{here}.
}
\references{
Haghverdi, L., Büttner, M., Wolf, F.A., Buettner, F., Theis, F.J.,
2016. Diffusion pseudotime robustly reconstructs lineage branching. Nature
Methods 13, 845–848.
}
\keyword{method}
