% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_container.R
\name{ti_elpicycle}
\alias{ti_elpicycle}
\title{Inferring a trajectory inference using \href{https://doi.org/https://github.com/Albluca/ElPiGraph.R}{ElPiGraph cyclic}}
\usage{
ti_elpicycle(run_environment = NULL, topology = "cycle", NumNodes = 50L,
  NumEdges = 100000L, InitNodes = 2L, Mu = 0.1, Lambda = 0.01,
  MaxNumberOfIterations = 10L, eps = 0.01, CenterData = FALSE)
}
\arguments{
\item{run_environment}{In which environment to run the method, can be 'docker' or 'singularity'}

\item{topology}{The kind of topology to detect \cr
discrete; default: "cycle"; possible values: cycle}

\item{NumNodes}{The number of nodes of the principal graph \cr
integer; default: 50L; possible values between 2 and 1000}

\item{NumEdges}{The maximum number of edges \cr
integer; default: 100000L; possible values between 2 and 100000}

\item{InitNodes}{Number of points to include in the initial graph \cr
integer; default: 2L; possible values between 2 and 1000}

\item{Mu}{Controls the elastic energy \cr
numeric; default: 0.1; possible values between 0.001 and 1}

\item{Lambda}{Controls the elastic energy \cr
numeric; default: 0.01; possible values between 0.001 and 1}

\item{MaxNumberOfIterations}{Maximum number of steps to embed the nodes \cr
integer; default: 10L; possible values between 1 and 1000}

\item{eps}{Minimal relative change in the position of the nodes to stop embedment \cr
numeric; default: 0.01; possible values between 0.001 and 1}

\item{CenterData}{Should data and initial node positions be centered? \cr}
}
\value{
The trajectory model
}
\description{
Will generate a trajectory using \href{https://doi.org/https://github.com/Albluca/ElPiGraph.R}{ElPiGraph cyclic}. This method was wrapped inside a \href{https://github.com/dynverse/dynmethods/tree/master/containers/elpicycle}{container}.
}
\details{
The original code of this method is available \href{https://github.com/Albluca/ElPiGraph.R}{here}.
}
