% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_embeddr.R, R/ti_container.R
\name{ti_embeddr}
\alias{ti_embeddr}
\alias{ti_embeddr}
\title{Inferring trajectories with Embeddr}
\usage{
ti_embeddr(docker = dynwrap::test_docker_installation(), ndim = 2,
  kernel = "nn", metric = "correlation", nn_pct = 0, eps = 0, t = 0,
  symmetrize = "mean", measure_type = "unorm", thresh = 0.001,
  maxit = 10, stretch = 2, smoother = "smooth.spline")

ti_embeddr(docker = dynwrap::test_docker_installation(), ndim = 2,
  kernel = "nn", metric = "correlation", nn_pct = 0, eps = 0, t = 0,
  symmetrize = "mean", measure_type = "unorm", thresh = 0.001,
  maxit = 10, stretch = 2, smoother = "smooth.spline")
}
\arguments{
\item{docker}{Whether to use the docker container or the R wrapper}

\item{ndim}{Dimension of the embedded space, default is 2}

\item{kernel}{The choice of kernel. 'nn' will give nearest neighbours, 'dist' gives minimum distance and
'heat' gives a heat kernel. Discussed in detail in 'Laplacian Eigenmaps and Spectral Techniques for Embedding and Clustering',
Belkin & Niyogi}

\item{metric}{The metric with which to assess 'closeness' for nearest neighbour selection, one of
'correlation' (pearson) or 'euclidean'. Default is 'correlation'.}

\item{nn_pct}{The percentage of cells to use as tge number of nearest neighbours if kernel == 'nn'.}

\item{eps}{Maximum distance parameter if kernel == 'dist'}

\item{t}{'time' for heat kernel if kernel == 'heat'}

\item{symmetrize}{How to make the adjacency matrix symmetric. Note that slightly
counterintuitively, node i having node j as a nearest neighbour doesn't guarantee node
j has node i. There are several ways to get round this:
\itemize{
\item \code{mean} If the above case occurs make the link weight 0.5 so the adjacency matrix becomes \eqn{0.5(A + A')}
\item \code{ceil} If the above case occurs set the link weight to 1 (ie take the ceiling of the mean case)
\item \code{floor} If the above case occurs set the link weight to 0 (ie take the floor of the mean case)
}}

\item{measure_type}{Type of laplacian eigenmap, which corresponds to the constraint on the eigenvalue problem. If
type is 'unorm' (default), then the graph measure used is the identity matrix, while if type is 'norm' then the measure
used is the degree matrix.}

\item{thresh}{convergence threshold on shortest distances to the curve.}

\item{maxit}{maximum number of iterations.}

\item{stretch}{a factor by which the curve can be extrapolated when
points are projected.  Default is 2 (times the last segment
length). The default is 0 for \code{smoother} equal to
\code{"periodic_lowess"}.}

\item{smoother}{choice of smoother. The default is
\code{"smooth_spline"}, and other choices are \code{"lowess"} and
\code{"periodic_lowess"}. The latter allows one to fit closed curves.
Beware, you may want to use \code{iter = 0} with \code{lowess()}.}
}
\value{
The trajectory model
}
\description{
Inferring trajectories with Embeddr

Will generate a trajectory using \href{https://doi.org/10.1101/027219}{Embeddr}. This method was wrapped inside a \href{https://github.com/dynverse/dynmethods/tree/master/containers/embeddr}{container}.
}
\details{
This methods was first wrapped inside R, see \link{ti_embeddr}

The original code of this method is available \href{https://github.com/kieranrcampbell/embeddr}{here}.

The method is described in: \href{https://doi.org/10.1101/027219}{Campbell, K., Ponting, C.P., Webber, C., 2015. Laplacian eigenmaps and principal curves for high resolution pseudotemporal ordering of single-cell RNA-seq profiles.}
}
