% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_embeddr.R
\name{ti_embeddr}
\alias{ti_embeddr}
\title{Inferring a trajectory inference using Embeddr}
\usage{
ti_embeddr(ndim = 2L, kernel = "nn", metric = "correlation",
  nn_pct = 0, eps = 0, t = 0, symmetrize = "mean",
  measure_type = "unorm", thresh = 0.001, maxit = 10L, stretch = 2,
  smoother = "smooth.spline")
}
\arguments{
\item{ndim}{integer; Dimension of the embedded space, default is 2 (default:
\code{2L}; range: from \code{2L} to \code{10L})}

\item{kernel}{discrete; The choice of kernel. 'nn' will give nearest
neighbours, 'dist' gives minimum distance and'heat' gives a heat kernel.
Discussed in detail in 'Laplacian Eigenmaps and Spectral Techniques for
Embedding and Clustering',Belkin & Niyogi (default: \code{"nn"}; values: {\code{"nn"},
\code{"dist"}, \code{"heat"}})}

\item{metric}{discrete; The metric with which to assess 'closeness' for nearest
neighbour selection, one of'correlation' (pearson) or 'euclidean'. Default is
'correlation'. (default: \code{"correlation"}; values: {\code{"correlation"},
\code{"euclidean"}, \code{"cosine"}})}

\item{nn_pct}{numeric; The percentage of cells to use as tge number of nearest
neighbours if kernel == 'nn'. (default: \code{0}; range: from \code{-2} to \code{1})}

\item{eps}{numeric; Maximum distance parameter if kernel == 'dist' (default:
\code{0}; range: from \code{-5} to \code{5})}

\item{t}{numeric; 'time' for heat kernel if kernel == 'heat' (default: \code{0};
range: from \code{-5} to \code{5})}

\item{symmetrize}{discrete; How to make the adjacency matrix symmetric. Note
that slightlycounterintuitively, node i having node j as a nearest neighbour
doesn't guarantee nodej has node i. There are several ways to get round
this:\itemize{\item \code{mean} If the above case occurs make the link weight
0.5 so the adjacency matrix becomes \eqn{0.5(A + A')}\item \code{ceil} If the
above case occurs set the link weight to 1 (ie take the ceiling of the mean
case)\item \code{floor} If the above case occurs set the link weight to 0 (ie
take the floor of the mean case)} (default: \code{"mean"}; values: {\code{"mean"},
\code{"ceil"}, \code{"floor"}})}

\item{measure_type}{discrete; Type of laplacian eigenmap, which corresponds to
the constraint on the eigenvalue problem. Iftype is 'unorm' (default), then the
graph measure used is the identity matrix, while if type is 'norm' then the
measureused is the degree matrix. (default: \code{"unorm"}; values: {\code{"unorm"},
\code{"norm"}})}

\item{thresh}{numeric; Convergence threshold on shortest distances to the
curve. (default: \code{0.001}; range: from \code{1e-05} to \code{1e+05})}

\item{maxit}{integer; Maximum number of iterations. (default: \code{10L}; range:
from \code{0L} to \code{50L})}

\item{stretch}{numeric; A factor by which the curve can be extrapolated
whenpoints are projected.  Default is 2 (times the last segmentlength). The
default is 0 for \code{smoother} equal to\code{"periodic_lowess"}. (default:
\code{2}; range: from \code{0} to \code{5})}

\item{smoother}{discrete; Choice of smoother. The default
is\code{"smooth_spline"}, and other choices are \code{"lowess"}
and\code{"periodic_lowess"}. The latter allows one to fit closed curves.Beware,
you may want to use \code{iter = 0} with \code{lowess()}. (default:
\code{"smooth.spline"}; values: {\code{"smooth.spline"}, \code{"lowess"},
\code{"periodic.lowess"}})}
}
\value{
A TI method wrapper to be used together with
\code{\link[dynwrap:infer_trajectories]{infer_trajectory}}
}
\description{
Will generate a trajectory using \href{https://doi.org/10.1101/027219}{Embeddr}.

This method was wrapped inside a
\href{https://github.com/dynverse/ti_embeddr}{container}.
The original code of this method is available
\href{https://github.com/kieranrcampbell/embeddr}{here}.
}
\references{
Campbell, K., Ponting, C.P., Webber, C., 2015. Laplacian eigenmaps
and principal curves for high resolution pseudotemporal ordering of single-cell
RNA-seq profiles.
}
