% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_error.R
\name{ti_error}
\alias{ti_error}
\title{Inferring a trajectory inference using Error}
\usage{
ti_error(dummy_param = 0.5)
}
\arguments{
\item{dummy_param}{Parameter; Dummy parameter., Domain: U(0, 1), Default: 0.5,
Format: numeric.}
}
\value{
A TI method wrapper to be used together with
\code{\link[dynwrap:infer_trajectories]{infer_trajectory}}
}
\description{
Will generate a trajectory using Error.

This method was wrapped inside a
\href{https://github.com/dynverse/ti_error}{container}.
}
