% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_error.R, R/ti_container.R
\name{ti_error}
\alias{ti_error}
\alias{ti_error}
\title{Inferring trajectories with Control: error}
\usage{
ti_error(docker = TRUE, dummy_param = 0.5)

ti_error(docker = TRUE, dummy_param = 0.5)
}
\arguments{
\item{docker}{Whether to use the docker container or the R wrapper}

\item{dummy_param}{This parameter does not do anything.}
}
\value{
The trajectory model
}
\description{
This control method will always produce an error.

Will generate a trajectory using Control: error. This method was wrapped inside a \href{https://github.com/dynverse/dynmethods/tree/master/containers/error}{container}.
}
\details{
This methods was first wrapped inside R, see \link{ti_error}
}
