% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_forks.R
\name{ti_forks}
\alias{ti_forks}
\title{Inferring a trajectory inference using FORKS}
\usage{
ti_forks(norm_function = "median", norm_quantile = 75L,
  cum_sum_exp_var = 0.9, min_cluster = 4L, max_cluster = 10L,
  mapping_type = "Isomap", initialization = "kmeans",
  iterMax = 1000L, eta = 0.01, C = 1L)
}
\arguments{
\item{norm_function}{discrete; No description provided by the author. (default:
\code{"median"}; values: {\code{"mean"}, \code{"median"}, \code{"quantile"}})}

\item{norm_quantile}{numeric; No description provided by the author. (default:
\code{75L}; range: from \code{0L} to \code{100L})}

\item{cum_sum_exp_var}{numeric; No description provided by the author.
(default: \code{0.9}; range: from \code{0} to \code{1})}

\item{min_cluster}{integer; No description provided by the author. (default:
\code{4L}; range: from \code{3L} to \code{20L})}

\item{max_cluster}{integer; No description provided by the author. (default:
\code{10L}; range: from \code{3L} to \code{20L})}

\item{mapping_type}{discrete; No description provided by the author. (default:
\code{"Isomap"}; values: {\code{"Isomap"}, \code{"MDS"}, \code{"PCA"}, \code{"RandomForest"},
\code{"SpectralEmbedding"}, \code{"LLE_standard"}, \code{"tSNE"}})}

\item{initialization}{discrete; No description provided by the author.
(default: \code{"kmeans"}; values: {\code{"kmeans"}, \code{"kmedoids"}, \code{"random"}})}

\item{iterMax}{integer; No description provided by the author. (default:
\code{1000L}; range: from \code{100L} to \code{100000L})}

\item{eta}{numeric; No description provided by the author. (default: \code{0.01};
range: from \code{1e-04} to \code{100L})}

\item{C}{numeric; No description provided by the author. (default: \code{1L}; range:
from \code{1e-04} to \code{1000L})}
}
\value{
A TI method wrapper to be used together with
\code{\link[dynwrap:infer_trajectories]{infer_trajectory}}
}
\description{
Will generate a trajectory using \href{https://doi.org/10.1101/132811}{FORKS}.

This method was wrapped inside a
\href{https://github.com/dynverse/ti_forks}{container}.
The original code of this method is available
\href{https://github.com/macsharma/FORKS}{here}.
}
\references{
Sharma, M., Li, H., Sengupta, D., Prabhakar, S., Jayadeva, J.,
2017. FORKS: Finding Orderings Robustly using K-means and Steiner trees.
}
