% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_forks.R
\name{ti_forks}
\alias{ti_forks}
\title{FORKS}
\usage{
ti_forks(norm_function = "median", norm_quantile = 75L,
  cum_sum_exp_var = 0.9, cluster = c(4L, 10L),
  mapping_type = "Isomap", initialization = "kmeans",
  iterMax = 1000L, eta = 0.01, C = 1L)
}
\arguments{
\item{norm_function}{No description provided by the author. Domain: {mean,
median, quantile}. Default: median. Format: character.}

\item{norm_quantile}{No description provided by the author. Domain: U(0, 100).
Default: 75. Format: numeric.}

\item{cum_sum_exp_var}{No description provided by the author. Domain: U(0, 1).
Default: 0.9. Format: numeric.}

\item{cluster}{No description provided by the author. Domain: ( U(3, 20), U(3,
20) ). Default: (4, 10). Format: integer_range.}

\item{mapping_type}{No description provided by the author. Domain: {Isomap,
MDS, PCA, RandomForest, SpectralEmbedding, LLE_standard, tSNE}. Default:
Isomap. Format: character.}

\item{initialization}{No description provided by the author. Domain: {kmeans,
kmedoids, random}. Default: kmeans. Format: character.}

\item{iterMax}{No description provided by the author. Domain: e^U(4.61, 11.51).
Default: 1000. Format: integer.}

\item{eta}{No description provided by the author. Domain: e^U(-9.21, 4.61).
Default: 0.01. Format: numeric.}

\item{C}{No description provided by the author. Domain: e^U(-9.21, 6.91).
Default: 1. Format: numeric.}
}
\value{
A TI method wrapper to be used together with
\code{\link[dynwrap:infer_trajectories]{infer_trajectory}}
}
\description{
Will generate a trajectory using \href{https://doi.org/10.1101/132811}{FORKS}.

This method was wrapped inside a
\href{https://github.com/dynverse/ti_forks}{container}.
The original code of this method is available
\href{https://github.com/macsharma/FORKS}{here}.
}
\references{
Sharma, M., Li, H., Sengupta, D., Prabhakar, S., Jayadeva, J.,
2017. FORKS: Finding Orderings Robustly using K-means and Steiner trees.
}
\keyword{method}
