% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_gng.R
\name{ti_gng}
\alias{ti_gng}
\title{Inferring a trajectory inference using GNG}
\usage{
ti_gng(dimred = "landmark_mds", ndim = 5L, max_iter = 15000L,
  max_nodes = 8L, apply_mst = TRUE)
}
\arguments{
\item{dimred}{Parameter; Which dimensionality reduction method to use, Domain:
{pca, mds, tsne, ica, lle, landmark_mds, mds_sammon, mds_isomds, mds_smacof,
umap, dm_diffusionMap}, Default: landmark_mds, Format: character.}

\item{ndim}{Parameter; The number of dimensions, Domain: U(2, 10), Default: 5,
Format: integer.}

\item{max_iter}{Parameter; The max number of iterations, Domain: e^U(3.22,
13.82), Default: 15000, Format: numeric.}

\item{max_nodes}{Parameter; The maximum number of nodes, Domain: U(2, 30),
Default: 8, Format: integer.}

\item{apply_mst}{Parameter; If true, an MST post-processing of the GNG is
performed, Default: TRUE, Format: logical.}
}
\value{
A TI method wrapper to be used together with
\code{\link[dynwrap:infer_trajectories]{infer_trajectory}}
}
\description{
Will generate a trajectory using \href{https://github.com/rcannood/gng}{GNG}.

This method was wrapped inside a
\href{https://github.com/dynverse/ti_gng}{container}.
The original code of this method is available
\href{https://github.com/rcannood/gng}{here}.
}
