% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_gng.R, R/ti_container.R
\name{ti_gng}
\alias{ti_gng}
\alias{ti_gng}
\title{Inferring trajectories with Growing Neural Gas}
\usage{

  ti_gng(run_environment = ifelse(is.null(getOption("dynwrap_run_environment")),
  "local", getOption("dynwrap_run_environment")), dimred = "pca", ndim = 5,
  max_iter = 15000, max_nodes = 8, apply_mst = TRUE)


  ti_gng(run_environment = ifelse(is.null(getOption("dynwrap_run_environment")),
  "local", getOption("dynwrap_run_environment")), dimred = "pca", ndim = 5,
  max_iter = 15000, max_nodes = 8, apply_mst = TRUE)
}
\arguments{
\item{run_environment}{In which environment to run the method, can be 'local', 'docker' or 'singularity'}

\item{dimred}{A character vector specifying which dimensionality reduction method to use.
See \link[dyndimred:dimred]{dyndimred::dimred} for the list of available dimensionality reduction methods.}

\item{ndim}{The number of dimensions}

\item{max_iter}{The max number of iterations}

\item{max_nodes}{The maximum number of nodes}

\item{apply_mst}{If true, an MST post-processing of the GNG is performed.}
}
\value{
The trajectory model
}
\description{
Inferring trajectories with Growing Neural Gas

Will generate a trajectory using \href{https://doi.org/https://github.com/rcannood/gng}{Growing Neural Gas}. This method was wrapped inside a \href{https://github.com/dynverse/dynmethods/tree/master/containers/gng}{container}.
}
\details{
This methods was first wrapped inside R, see \link{ti_gng}

The original code of this method is available \href{https://github.com/rcannood/gng}{here}.
}
