% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_grandprix.R
\name{ti_grandprix}
\alias{ti_grandprix}
\title{Inferring a trajectory inference using GrandPrix}
\usage{
ti_grandprix(n_inducing_points = 40L, latent_prior_var = 0.1,
  latent_var = 0.028)
}
\arguments{
\item{n_inducing_points}{Parameter; , Domain: U(0, 1), Default: 40, Format:
integer.}

\item{latent_prior_var}{Parameter; , Domain: U(0, 1), Default: 0.1, Format:
numeric.}

\item{latent_var}{Parameter; , Domain: U(0, 1), Default: 0.028, Format:
numeric.}
}
\value{
A TI method wrapper to be used together with
\code{\link[dynwrap:infer_trajectories]{infer_trajectory}}
}
\description{
Will generate a trajectory using \href{https://doi.org/10.1101/227843}{GrandPrix}.

This method was wrapped inside a
\href{https://github.com/dynverse/ti_grandprix}{container}.
The original code of this method is available
\href{https://github.com/ManchesterBioinference/GrandPrix}{here}.
}
\references{
Ahmed, S., Rattray, M., Boukouvalas, A., 2017. GrandPrix: Scaling
up the Bayesian GPLVM for single-cell data.
}
