% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_identity.R
\name{ti_identity}
\alias{ti_identity}
\title{Inferring a trajectory inference using Identity}
\usage{
ti_identity(dummy_param = 0.5)
}
\arguments{
\item{dummy_param}{numeric; Dummy parameter (default: \code{0.5}; range: from \code{0} to
\code{1})}
}
\value{
A TI method wrapper to be used together with
\code{\link[dynwrap:infer_trajectories]{infer_trajectory}}
}
\description{
Will generate a trajectory using Identity.

This method was wrapped inside a
\href{https://github.com/dynverse/ti_identity}{container}.
}
