% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_identity.R, R/ti_container.R
\name{ti_identity}
\alias{ti_identity}
\alias{ti_identity}
\title{Inferring trajectories with Control: identity}
\usage{
ti_identity(docker = TRUE, dummy_param = 0.5)

ti_identity(docker = TRUE, dummy_param = 0.5)
}
\arguments{
\item{docker}{Whether to use the docker container or the R wrapper}

\item{dummy_param}{This parameter does not do anything.}
}
\value{
The trajectory model
}
\description{
This control method will return the gold standard.

Will generate a trajectory using Control: identity. This method was wrapped inside a \href{https://github.com/dynverse/dynmethods/tree/master/containers/identity}{container}.
}
\details{
This methods was first wrapped inside R, see \link{ti_identity}
}
