% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_matcher.R
\name{ti_matcher}
\alias{ti_matcher}
\title{Inferring a trajectory inference using MATCHER}
\usage{
ti_matcher(quantiles = 50L, method = "linear",
  config = dynwrap::container_config())
}
\arguments{
\item{quantiles}{integer; Quantiles How many quantiles to use when computing
warp functions (integer) (default: \code{50L}; range: from \code{2L} to \code{500L})}

\item{method}{discrete; Gaussian process regression or linear interpolation?
("gp" or "linear) (default: \code{"linear"}; values: {\code{"gp"}, \code{"linear"}})}

\item{config}{A container config. See \code{\link[=container_config]{container_config()}} for more information.}
}
\value{
A TI method wrapper to be used together with
\code{\link[dynwrap:infer_trajectories]{infer_trajectory}}
}
\description{
Will generate a trajectory using
\href{https://doi.org/10.1186/s13059-017-1269-0}{MATCHER}.

This method was wrapped inside a
\href{https://github.com/dynverse/ti_matcher}{container}.
The original code of this method is available
\href{https://github.com/jw156605/MATCHER}{here}.
}
\references{
Welch, J.D., Hartemink, A.J., Prins, J.F., 2017. MATCHER: manifold
alignment reveals correspondence between single cell transcriptome and
epigenome dynamics. Genome Biology 18.
}
