% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_merlot.R
\name{ti_merlot}
\alias{ti_merlot}
\title{Inferring trajectories with MERLoT}
\usage{
ti_merlot(n_components = 20L, n_components_to_use = 3L,
  NumberOfNodes = 100L)
}
\description{
Inferring trajectories with MERLoT
}
