% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_merlot.R
\name{ti_merlot}
\alias{ti_merlot}
\title{MERLoT}
\usage{
ti_merlot(sigma = "local", distance = "euclidean", ndim = 20L,
  density_norm = TRUE, n_local = c(5L, 7L), w_width = 0.01,
  n_components_to_use = 3L, N_yk = 100L, lambda_0 = 8e-10,
  mu_0 = 0.0025, increaseFactor_mu = 20L,
  increaseFactor_lambda = 20L, FixEndpoints = FALSE)
}
\arguments{
\item{sigma}{Diffusion scale parameter of the Gaussian kernel. A larger sigma
might be necessary if the eigenvalues can not be found because of a singularity
in the matrix. Must a character vector -- \code{"local"} (default) or \code{"global"}.
Domain: {local, global}. Default: local. Format: character.}

\item{distance}{A character vector specifying which distance metric to use.
Allowed measures are the Euclidean distance (default), the cosine distance
(\code{1-corr(c_1, c_2)}), or the rank correlation distance (\code{1-corr(rank(c_1), rank(c_2))}). Domain: {euclidean, cosine, rankcor}. Default: euclidean. Format:
character.}

\item{ndim}{Number of eigenvectors/dimensions to return. Domain: U(2, 20).
Default: 20. Format: integer.}

\item{density_norm}{Logical. If TRUE, use density normalisation. Default: TRUE.
Format: logical.}

\item{n_local}{If sigma == 'local', the \code{n_local} nearest neighbor(s)
determine(s) the local sigma. Domain: ( U(2, 20), U(2, 20) ). Default: (5, 7).
Format: integer_range.}

\item{w_width}{Window width to use for deciding the branch cutoff. Domain:
e^U(-9.21, 0.00). Default: 0.01. Format: numeric.}

\item{n_components_to_use}{Which components to use in downstream analysis.
Domain: U(2, 20). Default: 3. Format: integer.}

\item{N_yk}{Number of nodes for the elastic principal tree. Domain: U(2, 1000).
Default: 100. Format: integer.}

\item{lambda_0}{Principal elastic tree energy function parameter. Domain:
e^U(-27.63, -13.82). Default: 8e-10. Format: numeric.}

\item{mu_0}{Principal elastic tree energy function parameter. Domain: U(5e-04,
0.005). Default: 0.0025. Format: numeric.}

\item{increaseFactor_mu}{Factor by which the mu will be increased for the
embedding. Domain: U(2, 50). Default: 20. Format: numeric.}

\item{increaseFactor_lambda}{Factor by which the mu will be increased for the
embedding. Domain: U(2, 50). Default: 20. Format: numeric.}

\item{FixEndpoints}{Documentation not provided by authors. Default: FALSE.
Format: logical.}
}
\value{
A TI method wrapper to be used together with
\code{\link[dynwrap:infer_trajectories]{infer_trajectory}}
}
\description{
Will generate a trajectory using \href{https://doi.org/10.1101/261768}{MERLoT}.

This method was wrapped inside a
\href{https://github.com/dynverse/ti_merlot}{container}.
The original code of this method is available
\href{https://github.com/soedinglab/merlot}{here}.
}
\references{
Parra, R.G., Papadopoulos, N., Ahumada-Arranz, L., Kholtei, J.E.,
Mottelson, N., Horokhovsky, Y., Treutlein, B., Soeding, J., 2018.
Reconstructing complex lineage trees from scRNA-seq data using MERLoT.
}
\keyword{method}
