% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_mfa.R
\name{ti_mfa}
\alias{ti_mfa}
\title{Inferring a trajectory inference using MFA}
\usage{
ti_mfa(iter = 2000L, thin = 1L, pc_initialise = 1L,
  prop_collapse = 0, scale_input = TRUE, zero_inflation = FALSE)
}
\arguments{
\item{iter}{integer; Number of MCMC iterations (default: \code{2000L}; range: from
\code{20L} to \code{5000L})}

\item{thin}{integer; MCMC samples to thin (default: \code{1L}; range: from \code{1L} to
\code{20L})}

\item{pc_initialise}{integer; Which principal component to initialise
pseudotimes to (default: \code{1L}; range: from \code{1L} to \code{5L})}

\item{prop_collapse}{numeric; Proportion of Gibbs samples which should
marginalise over c (default: \code{0}; range: from \code{0} to \code{1})}

\item{scale_input}{logical; Logical. If true, input is scaled to have mean 0
variance 1}

\item{zero_inflation}{logical; Logical, should zero inflation be enabled?}
}
\value{
A TI method wrapper to be used together with
\code{\link[dynwrap:infer_trajectories]{infer_trajectory}}
}
\description{
Will generate a trajectory using
\href{https://doi.org/10.12688/wellcomeopenres.11087.1}{MFA}.

This method was wrapped inside a
\href{https://github.com/dynverse/ti_mfa}{container}.
The original code of this method is available
\href{https://github.com/kieranrcampbell/mfa}{here}.
}
\references{
Campbell, K.R., Yau, C., 2017. Probabilistic modeling of
bifurcations in single-cell gene expression data using a Bayesian mixture of
factor analyzers. Wellcome Open Research 2, 19.
}
