% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_mfa.R, R/ti_container.R
\name{ti_mfa}
\alias{ti_mfa}
\alias{ti_mfa}
\title{Inferring trajectories with mfa}
\usage{

  ti_mfa(run_environment = ifelse(is.null(getOption("dynwrap_run_environment")),
  "local", getOption("dynwrap_run_environment")), b = 2, iter = 2000,
  thin = 1, pc_initialise = 1, prop_collapse = 0, scale_input = TRUE,
  zero_inflation = FALSE)


  ti_mfa(run_environment = ifelse(is.null(getOption("dynwrap_run_environment")),
  "local", getOption("dynwrap_run_environment")), b = 2, iter = 2000,
  thin = 1, pc_initialise = 1, prop_collapse = 0, scale_input = TRUE,
  zero_inflation = FALSE)
}
\arguments{
\item{run_environment}{In which environment to run the method, can be 'local', 'docker' or 'singularity'}

\item{b}{Number of branches to model}

\item{iter}{Number of MCMC iterations}

\item{thin}{MCMC samples to thin}

\item{pc_initialise}{Which principal component to initialise pseudotimes to}

\item{prop_collapse}{Proportion of Gibbs samples which should marginalise over c}

\item{scale_input}{Logical. If true, input is scaled to have mean 0 variance 1}

\item{zero_inflation}{Logical, should zero inflation be enabled?}
}
\value{
The trajectory model
}
\description{
Inferring trajectories with mfa

Will generate a trajectory using \href{https://doi.org/10.12688/wellcomeopenres.11087.1}{mfa}. This method was wrapped inside a \href{https://github.com/dynverse/dynmethods/tree/master/containers/mfa}{container}.
}
\details{
This methods was first wrapped inside R, see \link{ti_mfa}

The original code of this method is available \href{https://github.com/kieranrcampbell/mfa}{here}.

The method is described in: \href{https://doi.org/10.12688/wellcomeopenres.11087.1}{Campbell, K.R., Yau, C., 2017. Probabilistic modeling of bifurcations in single-cell gene expression data using a Bayesian mixture of factor analyzers. Wellcome Open Research 2, 19.}
}
\seealso{
\code{\link[mfa:mfa]{mfa::mfa()}}
}
