% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_mfa.R
\name{ti_mfa}
\alias{ti_mfa}
\title{Inferring trajectories with mfa}
\usage{
ti_mfa(b = 2L, iter = 2000L, thin = 1L, pc_initialise = 1L,
  prop_collapse = 0, scale_input = TRUE, zero_inflation = FALSE)
}
\description{
Inferring trajectories with mfa
}
