% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_monocle.R
\name{ti_monocle_ica}
\alias{ti_monocle_ica}
\title{Inferring trajectories with Monocle v1}
\usage{
ti_monocle_ica(reduction_method = "ICA", max_components = 2,
  norm_method = "vstExprs")
}
\arguments{
\item{reduction_method}{A character string specifying the algorithm to use for dimensionality reduction.}

\item{max_components}{the dimensionality of the reduced space}

\item{norm_method}{Determines how to transform expression values prior to reducing dimensionality}
}
\description{
Inferring trajectories with Monocle v1
}
\seealso{
\code{\link[monocle:reduceDimension]{monocle::reduceDimension()}}, \code{\link[monocle:orderCells]{monocle::orderCells()}}
}
