% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_monocle_ica.R
\name{ti_monocle_ica}
\alias{ti_monocle_ica}
\title{Monocle ICA}
\usage{
ti_monocle_ica(
  reduction_method = "ICA",
  max_components = 2L,
  norm_method = "log",
  filter_features = TRUE,
  filter_features_mean_expression = 0.1
)
}
\arguments{
\item{reduction_method}{A character string specifying the algorithm to use for
dimensionality reduction. Domain: {ICA}. Default: ICA. Format: character.}

\item{max_components}{The dimensionality of the reduced space. Domain: U(2,
20). Default: 2. Format: integer.}

\item{norm_method}{Determines how to transform expression values prior to
reducing dimensionality. Domain: {vstExprs, log, none}. Default: log. Format:
character.}

\item{filter_features}{Whether to include monocle feature filtering. Default:
TRUE. Format: logical.}

\item{filter_features_mean_expression}{Minimal mean feature expression, only
used when \code{filter_features} is set to TRUE. Domain: U(0, 10). Default: 0.1.
Format: numeric.}
}
\value{
A TI method wrapper to be used together with
\code{\link[dynwrap:infer_trajectories]{infer_trajectory}}
}
\description{
Will generate a trajectory using \href{https://doi.org/10.1038/nmeth.4402}{Monocle ICA}.

This method was wrapped inside a
\href{https://github.com/dynverse/ti_monocle_ica}{container}.
The original code of this method is available
\href{https://github.com/cole-trapnell-lab/monocle-release}{here}.
}
\references{
Qiu, X., Mao, Q., Tang, Y., Wang, L., Chawla, R., Pliner, H.A.,
Trapnell, C., 2017. Reversed graph embedding resolves complex single-cell
trajectories. Nature Methods 14, 979–982.
}
\keyword{method}
