% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_monocle.R, R/ti_container.R
\name{ti_monocle_ica}
\alias{ti_monocle_ica}
\alias{ti_monocle_ica}
\title{Inferring trajectories with Monocle v1}
\usage{

  ti_monocle_ica(run_environment = ifelse(is.null(getOption("dynwrap_run_environment")),
  "local", getOption("dynwrap_run_environment")), reduction_method = "ICA",
  max_components = 2, norm_method = "vstExprs")


  ti_monocle_ica(run_environment = ifelse(is.null(getOption("dynwrap_run_environment")),
  "local", getOption("dynwrap_run_environment")), reduction_method = "ICA",
  max_components = 2, norm_method = "vstExprs")
}
\arguments{
\item{run_environment}{In which environment to run the method, can be 'local', 'docker' or 'singularity'}

\item{reduction_method}{A character string specifying the algorithm to use for dimensionality reduction.}

\item{max_components}{the dimensionality of the reduced space}

\item{norm_method}{Determines how to transform expression values prior to reducing dimensionality}
}
\value{
The trajectory model
}
\description{
Inferring trajectories with Monocle v1

Will generate a trajectory using \href{https://doi.org/10.1038/nmeth.4402}{Monocle ICA}. This method was wrapped inside a \href{https://github.com/dynverse/dynmethods/tree/master/containers/monocle_ica}{container}.
}
\details{
This methods was first wrapped inside R, see \link{ti_monocle_ica}

The original code of this method is available \href{https://github.com/cole-trapnell-lab/monocle-release}{here}.

The method is described in: \href{https://doi.org/10.1038/nmeth.4402}{Qiu, X., Mao, Q., Tang, Y., Wang, L., Chawla, R., Pliner, H.A., Trapnell, C., 2017. Reversed graph embedding resolves complex single-cell trajectories. Nature Methods 14, 979–982.}
}
\seealso{
\code{\link[monocle:reduceDimension]{monocle::reduceDimension()}}, \code{\link[monocle:orderCells]{monocle::orderCells()}}
}
