% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_mpath.R
\name{ti_mpath}
\alias{ti_mpath}
\title{Mpath}
\usage{
ti_mpath(
  distMethod = "euclidean",
  method = "kmeans",
  numcluster = 11L,
  numcluster_null = TRUE,
  diversity_cut = 0.6,
  size_cut = 0.05
)
}
\arguments{
\item{distMethod}{The method for calculating dissimilarity between cells.
distMethod can be one of "pearson", "kendall", "spearman" or "euclidean".
Default is "euclidean". Domain: {pearson, kendall, spearman, euclidean}.
Default: euclidean. Format: character.}

\item{method}{Method for distinguishing landmark clusters from non-landmark
clusters.method can be "kmeans" or "diversity" or "size" or "diversity_size".
When method="diversity", numlm needs to be specified. Default is
"diversity_size". Domain: {kmeans, diversity, size, diversity_size}. Default:
kmeans. Format: character.}

\item{numcluster}{Number of initial clusters. Domain: U(3, 30). Default: 11.
Format: integer.}

\item{numcluster_null}{If TRUE, will automatically select the number of
clusters. Default: TRUE. Format: logical.}

\item{diversity_cut}{The cutoff value of diversity for differentiating landmark
clusters from non-landmark clusters. The diversity of a landmark cluster must
be below this cutoff. Domain: U(0.1, 1). Default: 0.6. Format: numeric.}

\item{size_cut}{The cutoff value of size i.e. number of cells for
differentiating landmark clusters from non-landmark clusters. The number of
cells in a landmark cluster must be greater than this cutoff. Domain: U(0.01,
1). Default: 0.05. Format: numeric.}
}
\value{
A TI method wrapper to be used together with
\code{\link[dynwrap:infer_trajectories]{infer_trajectory}}
}
\description{
Will generate a trajectory using \href{https://doi.org/10.1038/ncomms11988}{Mpath}.

This method was wrapped inside a
\href{https://github.com/dynverse/ti_mpath}{container}.
The original code of this method is available
\href{https://github.com/JinmiaoChenLab/Mpath}{here}.
}
\references{
Chen, J., Schlitzer, A., Chakarov, S., Ginhoux, F., Poidinger, M.,
2016. Mpath maps multi-branching single-cell trajectories revealing progenitor
cell progression during development. Nature Communications 7.
}
\keyword{method}
