% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_mpath.R, R/ti_container.R
\name{ti_mpath}
\alias{ti_mpath}
\alias{ti_mpath}
\title{Inferring trajectories with Mpath}
\usage{
ti_mpath(docker = TRUE),
  distMethod = "euclidean", method = "kmeans", numcluster = 11,
  numcluster_null = TRUE, diversity_cut = 0.6, size_cut = 0.05)

ti_mpath(docker = TRUE),
  distMethod = "euclidean", method = "kmeans", numcluster = 11,
  numcluster_null = TRUE, diversity_cut = 0.6, size_cut = 0.05)
}
\arguments{
\item{docker}{Whether to use the docker container or the R wrapper}

\item{distMethod}{the method for calculating dissimilarity between cells. distMethod can be one of "pearson", "kendall", "spearman" or "euclidean". Default is "euclidean".}

\item{method}{method for distinguishing landmark clusters from non-landmark clusters.method can be "kmeans" or "diversity" or "size" or "diversity_size". When method="diversity", numlm needs to be specified. Default is "diversity_size".}

\item{numcluster}{number of initial clusters}

\item{numcluster_null}{If TRUE, will automatically select the number of clusters}

\item{diversity_cut}{the cutoff value of diversity for differentiating landmark clusters from non-landmark clusters. The diversity of a landmark cluster must be below this cutoff.}

\item{size_cut}{the cutoff value of size i.e. number of cells for differentiating landmark clusters from non-landmark clusters. The number of cells in a landmark cluster must be greater than this cutoff.}
}
\value{
The trajectory model
}
\description{
Inferring trajectories with Mpath

Will generate a trajectory using \href{https://doi.org/10.1038/ncomms11988}{Mpath}. This method was wrapped inside a \href{https://github.com/dynverse/dynmethods/tree/master/containers/mpath}{container}.
}
\details{
This methods was first wrapped inside R, see \link{ti_mpath}

The original code of this method is available \href{https://github.com/JinmiaoChenLab/Mpath}{here}.

The method is described in: \href{https://doi.org/10.1038/ncomms11988}{Chen, J., Schlitzer, A., Chakarov, S., Ginhoux, F., Poidinger, M., 2016. Mpath maps multi-branching single-cell trajectories revealing progenitor cell progression during development. Nature Communications 7, 11988.}
}
