% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_mst.R
\name{ti_mst}
\alias{ti_mst}
\title{MST}
\usage{
ti_mst(dimred = "pca", ndim = 2L)
}
\arguments{
\item{dimred}{Which dimensionality reduction method to use. Domain: {pca, mds,
tsne, ica, lle, landmark_mds, mds_sammon, mds_isomds, mds_smacof, umap,
dm_diffusionMap}. Default: pca. Format: character.}

\item{ndim}{. Domain: U(2, 30). Default: 2. Format: integer.}
}
\value{
A TI method wrapper to be used together with
\code{\link[dynwrap:infer_trajectories]{infer_trajectory}}
}
\description{
Will generate a trajectory using MST.

This method was wrapped inside a
\href{https://github.com/dynverse/ti_mst}{container}.
}
\keyword{method}
