% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_ouija.R
\name{ti_ouija}
\alias{ti_ouija}
\title{Inferring a trajectory inference using ouija}
\usage{
ti_ouija(iter = 100L, response_type = "switch",
  inference_type = "hmc", normalise_expression = TRUE)
}
\arguments{
\item{iter}{Parameter; Number of iterations, Domain: e^U(2.30, 6.91), Default:
100, Format: numeric.}

\item{response_type}{Parameter; A vector declaring whether each gene exhibits
"switch" or "transient"expression. Defaults to "switch" for all genes, Domain:
{switch, transient}, Default: switch, Format: character.}

\item{inference_type}{Parameter; The type of inference to be performed, either
\code{hmc} for HamiltonianMonte Carlo or \code{vb} for ADVI (Variational
Bayes). Note that HMC is typically more accuratebut VB will be orders of
magnitude faster, Domain: {hmc, vb}, Default: hmc, Format: character.}

\item{normalise_expression}{Parameter; Logical, default TRUE. If TRUE the data
is pre-normalisedso the average peak expression is approximately 1. This makes
the strength parametersapproximately comparable between genes, Default: TRUE,
Format: logical.}
}
\value{
A TI method wrapper to be used together with
\code{\link[dynwrap:infer_trajectories]{infer_trajectory}}
}
\description{
Will generate a trajectory using \href{https://doi.org/10.1101/060442}{ouija}.

This method was wrapped inside a
\href{https://github.com/dynverse/ti_ouija}{container}.
The original code of this method is available
\href{https://github.com/kieranrcampbell/ouija}{here}.
}
\references{
Campbell, K.R., Yau, C., 2016. A descriptive marker gene approach
to single-cell pseudotime inference.
}
