% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_ouija.R, R/ti_container.R
\name{ti_ouija}
\alias{ti_ouija}
\alias{ti_ouija}
\title{Inferring trajectories with ouija}
\usage{
ti_ouija(docker = "stevedore" \%in\% rownames(installed.packages()),
  iter = 100, response_type = "switch", inference_type = "hmc",
  normalise_expression = TRUE)

ti_ouija(docker = "stevedore" \%in\% rownames(installed.packages()),
  iter = 100, response_type = "switch", inference_type = "hmc",
  normalise_expression = TRUE)
}
\arguments{
\item{docker}{Whether to use the docker container or the R wrapper}

\item{iter}{Number of iterations}

\item{response_type}{A vector declaring whether each gene exhibits "switch" or "transient"
expression. Defaults to "switch" for all genes}

\item{inference_type}{The type of inference to be performed, either \code{hmc} for Hamiltonian
Monte Carlo or \code{vb} for ADVI (Variational Bayes). Note that HMC is typically more accurate
but VB will be orders of magnitude faster.}

\item{normalise_expression}{Logical, default TRUE. If TRUE the data is pre-normalised
so the average peak expression is approximately 1. This makes the strength parameters
approximately comparable between genes.}
}
\value{
The trajectory model
}
\description{
Inferring trajectories with ouija

Will generate a trajectory using \href{https://doi.org/10.1101/060442}{ouija}. This method was wrapped inside a \href{https://github.com/dynverse/dynmethods/tree/master/containers/ouija}{container}.
}
\details{
This methods was first wrapped inside R, see \link{ti_ouija}

The original code of this method is available \href{https://github.com/kieranrcampbell/ouija}{here}.

The method is described in: \href{https://doi.org/10.1101/060442}{Campbell, K.R., Yau, C., 2016. A descriptive marker gene approach to single-cell pseudotime inference.}
}
\seealso{
\code{\link[ouija:ouija]{ouija::ouija()}}
}
