% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_ouijaflow.R
\name{ti_ouijaflow}
\alias{ti_ouijaflow}
\title{Inferring a trajectory inference using ouijaflow}
\usage{
ti_ouijaflow(iter = 1000L, run_environment = NULL)
}
\arguments{
\item{iter}{integer; (default: \code{1000L}; range: from \code{2L} to \code{50000L})}

\item{run_environment}{In which environment to run the method, can be \code{"docker"} or \code{"singularity"}.}
}
\value{
A TI method wrapper to be used together with
\code{\link[dynwrap:infer_trajectories]{infer_trajectory}}
}
\description{
Will generate a trajectory using \href{https://doi.org/10.1101/060442}{ouijaflow}.

This method was wrapped inside a
\href{https://github.com/dynverse/dynmethods/tree/master/containers/ouijaflow}{container}.
The original code of this method is available
\href{https://github.com/kieranrcampbell/ouija}{here}.
}
\references{
Campbell, K.R., Yau, C., 2016. A descriptive marker gene approach
to single-cell pseudotime inference.
}
