% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_ouijaflow.R
\name{ti_ouijaflow}
\alias{ti_ouijaflow}
\title{Inferring a trajectory inference using ouijaflow}
\usage{
ti_ouijaflow(iter = 1000L)
}
\arguments{
\item{iter}{Parameter; ., Domain: e^U(0.69, 10.82), Default: 1000, Format:
integer.}
}
\value{
A TI method wrapper to be used together with
\code{\link[dynwrap:infer_trajectories]{infer_trajectory}}
}
\description{
Will generate a trajectory using \href{https://doi.org/10.1101/060442}{ouijaflow}.

This method was wrapped inside a
\href{https://github.com/dynverse/ti_ouijaflow}{container}.
The original code of this method is available
\href{https://github.com/kieranrcampbell/ouija}{here}.
}
\references{
Campbell, K.R., Yau, C., 2016. A descriptive marker gene approach
to single-cell pseudotime inference.
}
