% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_container.R
\name{ti_paga}
\alias{ti_paga}
\title{Inferring a trajectory inference using \href{https://doi.org/10.1101/208819}{PAGA}}
\usage{
ti_paga(n_neighbors = 30L, n_comps = 50L, n_dcs = 15L, resolution = 2.5,
  embedding_type = "fa")
}
\arguments{
\item{n_neighbors}{Number of neighbours for knn \cr
integer; default: 30L; possible values between 1 and 100}

\item{n_comps}{Number of principal components \cr
integer; default: 50L; possible values between 0 and 100}

\item{n_dcs}{Number of diffusion components for denoising graph, 0 means no denoising. \cr
integer; default: 15L; possible values between 0 and 40}

\item{resolution}{Resolution of louvain clustering, which determines the granularity of the clustering. Higher values will result in more clusters. \cr
numeric; default: 2.5; possible values between 0.1 and 10}

\item{embedding_type}{Either 'umap' (scales very well, recommended for very large datasets) or 'fa' (ForceAtlas2, often a bit more intuitive for small datasets). \cr}
}
\value{
The trajectory model
}
\description{
Will generate a trajectory using \href{https://doi.org/10.1101/208819}{PAGA}. This method was wrapped inside a \href{https://github.com/dynverse/dynmethods/tree/master/containers/paga}{container}.
}
\details{
The original code of this method is available \href{https://github.com/theislab/graph_abstraction}{here}.

The method is described in: \href{https://doi.org/10.1101/208819}{Wolf, F.A., Hamey, F., Plass, M., Solana, J., Dahlin, J.S., Gottgens, B., Rajewsky, N., Simon, L., Theis, F.J., 2017. Graph abstraction reconciles clustering with trajectory inference through a topology preserving map of single cells.}
}
