% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_paga.R
\name{ti_paga}
\alias{ti_paga}
\title{Inferring a trajectory inference using PAGA}
\usage{
ti_paga(n_neighbors = 30L, n_comps = 50L, n_dcs = 15L,
  resolution = 2.5, embedding_type = "fa", run_environment = NULL)
}
\arguments{
\item{n_neighbors}{integer; Number of neighbours for knn (default: \code{30L};
range: from \code{1L} to \code{100L})}

\item{n_comps}{integer; Number of principal components (default: \code{50L}; range:
from \code{0L} to \code{100L})}

\item{n_dcs}{integer; Number of diffusion components for denoising graph, 0
means no denoising. (default: \code{15L}; range: from \code{0L} to \code{40L})}

\item{resolution}{numeric; Resolution of louvain clustering, which determines
the granularity of the clustering. Higher values will result in more clusters.
(default: \code{2.5}; range: from \code{0.1} to \code{10L})}

\item{embedding_type}{discrete; Either 'umap' (scales very well, recommended
for very large datasets) or 'fa' (ForceAtlas2, often a bit more intuitive for
small datasets). (default: \code{"fa"}; values: {``})}

\item{run_environment}{In which environment to run the method, can be \code{"docker"} or \code{"singularity"}.}
}
\value{
A TI method wrapper to be used together with
\code{\link[dynwrap:infer_trajectories]{infer_trajectory}}
}
\description{
Will generate a trajectory using \href{https://doi.org/10.1101/208819}{PAGA}.

This method was wrapped inside a
\href{https://github.com/dynverse/dynmethods/tree/master/containers/paga}{container}.
The original code of this method is available
\href{https://github.com/theislab/graph_abstraction}{here}.
}
\references{
Wolf, F.A., Hamey, F., Plass, M., Solana, J., Dahlin, J.S.,
Gottgens, B., Rajewsky, N., Simon, L., Theis, F.J., 2017. Graph abstraction
reconciles clustering with trajectory inference through a topology preserving
map of single cells.
}
