% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_pcreode.R
\name{ti_pcreode}
\alias{ti_pcreode}
\title{Inferring a trajectory inference using \href{https://doi.org/10.1016/j.cels.2017.10.012}{pCreode}}
\usage{
ti_pcreode(n_pca_components = 3L, radius = 1L, noise = 8L, target = 25L,
  num_runs = 10L)
}
\arguments{
\item{n_pca_components}{\cr
integer; default: 3L; possible values between 2 and 10}

\item{radius}{\cr
numeric; default: 1L; possible values between 0.01 and 10}

\item{noise}{\cr
numeric; default: 8L; possible values between 1 and 20}

\item{target}{\cr
numeric; default: 25L; possible values between 5 and 100}

\item{num_runs}{\cr
integer; default: 10L; possible values between 10 and 1000}
}
\value{
The trajectory model
}
\description{
Will generate a trajectory using \href{https://doi.org/10.1016/j.cels.2017.10.012}{pCreode}. This method was wrapped inside a \href{https://github.com/dynverse/dynmethods/tree/master/containers/pcreode}{container}.
}
\details{
The original code of this method is available \href{https://github.com/KenLauLab/pCreode}{here}.

The method is described in: \href{https://doi.org/10.1016/j.cels.2017.10.012}{Herring, C.A., Banerjee, A., McKinley, E.T., Simmons, A.J., Ping, J., Roland, J.T., Franklin, J.L., Liu, Q., Gerdes, M.J., Coffey, R.J., Lau, K.S., 2018. Unsupervised Trajectory Analysis of Single-Cell RNA-Seq and Imaging Data Reveals Alternative Tuft Cell Origins in the Gut. Cell Systems 6, 37–51.e9.}
}
