% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_pcreode.R
\name{ti_pcreode}
\alias{ti_pcreode}
\title{pCreode}
\usage{
ti_pcreode(n_pca_components = 3L, num_runs = 10L)
}
\arguments{
\item{n_pca_components}{. Domain: U(2, 10). Default: 3. Format: integer.}

\item{num_runs}{. Domain: e^U(2.30, 4.61). Default: 10. Format: integer.}
}
\value{
A TI method wrapper to be used together with
\code{\link[dynwrap:infer_trajectories]{infer_trajectory}}
}
\description{
Will generate a trajectory using
\href{https://doi.org/10.1016/j.cels.2017.10.012}{pCreode}.

This method was wrapped inside a
\href{https://github.com/dynverse/ti_pcreode}{container}.
The original code of this method is available
\href{https://github.com/KenLauLab/pCreode}{here}.
}
\references{
Herring, C.A., Banerjee, A., McKinley, E.T., Simmons, A.J., Ping,
J., Roland, J.T., Franklin, J.L., Liu, Q., Gerdes, M.J., Coffey, R.J., Lau,
K.S., 2018. Unsupervised Trajectory Analysis of Single-Cell RNA-Seq and Imaging
Data Reveals Alternative Tuft Cell Origins in the Gut. Cell Systems 6,
37–51.e9.
}
\keyword{method}
