% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_periodpc.R, R/ti_container.R
\name{ti_periodpc}
\alias{ti_periodpc}
\alias{ti_periodpc}
\title{Inferring trajectories with Periodic PrinCurvee}
\usage{
ti_periodpc(docker = dynwrap::test_docker_installation(), ndim = 3,
  maxit = 10)

ti_periodpc(docker = dynwrap::test_docker_installation(), ndim = 3,
  maxit = 10)
}
\arguments{
\item{docker}{Whether to use the docker container or the R wrapper}

\item{ndim}{The number of dimensions}

\item{maxit}{maximum number of iterations.}
}
\value{
The trajectory model
}
\description{
Inferring trajectories with Periodic PrinCurvee

Will generate a trajectory using Periodic PrinCurve. This method was wrapped inside a \href{https://github.com/dynverse/dynmethods/tree/master/containers/periodpc}{container}.
}
\details{
This methods was first wrapped inside R, see \link{ti_periodpc}
}
\seealso{
\code{\link[princurve:principal.curve]{princurve::principal.curve()}}
}
