% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_periodpc.R
\name{ti_periodpc}
\alias{ti_periodpc}
\title{Inferring trajectories with Periodic PrinCurvee}
\usage{
ti_periodpc(ndim = 3L, maxit = 10L)
}
\description{
Arguments passed to this function will be used as default parameters for the method.
}
