% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_periodpc.R
\name{ti_periodpc}
\alias{ti_periodpc}
\title{Inferring a trajectory inference using Periodic PrinCurve}
\usage{
ti_periodpc(ndim = 3, maxit = 10)
}
\arguments{
\item{ndim}{integer; (default: \code{3}; range: from \code{2} to \code{10})}

\item{maxit}{integer; (default: \code{10}; range: from \code{0} to \code{100})}
}
\value{
A TI method wrapper to be used together with
\code{\link[dynwrap:infer_trajectories]{infer_trajectory}}
}
\description{
Will generate a trajectory using Periodic PrinCurve.

This method was wrapped inside a
\href{https://github.com/dynverse/ti_periodpc}{container}.
}
