% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_phenopath.R
\name{ti_phenopath}
\alias{ti_phenopath}
\title{Inferring trajectories with PhenoPath}
\usage{
ti_phenopath(thin = 40L, z_init = 1, model_mu = FALSE, scale_y = TRUE)
}
\description{
Inferring trajectories with PhenoPath
}
