% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_phenopath.R
\name{ti_phenopath}
\alias{ti_phenopath}
\title{Inferring a trajectory inference using PhenoPath}
\usage{
ti_phenopath(thin = 40L, z_init = "1", model_mu = FALSE,
  scale_y = TRUE, container_type = NULL)
}
\arguments{
\item{thin}{integer; The number of iterations to wait each time
beforere-calculating the elbo (default: \code{40L}; range: from \code{2L} to \code{500L})}

\item{z_init}{discrete; The initialisation of the latent trajectory. Should be
one of\enumerate{\item A positive integer describing which principal component
of the data shouldbe used for initialisation (default 1), \emph{or}\item A
numeric vector of length number of samples to be used directly for
initialisation, \emph{or}\item The text character \code{"random"}, for random
initialisation from a standard normal distribution.} (default: \code{"1"}; values:
{\code{"1"}, \code{"2"}, \code{"3"}, \code{"4"}, \code{"5"}, \code{"random"}})}

\item{model_mu}{logical; Logical - should a gene-specific intercept term be
modelled?}

\item{scale_y}{logical; Logical - should the expression matrix be centre
scaled?}

\item{container_type}{In which environment to run the method, can be \code{"docker"} or \code{"singularity"}.
By default, docker will be used.}
}
\value{
A TI method wrapper to be used together with
\code{\link[dynwrap:infer_trajectories]{infer_trajectory}}
}
\description{
Will generate a trajectory using \href{https://doi.org/10.1101/159913}{PhenoPath}.

This method was wrapped inside a
\href{https://github.com/dynverse/dynmethods/tree/master/containers/phenopath}{container}.
The original code of this method is available
\href{https://github.com/kieranrcampbell/phenopath}{here}.
}
\references{
Campbell, K., Yau, C., 2017. Uncovering genomic trajectories with
heterogeneous genetic and environmental backgrounds across single-cells and
populations.
}
