% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_phenopath.R
\name{ti_phenopath}
\alias{ti_phenopath}
\title{PhenoPath}
\usage{
ti_phenopath(thin = 40L, z_init = "1", model_mu = FALSE, scale_y = TRUE)
}
\arguments{
\item{thin}{The number of iterations to wait each time before re-calculating
the elbo. Domain: U(2, 500). Default: 40. Format: integer.}

\item{z_init}{The initialisation of the latent trajectory. Should be one of* A
positive integer describing which principal component of the data should be
used for initialisation (default 1), or * The text character \code{"random"}, for
random initialisation from a standard normal distribution. Domain: {1, 2, 3, 4,
5, random}. Default: 1. Format: character.}

\item{model_mu}{Logical - should a gene-specific intercept term be modelled?.
Default: FALSE. Format: logical.}

\item{scale_y}{Logical - should the expression matrix be centre scaled?.
Default: TRUE. Format: logical.}
}
\value{
A TI method wrapper to be used together with
\code{\link[dynwrap:infer_trajectories]{infer_trajectory}}
}
\description{
Will generate a trajectory using \href{https://doi.org/10.1101/159913}{PhenoPath}.

This method was wrapped inside a
\href{https://github.com/dynverse/ti_phenopath}{container}.
The original code of this method is available
\href{https://github.com/kieranrcampbell/phenopath}{here}.
}
\references{
Campbell, K., Yau, C., 2017. Uncovering genomic trajectories with
heterogeneous genetic and environmental backgrounds across single-cells and
populations.
}
\keyword{method}
