% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_phenopath.R, R/ti_container.R
\name{ti_phenopath}
\alias{ti_phenopath}
\alias{ti_phenopath}
\title{Inferring trajectories with PhenoPath}
\usage{
ti_phenopath(docker = "stevedore" \%in\% rownames(installed.packages()),
  thin = 40, z_init = "1", model_mu = FALSE, scale_y = TRUE)

ti_phenopath(docker = "stevedore" \%in\% rownames(installed.packages()),
  thin = 40, z_init = "1", model_mu = FALSE, scale_y = TRUE)
}
\arguments{
\item{docker}{Whether to use the docker container or the R wrapper}

\item{thin}{The number of iterations to wait each time before
re-calculating the elbo}

\item{z_init}{The initialisation of the latent trajectory. Should be one of
\enumerate{
\item A positive integer describing which principal component of the data should
be used for initialisation (default 1), \emph{or}
\item A numeric vector of length number of samples to be used 
directly for initialisation, \emph{or}
\item The text character \code{"random"}, for random initialisation 
from a standard normal distribution.
}}

\item{model_mu}{Logical - should a gene-specific intercept term be modelled?}

\item{scale_y}{Logical - should the expression matrix be centre scaled?}
}
\value{
The trajectory model
}
\description{
Inferring trajectories with PhenoPath

Will generate a trajectory using \href{https://doi.org/10.1101/159913}{PhenoPath}. This method was wrapped inside a \href{https://github.com/dynverse/dynmethods/tree/master/containers/phenopath}{container}.
}
\details{
This methods was first wrapped inside R, see \link{ti_phenopath}

The original code of this method is available \href{https://github.com/kieranrcampbell/phenopath}{here}.

The method is described in: \href{https://doi.org/10.1101/159913}{Campbell, K., Yau, C., 2017. Uncovering genomic trajectories with heterogeneous genetic and environmental backgrounds across single-cells and populations.}
}
\seealso{
\code{\link[phenopath:phenopath]{phenopath::phenopath()}}, \code{\link[phenopath:clvm]{phenopath::clvm()}}
}
