% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_projected_monocle.R
\name{ti_projected_monocle}
\alias{ti_projected_monocle}
\title{Projected Monocle}
\usage{
ti_projected_monocle(
  reduction_method = "DDRTree",
  max_components = 2L,
  norm_method = "vstExprs",
  auto_param_selection = TRUE,
  filter_features = TRUE,
  filter_features_mean_expression = 0.1
)
}
\arguments{
\item{reduction_method}{A character string specifying the algorithm to use for
dimensionality reduction. Domain: {DDRTree}. Default: DDRTree. Format:
character.}

\item{max_components}{The dimensionality of the reduced space. Domain: U(2,
20). Default: 2. Format: integer.}

\item{norm_method}{Determines how to transform expression values prior to
reducing dimensionality. Domain: {vstExprs, log, none}. Default: vstExprs.
Format: character.}

\item{auto_param_selection}{When this argument is set to TRUE (default), it
will automatically calculate the proper value for the ncenter (number of
centroids) parameters which will be passed into DDRTree call. Default: TRUE.
Format: logical.}

\item{filter_features}{Whether to include monocle feature filtering. Default:
TRUE. Format: logical.}

\item{filter_features_mean_expression}{Minimal mean feature expression, only
used when \code{filter_features} is set to TRUE. Domain: U(0, 10). Default: 0.1.
Format: numeric.}
}
\value{
A TI method wrapper to be used together with
\code{\link[dynwrap:infer_trajectories]{infer_trajectory}}
}
\description{
Will generate a trajectory using \href{https://github.com/cole-trapnell-lab/monocle-release}{Projected Monocle}.

This method was wrapped inside a
\href{https://github.com/dynverse/ti_projected_monocle}{container}.
The original code of this method is available
\href{https://github.com/cole-trapnell-lab/monocle-release}{here}.
}
\keyword{method}
