% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_projected_monocle.R
\name{ti_projected_monocle}
\alias{ti_projected_monocle}
\title{Inferring a trajectory inference using Projected Monocle}
\usage{
ti_projected_monocle(reduction_method = "DDRTree", max_components = 2L,
  norm_method = "vstExprs", auto_param_selection = TRUE,
  run_environment = NULL)
}
\arguments{
\item{reduction_method}{discrete; A character string specifying the algorithm
to use for dimensionality reduction. (default: \code{"DDRTree"}; values:
{\code{"DDRTree"}})}

\item{max_components}{integer; The dimensionality of the reduced space
(default: \code{2L}; range: from \code{2L} to \code{20L})}

\item{norm_method}{discrete; Determines how to transform expression values
prior to reducing dimensionality (default: \code{"vstExprs"}; values: {\code{"vstExprs"},
\code{"log"}, \code{"none"}})}

\item{auto_param_selection}{logical; When this argument is set to TRUE
(default), it will automatically calculate the proper value for the ncenter
(number of centroids) parameters which will be passed into DDRTree call.}

\item{run_environment}{In which environment to run the method, can be \code{"docker"} or \code{"singularity"}.}
}
\value{
A TI method wrapper to be used together with
\code{\link[dynwrap:infer_trajectories]{infer_trajectory}}
}
\description{
Will generate a trajectory using \href{https://github.com/cole-trapnell-lab/monocle-release}{Projected Monocle}.

This method was wrapped inside a
\href{https://github.com/dynverse/dynmethods/tree/master/containers/projected_monocle}{container}.
The original code of this method is available
\href{https://github.com/cole-trapnell-lab/monocle-release}{here}.
}
