% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_projected_tscan.R
\name{ti_projected_tscan}
\alias{ti_projected_tscan}
\title{Inferring a trajectory inference using Projected TSCAN}
\usage{
ti_projected_tscan(minexpr_percent = 0, minexpr_value = 0,
  cvcutoff = 0, clusternum_lower = 2L, clusternum_upper = 9L,
  modelNames = "VVV")
}
\arguments{
\item{minexpr_percent}{Parameter; ., Domain: U(0, 1), Default: 0, Format:
numeric.}

\item{minexpr_value}{Parameter; ., Domain: U(0, 10), Default: 0, Format:
numeric.}

\item{cvcutoff}{Parameter; ., Domain: U(0, 5), Default: 0, Format: numeric.}

\item{clusternum_lower}{Parameter; ., Domain: U(2, 20), Default: 2, Format:
integer.}

\item{clusternum_upper}{Parameter; ., Domain: U(2, 20), Default: 9, Format:
integer.}

\item{modelNames}{Parameter; ., Domain: {EII, VII, EEI, VEI, EVI, VVI, EEE,
EVE, VEE, VVE, EEV, VEV, EVV, VVV}, Default: VVV, Format: character.}
}
\value{
A TI method wrapper to be used together with
\code{\link[dynwrap:infer_trajectories]{infer_trajectory}}
}
\description{
Will generate a trajectory using \href{https://github.com/zji90/TSCAN}{Projected TSCAN}.

This method was wrapped inside a
\href{https://github.com/dynverse/ti_tscan}{container}.
The original code of this method is available
\href{https://github.com/zji90/TSCAN}{here}.
}
