% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_pseudogp.R
\name{ti_pseudogp}
\alias{ti_pseudogp}
\title{Inferring a trajectory inference using pseudogp}
\usage{
ti_pseudogp(smoothing_alpha = 10, smoothing_beta = 3,
  pseudotime_mean = 0.5, pseudotime_var = 1, chains = 3L,
  iter = 100L, dimreds = c(TRUE, TRUE, FALSE, FALSE, FALSE, FALSE,
  FALSE, FALSE, FALSE, FALSE, FALSE), initialise_from = "random",
  config = dynwrap::container_config())
}
\arguments{
\item{smoothing_alpha}{numeric; The hyperparameter for the Gamma distribution
that controls arc-length (default: \code{10}; range: from \code{1} to \code{20})}

\item{smoothing_beta}{numeric; The hyperparameter for the Gamma distribution
that controls arc-length (default: \code{3}; range: from \code{1} to \code{20})}

\item{pseudotime_mean}{numeric; The mean of the constrained normal prior on the
pseudotimes (default: \code{0.5}; range: from \code{0} to \code{1})}

\item{pseudotime_var}{numeric; The variance of the constrained normal prior on
the pseudotimes (default: \code{1}; range: from \code{0.01} to \code{1})}

\item{chains}{integer; The number of chains for the MCMC trace (default: \code{3L};
range: from \code{1L} to \code{20L})}

\item{iter}{integer; The number of iterations for the MCMC trace (default:
\code{100L}; range: from \code{100L} to \code{1000L})}

\item{dimreds}{logical_vector; A character vector specifying which
dimensionality reduction methods to use.See
\code{\link[dyndimred:dimred]{dyndimred::dimred}} for the list of available
dimensionality reduction methods.}

\item{initialise_from}{discrete; How to initialise the MCMC chain. One of
"random" (stan decides),"principal_curve", or "pca" (the first component of PCA
rescaled is taken to be the pseudotimes).Note: if multiple representations are
provided, \code{pseudogp} will take the principal curve orpca from the first
rather than combining them. If a particular representation is required, it isup
to the user to re-order them. (default: \code{"random"}; values: {\code{"random"},
\code{"principal_curve"}, \code{"pca"}})}

\item{config}{A container config. See \code{\link[=container_config]{container_config()}} for more information.}
}
\value{
A TI method wrapper to be used together with
\code{\link[dynwrap:infer_trajectories]{infer_trajectory}}
}
\description{
Will generate a trajectory using
\href{https://doi.org/10.1371/journal.pcbi.1005212}{pseudogp}.

This method was wrapped inside a
\href{https://github.com/dynverse/ti_pseudogp}{container}.
The original code of this method is available
\href{https://github.com/kieranrcampbell/pseudogp}{here}.
}
\references{
Campbell, K.R., Yau, C., 2016. Order Under Uncertainty: Robust
Differential Expression Analysis Using Probabilistic Models for Pseudotime
Inference. PLOS Computational Biology 12, e1005212.
}
