% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_pseudogp.R
\name{ti_pseudogp}
\alias{ti_pseudogp}
\title{Inferring trajectories with pseudogp}
\usage{
ti_pseudogp(smoothing_alpha = 10, smoothing_beta = 3,
  pseudotime_mean = 0.5, pseudotime_var = 1, chains = 3L, iter = 100,
  dimreds = c(TRUE, TRUE, FALSE, FALSE, FALSE), initialise_from = "random")
}
\description{
Arguments passed to this function will be used as default parameters for the method.
}
