% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_raceid_stemid.R
\name{ti_raceid_stemid}
\alias{ti_raceid_stemid}
\title{Inferring a trajectory inference using RaceID / StemID}
\usage{
ti_raceid_stemid(knn = 10L, ccor = 0.4, metric = "pearson",
  sat = TRUE, samp = 1000L, cln = 30L, clustnr = 30L,
  bootnr = 50L, FUNcluster = "kmedoids", probthr = 0.001,
  outminc = 5L, outlg = 2L, outdistquant = 0.95,
  initial_cmd = TRUE, perplexity = 30L, cthr = 5L, nmode = TRUE,
  projcells_knn = 3L, fr = FALSE, pdishuf = 500L, fast = FALSE,
  pthr = 0.01, scthr = 0.2, container_type = NULL)
}
\arguments{
\item{knn}{integer; Number of nearest neighbors used to infer corresponding
cell types in different batches. (default: \code{10L}; range: from \code{5L} to \code{50L})}

\item{ccor}{numeric; Correlation coefficient used as a treshhold for
determining genes correlated to eachother. (default: \code{0.4}; range: from \code{0} to
\code{1})}

\item{metric}{discrete; Distances are computed from the filtered expression
matrix after optional feature selection, dimensional reduction, and/or
transformation (batch correction). (default: \code{"pearson"}; values: {\code{"pearson"},
\code{"spearman"}, \code{"logpearson"}, \code{"euclidean"}})}

\item{sat}{logical; If \code{TRUE}, then the number of clusters is determined
based on finding the saturation point of the mean within-cluster dispersion as
a function of the cluster number. If \code{FALSE}, then cluster number needs to
be given as \code{cln}.}

\item{samp}{integer; Number of booststrapping runs for clusterboot. Default is
50 (default: \code{1000L}; range: from \code{50L} to \code{10000L})}

\item{cln}{integer; Number of clusters to be used. If \code{sat} is
\code{TRUE}, this number is inferred by the saturation criterion. (default:
\code{30L}; range: from \code{10L} to \code{100L})}

\item{clustnr}{integer; Maximum number of clusters for the derivation of the
cluster number by the saturation of mean within-cluster-dispersion. (default:
\code{30L}; range: from \code{10L} to \code{100L})}

\item{bootnr}{integer; Number of booststrapping runs for clusterboot. (default:
\code{50L}; range: from \code{20L} to \code{100L})}

\item{FUNcluster}{discrete; Clustering method used by RaceID3. (default:
\code{"kmedoids"}; values: {\code{"kmedoids"}, \code{"kmeans"}, \code{"hclust"}})}

\item{probthr}{numeric; Outlier probability threshold for a minimum of outlg
genes to be an outlier cell. This probability is computed from a negative
binomial background model of expression in a cluster. (default: \code{0.001}; range:
from \code{1e-05} to \code{1L})}

\item{outminc}{integer; Minimal transcript count of a gene in a clusters to be
tested for being an outlier gene. (default: \code{5L}; range: from \code{0L} to \code{100L})}

\item{outlg}{integer; Minimum number of outlier genes required for being an
outlier cell. (default: \code{2L}; range: from \code{0L} to \code{100L})}

\item{outdistquant}{numeric; Real number between zero and one. Outlier cells
are merged to outlier clusters if their distance smaller than the
outdistquant-quantile of the distance distribution of pairs of cells in the
orginal clusters after outlier removal. (default: \code{0.95}; range: from \code{0} to
\code{1})}

\item{initial_cmd}{logical; If TRUE, then the t-SNE map computation is
initialized with a configuration obtained by classical multidimensional
scaling.}

\item{perplexity}{integer; Perplexity of the t-SNE map. (default: \code{30L}; range:
from \code{5L} to \code{100L})}

\item{cthr}{integer; Clusters to be included into the StemID2 analysis must
contain more than cthr cells. D (default: \code{5L}; range: from \code{1L} to \code{25L})}

\item{nmode}{logical; If \code{TRUE}, then a cell of given cluster is assigned
to the link to the cluster with the smallest average distance of the knn
nearest neighbours within this cluster.}

\item{projcells_knn}{integer; See \code{nmode}. (default: \code{3L}; range: from
\code{3L} to \code{20L})}

\item{fr}{logical; Use Fruchterman-Rheingold layout instead of t-SNE for
dimensional-reduction representation of the lineage tree.}

\item{pdishuf}{integer; Number of randomizations of cell positions for which to
compute projections of cells on inter-cluster links. (default: \code{500L}; range:
from \code{10L} to \code{10000L})}

\item{fast}{logical; If \code{TRUE} and \code{nmode} is \code{FALSE} cells will
still be assigned to links based on maximum projections but a fast approximate
background model will be used to infer significance. The function will do
nothing in this case.}

\item{pthr}{numeric; P-value cutoff for link significance. This threshold is
applied for the calculation of link scores reflecting how uniformly a link is
occupied by cells. (default: \code{0.01}; range: from \code{1e-05} to \code{1L})}

\item{scthr}{numeric; Score threshold for links to be shown in the graph.
(default: \code{0.2}; range: from \code{0} to \code{1})}

\item{container_type}{In which environment to run the method, can be \code{"docker"} or \code{"singularity"}.
By default, docker will be used.}
}
\value{
A TI method wrapper to be used together with
\code{\link[dynwrap:infer_trajectories]{infer_trajectory}}
}
\description{
Will generate a trajectory using \href{https://doi.org/10.1016/j.stem.2016.05.010}{RaceID / StemID}.

This method was wrapped inside a
\href{https://github.com/dynverse/dynmethods/tree/master/containers/raceid_stemid}{container}.
The original code of this method is available
\href{https://github.com/dgrun/RaceID3_StemID2_package}{here}.
}
\references{
Grün, D., Muraro, M.J., Boisset, J.-C., Wiebrands, K., Lyubimova,
A., Dharmadhikari, G., van den Born, M., van Es, J., Jansen, E., Clevers, H.,
de Koning, E.J.P., van Oudenaarden, A., 2016. De Novo Prediction of Stem Cell
Identity using Single-Cell Transcriptome Data. Cell Stem Cell 19, 266–277.
}
