% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_random.R
\name{ti_random}
\alias{ti_random}
\title{Inferring a trajectory inference using Control: random}
\usage{
ti_random(dummy_param = 0.5, config = dynwrap::container_config())
}
\arguments{
\item{dummy_param}{numeric; Dummy parameter (default: \code{0.5}; range: from \code{0} to
\code{1})}

\item{config}{A container config. See \code{\link[=container_config]{container_config()}} for more information.}
}
\value{
A TI method wrapper to be used together with
\code{\link[dynwrap:infer_trajectories]{infer_trajectory}}
}
\description{
Will generate a trajectory using Control: random.

This method was wrapped inside a
\href{https://github.com/dynverse/dynmethods/tree/master/containers/random}{container}.
}
