% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_recat.R, R/ti_container.R
\name{ti_recat}
\alias{ti_recat}
\alias{ti_recat}
\title{Inferring trajectories with reCAT}
\usage{
ti_recat(docker = TRUE, TSPFold = 2, beginNum = 10, endNum = 15,
  step_size = 2, base_cycle_range_start = 6, base_cycle_range_end = 9,
  max_num = 300, clustMethod = "GMM")

ti_recat(docker = TRUE, TSPFold = 2, beginNum = 10, endNum = 15,
  step_size = 2, base_cycle_range_start = 6, base_cycle_range_end = 9,
  max_num = 300, clustMethod = "GMM")
}
\arguments{
\item{docker}{Whether to use the docker container or the R wrapper}

\item{TSPFold}{No documentation provided by authors}

\item{beginNum}{No documentation provided by authors}

\item{endNum}{No documentation provided by authors}

\item{step_size}{Determines the number of k to skip in your consensus path, ie if step_size = 2, then reCAT would only calculate and merge the paths for k = 12, 14, 16, 18, ', n-2, n. We recommend step_size of up to a maximum of 5 while preserving the performance of reCAT. Usually a step_size of 2 (by default) would suffice and bigger steps are recommended for larger datasets (>1000 cells) in order to reduce computational time.}

\item{base_cycle_range_start}{The minimal number of four k's for computing the reference cycle mentioned in the manuscript. Can be set to 6 or 7}

\item{base_cycle_range_end}{The maximal number of four k's for computing the reference cycle mentioned in the manuscript. Can be set to 6 or 7}

\item{max_num}{No documentation provided by authors}

\item{clustMethod}{No documentation provided by authors}
}
\value{
The trajectory model
}
\description{
Inferring trajectories with reCAT

Will generate a trajectory using \href{https://doi.org/10.1038/s41467-017-00039-z}{reCAT}. This method was wrapped inside a \href{https://github.com/dynverse/dynmethods/tree/master/containers/recat}{container}.
}
\details{
This methods was first wrapped inside R, see \link{ti_recat}

The original code of this method is available \href{https://github.com/tinglab/reCAT}{here}.

The method is described in: \href{https://doi.org/10.1038/s41467-017-00039-z}{Liu, Z., Lou, H., Xie, K., Wang, H., Chen, N., Aparicio, O.M., Zhang, M.Q., Jiang, R., Chen, T., 2017. Reconstructing cell cycle pseudo time-series via single-cell transcriptome data. Nature Communications 8.}
}
