% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_recat.R
\name{ti_recat}
\alias{ti_recat}
\title{Inferring a trajectory inference using reCAT}
\usage{
ti_recat(TSPFold = 2, beginNum = 10, endNum = 15, step_size = 2,
  base_cycle_range_start = 6, base_cycle_range_end = 9,
  max_num = 300, clustMethod = "GMM")
}
\arguments{
\item{TSPFold}{Parameter; No documentation provided by authors, Domain: U(2,
10), Default: 2, Format: integer.}

\item{beginNum}{Parameter; No documentation provided by authors, Domain: U(2,
20), Default: 10, Format: integer.}

\item{endNum}{Parameter; No documentation provided by authors, Domain: U(2,
20), Default: 15, Format: integer.}

\item{step_size}{Parameter; Determines the number of k to skip in your
consensus path, ie ifstep_size = 2, then reCAT would only calculate and merge
the paths fork = 12, 14, 16, 18, ..., n-2, n. We recommend step_size of up to a
maximum of 5 while preserving the performance of reCAT. Usually a step_size of
2 (by default) would suffice and bigger steps are recommended for larger
datasets (>1000 cells) in order to reduce computational time, Domain: U(2, 20),
Default: 2, Format: integer.}

\item{base_cycle_range_start}{Parameter; The minimal number of four k's for
computing the reference cycle mentioned in the manuscript. Can be set to 6 or
7, Domain: U(6, 7), Default: 6, Format: integer.}

\item{base_cycle_range_end}{Parameter; The maximal number of four k's for
computing the reference cycle mentioned in the manuscript. Can be set to 6 or
7, Domain: U(9, 10), Default: 9, Format: integer.}

\item{max_num}{Parameter; No documentation provided by authors, Domain: U(100,
500), Default: 300, Format: integer.}

\item{clustMethod}{Parameter; No documentation provided by authors, Domain:
{GMM, Pam, Kmeans}, Default: GMM, Format: character.}
}
\value{
A TI method wrapper to be used together with
\code{\link[dynwrap:infer_trajectories]{infer_trajectory}}
}
\description{
Will generate a trajectory using
\href{https://doi.org/10.1038/s41467-017-00039-z}{reCAT}.

This method was wrapped inside a
\href{https://github.com/dynverse/ti_recat}{container}.
The original code of this method is available
\href{https://github.com/tinglab/reCAT}{here}.
}
\references{
Liu, Z., Lou, H., Xie, K., Wang, H., Chen, N., Aparicio, O.M.,
Zhang, M.Q., Jiang, R., Chen, T., 2017. Reconstructing cell cycle pseudo
time-series via single-cell transcriptome data. Nature Communications 8.
}
