% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_recat.R
\name{ti_recat}
\alias{ti_recat}
\title{Inferring a trajectory inference using reCAT}
\usage{
ti_recat(TSPFold = 2, beginNum = 10, endNum = 15, step_size = 2,
  base_cycle_range_start = 6, base_cycle_range_end = 9,
  max_num = 300, clustMethod = "GMM", container_type = NULL)
}
\arguments{
\item{TSPFold}{integer; No documentation provided by authors (default: \code{2};
range: from \code{2} to \code{10})}

\item{beginNum}{integer; No documentation provided by authors (default: \code{10};
range: from \code{2} to \code{20})}

\item{endNum}{integer; No documentation provided by authors (default: \code{15};
range: from \code{2} to \code{20})}

\item{step_size}{integer; Determines the number of k to skip in your consensus
path, ie ifstep_size = 2, then reCAT would only calculate and merge the paths
fork = 12, 14, 16, 18, ..., n-2, n. We recommend step_size of up to a maximum
of 5 while preserving the performance of reCAT. Usually a step_size of 2 (by
default) would suffice and bigger steps are recommended for larger datasets
(>1000 cells) in order to reduce computational time. (default: \code{2}; range: from
\code{2} to \code{20})}

\item{base_cycle_range_start}{integer; The minimal number of four k's for
computing the reference cycle mentioned in the manuscript. Can be set to 6 or 7
(default: \code{6}; range: from \code{6} to \code{7})}

\item{base_cycle_range_end}{integer; The maximal number of four k's for
computing the reference cycle mentioned in the manuscript. Can be set to 6 or 7
(default: \code{9}; range: from \code{9} to \code{10})}

\item{max_num}{integer; No documentation provided by authors (default: \code{300};
range: from \code{100} to \code{500})}

\item{clustMethod}{discrete; No documentation provided by authors (default:
\code{"GMM"}; values: {\code{"GMM"}, \code{"Pam"}, \code{"Kmeans"}})}

\item{container_type}{In which environment to run the method, can be \code{"docker"} or \code{"singularity"}.
By default, docker will be used.}
}
\value{
A TI method wrapper to be used together with
\code{\link[dynwrap:infer_trajectories]{infer_trajectory}}
}
\description{
Will generate a trajectory using
\href{https://doi.org/10.1038/s41467-017-00039-z}{reCAT}.

This method was wrapped inside a
\href{https://github.com/dynverse/dynmethods/tree/master/containers/recat}{container}.
The original code of this method is available
\href{https://github.com/tinglab/reCAT}{here}.
}
\references{
Liu, Z., Lou, H., Xie, K., Wang, H., Chen, N., Aparicio, O.M.,
Zhang, M.Q., Jiang, R., Chen, T., 2017. Reconstructing cell cycle pseudo
time-series via single-cell transcriptome data. Nature Communications 8.
}
