% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_scimitar.R
\name{ti_scimitar}
\alias{ti_scimitar}
\title{SCIMITAR}
\usage{
ti_scimitar(
  covariance_type = "diag",
  degree = 3L,
  step_size = 0.07,
  cov_estimator = "identity",
  cov_reg = 0.05,
  max_iter = 3L
)
}
\arguments{
\item{covariance_type}{. Domain: {diag, spherical, full}. Default: diag.
Format: character.}

\item{degree}{. Domain: U(1, 20). Default: 3. Format: integer.}

\item{step_size}{. Domain: e^U(-4.61, -2.30). Default: 0.07. Format: numeric.}

\item{cov_estimator}{. Domain: {identity, diag, sample, global, glasso,
average}. Default: identity. Format: character.}

\item{cov_reg}{. Domain: e^U(-4.61, -2.30). Default: 0.05. Format: numeric.}

\item{max_iter}{. Domain: U(1, 20). Default: 3. Format: integer.}
}
\value{
A TI method wrapper to be used together with
\code{\link[dynwrap:infer_trajectories]{infer_trajectory}}
}
\description{
Will generate a trajectory using
\href{https://doi.org/10.1142/9789813207813_0053}{SCIMITAR}.

This method was wrapped inside a
\href{https://github.com/dynverse/ti_scimitar}{container}.
The original code of this method is available
\href{https://github.com/dimenwarper/scimitar}{here}.
}
\references{
CORDERO, P., STUART, J.M., 2016. TRACING CO-REGULATORY NETWORK
DYNAMICS IN NOISY, SINGLE-CELL TRANSCRIPTOME TRAJECTORIES. Biocomputing 2017.
}
\keyword{method}
