% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_scimitar.R
\name{ti_scimitar}
\alias{ti_scimitar}
\title{Inferring a trajectory inference using SCIMITAR}
\usage{
ti_scimitar(covariance_type = "diag", degree = 3L, step_size = 0.07,
  cov_estimator = "corpcor", cov_reg = 0.05, max_iter = 3L,
  config = dynwrap::container_config())
}
\arguments{
\item{covariance_type}{discrete; (default: \code{"diag"}; values: {\code{"diag"},
\code{"spherical"}, \code{"full"}})}

\item{degree}{integer; (default: \code{3L}; range: from \code{1L} to \code{20L})}

\item{step_size}{numeric; (default: \code{0.07}; range: from \code{0.01} to \code{0.1})}

\item{cov_estimator}{discrete; (default: \code{"corpcor"}; values: {\code{"identity"},
\code{"diag"}, \code{"sample"}, \code{"global"}, \code{"glasso"}, \code{"corpcor"}, \code{"average"}})}

\item{cov_reg}{numeric; (default: \code{0.05}; range: from \code{0.01} to \code{0.1})}

\item{max_iter}{integer; (default: \code{3L}; range: from \code{1L} to \code{20L})}

\item{config}{A container config. See \code{\link[=container_config]{container_config()}} for more information.}
}
\value{
A TI method wrapper to be used together with
\code{\link[dynwrap:infer_trajectories]{infer_trajectory}}
}
\description{
Will generate a trajectory using
\href{https://doi.org/10.1142/9789813207813_0053}{SCIMITAR}.

This method was wrapped inside a
\href{https://github.com/dynverse/ti_scimitar}{container}.
The original code of this method is available
\href{https://github.com/dimenwarper/scimitar}{here}.
}
\references{
CORDERO, P., STUART, J.M., 2016. TRACING CO-REGULATORY NETWORK
DYNAMICS IN NOISY, SINGLE-CELL TRANSCRIPTOME TRAJECTORIES. Biocomputing 2017.
}
