% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_scorpius.R, R/ti_container.R
\name{ti_scorpius}
\alias{ti_scorpius}
\alias{ti_scorpius_sparse}
\alias{ti_scorpius}
\title{Inferring trajectories with SCORPIUS}
\usage{
ti_scorpius(docker = "stevedore" \%in\% rownames(installed.packages()),
  distance_method = "spearman", ndim = 3, k = 4, thresh = 0.001,
  maxit = 10, stretch = 0, smoother = "smooth.spline", sparse = FALSE)

ti_scorpius_sparse(distance_method = "spearman", ndim = 3, k = 4,
  thresh = 0.001, maxit = 10, stretch = 0, smoother = "smooth.spline",
  sparse = TRUE)

ti_scorpius(docker = "stevedore" \%in\% rownames(installed.packages()),
  distance_method = "spearman", ndim = 3, k = 4, thresh = 0.001,
  maxit = 10, stretch = 0, smoother = "smooth.spline", sparse = FALSE)
}
\arguments{
\item{docker}{Whether to use the docker container or the R wrapper}

\item{distance_method}{A character string indicating which correlation
coefficient (or covariance) is to be computed. One of "pearson", "kendall", or "spearman".}

\item{ndim}{The number of dimensions in the new space.}

\item{k}{The number of clusters to cluster the data into.}

\item{thresh}{\code{\link[princurve]{principal.curve}} parameter: convergence threshhold on shortest distances to the curve}

\item{maxit}{\code{\link[princurve]{principal.curve}} parameter: maximum number of iterations}

\item{stretch}{\code{\link[princurve]{principal.curve}} parameter: a factor by which the curve can be extrapolated when points are projected}

\item{smoother}{\code{\link[princurve]{principal.curve}} parameter: choice of smoother}

\item{sparse}{Whether or not to use sparse MDS dimensionality reduction,
for datasets with large amounts of cells.}
}
\value{
The trajectory model
}
\description{
Inferring trajectories with SCORPIUS

Will generate a trajectory using \href{https://doi.org/10.1101/079509}{SCORPIUS}. This method was wrapped inside a \href{https://github.com/dynverse/dynmethods/tree/master/containers/scorpius}{container}.
}
\details{
This methods was first wrapped inside R, see \link{ti_scorpius}

The original code of this method is available \href{https://github.com/rcannood/SCORPIUS}{here}.

The method is described in: \href{https://doi.org/10.1101/079509}{Cannoodt, R., Saelens, W., Sichien, D., Tavernier, S., Janssens, S., Guilliams, M., Lambrecht, B.N., De Preter, K., Saeys, Y., 2016. SCORPIUS improves trajectory inference and identifies novel modules in dendritic cell development.}
}
