% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_scoup.R, R/ti_container.R
\name{ti_scoup}
\alias{ti_scoup}
\alias{ti_scoup}
\title{Inferring trajectories with SCOUP}
\usage{
ti_scoup(docker = dynwrap::test_docker_installation(), ndim = 2,
  max_ite1 = 100, max_ite2 = 100, alpha_min = 0.1, alpha_max = 100,
  t_min = 0.001, t_max = 2, sigma_squared_min = 0.1, thresh = 0.01)

ti_scoup(docker = dynwrap::test_docker_installation(), ndim = 2,
  max_ite1 = 100, max_ite2 = 100, alpha_min = 0.1, alpha_max = 100,
  t_min = 0.001, t_max = 2, sigma_squared_min = 0.1, thresh = 0.01)
}
\arguments{
\item{docker}{Whether to use the docker container or the R wrapper}

\item{ndim}{Number of pca dimensions}

\item{max_ite1}{Upper bound of EM iteration (without pseudo-time optimization). The detailed explanation is described in the supplementary text. (default is 1,000)}

\item{max_ite2}{Upper bound of EM iteration (including pseudo-time optimization) (default is 1,000).}

\item{alpha_min}{Lower bound of alpha (default is 0.1)}

\item{alpha_max}{Upper bound of alpha (default is 100)}

\item{t_min}{Lower bound of pseudo-time (default is 0.001)}

\item{t_max}{Upper bound of pseudo-time (default is 2.0)}

\item{sigma_squared_min}{Lower bound of sigma squared (default is 0.1)}

\item{thresh}{Threshold}
}
\value{
The trajectory model
}
\description{
Inferring trajectories with SCOUP

Will generate a trajectory using \href{https://doi.org/10.1186/s12859-016-1109-3}{SCOUP}. This method was wrapped inside a \href{https://github.com/dynverse/dynmethods/tree/master/containers/scoup}{container}.
}
\details{
This methods was first wrapped inside R, see \link{ti_scoup}

The original code of this method is available \href{https://github.com/hmatsu1226/SCOUP}{here}.

The method is described in: \href{https://doi.org/10.1186/s12859-016-1109-3}{Matsumoto, H., Kiryu, H., 2016. SCOUP: a probabilistic model based on the Ornstein–Uhlenbeck process to analyze single-cell expression data during differentiation. BMC Bioinformatics 17.}
}
