% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_scoup.R
\name{ti_scoup}
\alias{ti_scoup}
\title{Inferring a trajectory inference using SCOUP}
\usage{
ti_scoup(ndim = 2L, max_ite1 = 100L, max_ite2 = 100L,
  alpha = c(0.1, 100), t = c(0.001, 2), sigma_squared = 0.1,
  thresh = 0.01)
}
\arguments{
\item{ndim}{Parameter; Number of pca dimensions., Domain: U(2, 20), Default: 2,
Format: integer.}

\item{max_ite1}{Parameter; Upper bound of EM iteration (without pseudo-time
optimization)., Domain: e^U(0.69, 8.52), Default: 100, Format: integer.}

\item{max_ite2}{Parameter; Upper bound of EM iteration (including pseudo-time
optimization)., Domain: e^U(0.69, 13.12), Default: 100, Format: integer.}

\item{alpha}{Parameter; Bounds of alpha., Domain: ( e^U(-6.91, 2.30),
e^U(-6.91, 2.30) ), Default: (0.1, 100), Format: range.}

\item{t}{Parameter; Bounds of pseudo-time., Domain: ( e^U(-11.51, 0.00),
e^U(-11.51, 0.00) ), Default: (0.001, 2), Format: range.}

\item{sigma_squared}{Parameter; Lower bound of sigma squared., Domain:
e^U(-6.91, 2.30), Default: 0.1, Format: numeric.}

\item{thresh}{Parameter; Threshold., Domain: e^U(-4.61, 2.30), Default: 0.01,
Format: numeric.}
}
\value{
A TI method wrapper to be used together with
\code{\link[dynwrap:infer_trajectories]{infer_trajectory}}
}
\description{
Will generate a trajectory using
\href{https://doi.org/10.1186/s12859-016-1109-3}{SCOUP}.

This method was wrapped inside a
\href{https://github.com/dynverse/ti_scoup}{container}.
The original code of this method is available
\href{https://github.com/hmatsu1226/SCOUP}{here}.
}
\references{
Matsumoto, H., Kiryu, H., 2016. SCOUP: a probabilistic model based
on the Ornstein–Uhlenbeck process to analyze single-cell expression data during
differentiation. BMC Bioinformatics 17.
}
