% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_scuba.R
\name{ti_scuba}
\alias{ti_scuba}
\title{Inferring a trajectory inference using \href{https://doi.org/10.1073/pnas.1408993111}{scuba}}
\usage{
ti_scuba(rigorous_gap_stats = TRUE, N_dim = 2L, low_gene_threshold = 1L,
  low_gene_fraction_max = 0.7, min_split = 15L,
  min_percentage_split = 0.25)
}
\arguments{
\item{rigorous_gap_stats}{Whether to use rigorous gap statistics to determine number of clusters \cr}

\item{N_dim}{Number of TSNE dimensions \cr
integer; default: 2L; possible values between 2 and 3}

\item{low_gene_threshold}{Threshold value for genes of low expression levels \cr
numeric; default: 1L; possible values between 0 and 5}

\item{low_gene_fraction_max}{Maximum fraction of lowly-expressed cells allowed for each gene \cr
numeric; default: 0.7; possible values between 0 and 1}

\item{min_split}{Lower threshold on the number of cells in a cluster for this cluster to be split. \cr
integer; default: 15L; possible values between 1 and 100}

\item{min_percentage_split}{Minimum fraction of cells in the smaller cluster during a bifurcation. \cr
numeric; default: 0.25; possible values between 0 and 1}
}
\value{
The trajectory model
}
\description{
Will generate a trajectory using \href{https://doi.org/10.1073/pnas.1408993111}{scuba}. This method was wrapped inside a \href{https://github.com/dynverse/dynmethods/tree/master/containers/scuba}{container}.
}
\details{
The original code of this method is available \href{https://github.com/gcyuan/SCUBA}{here}.

The method is described in: \href{https://doi.org/10.1073/pnas.1408993111}{Marco, E., Karp, R.L., Guo, G., Robson, P., Hart, A.H., Trippa, L., Yuan, G.-C., 2014. Bifurcation analysis of single-cell gene expression data reveals epigenetic landscape. Proceedings of the National Academy of Sciences 111, E5643–E5650.}
}
