% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_scuba.R
\name{ti_scuba}
\alias{ti_scuba}
\title{Inferring a trajectory inference using SCUBA}
\usage{
ti_scuba(rigorous_gap_stats = TRUE, N_dim = 2L, low_gene_threshold = 1L,
  low_gene_fraction_max = 0.7, min_split = 15L,
  min_percentage_split = 0.25, run_environment = NULL)
}
\arguments{
\item{rigorous_gap_stats}{logical; Whether to use rigorous gap statistics to
determine number of clusters}

\item{N_dim}{integer; Number of TSNE dimensions (default: \code{2L}; range: from
\code{2L} to \code{3L})}

\item{low_gene_threshold}{numeric; Threshold value for genes of low expression
levels (default: \code{1L}; range: from \code{0L} to \code{5L})}

\item{low_gene_fraction_max}{numeric; Maximum fraction of lowly-expressed cells
allowed for each gene (default: \code{0.7}; range: from \code{0L} to \code{1L})}

\item{min_split}{integer; Lower threshold on the number of cells in a cluster
for this cluster to be split. (default: \code{15L}; range: from \code{1L} to \code{100L})}

\item{min_percentage_split}{numeric; Minimum fraction of cells in the smaller
cluster during a bifurcation. (default: \code{0.25}; range: from \code{0L} to \code{1L})}

\item{run_environment}{In which environment to run the method, can be \code{"docker"} or \code{"singularity"}.}
}
\value{
A TI method wrapper to be used together with
\code{\link[dynwrap:infer_trajectories]{infer_trajectory}}
}
\description{
Will generate a trajectory using
\href{https://doi.org/10.1073/pnas.1408993111}{SCUBA}.

This method was wrapped inside a
\href{https://github.com/dynverse/dynmethods/tree/master/containers/scuba}{container}.
The original code of this method is available
\href{https://github.com/gcyuan/SCUBA}{here}.
}
\references{
Marco, E., Karp, R.L., Guo, G., Robson, P., Hart, A.H., Trippa, L.,
Yuan, G.-C., 2014. Bifurcation analysis of single-cell gene expression data
reveals epigenetic landscape. Proceedings of the National Academy of Sciences
111, E5643–E5650.
}
