% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_shuffle.R, R/ti_container.R
\name{ti_shuffle}
\alias{ti_shuffle}
\alias{ti_shuffle}
\title{Inferring trajectories with Control: shuffle}
\usage{
ti_shuffle(docker = TRUE, dummy_param = 0.5)

ti_shuffle(docker = TRUE, dummy_param = 0.5)
}
\arguments{
\item{docker}{Whether to use the docker container or the R wrapper}

\item{dummy_param}{This parameter does not do anything.}
}
\value{
The trajectory model
}
\description{
This control method will return the milestone network of the provided
gold standard, but will shuffle the cell positions randomly.

Will generate a trajectory using Control: shuffle. This method was wrapped inside a \href{https://github.com/dynverse/dynmethods/tree/master/containers/shuffle}{container}.
}
\details{
This methods was first wrapped inside R, see \link{ti_shuffle}
}
