% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_shuffle.R
\name{ti_shuffle}
\alias{ti_shuffle}
\title{Inferring a trajectory inference using Control: shuffle}
\usage{
ti_shuffle(dummy_param = 0.5)
}
\arguments{
\item{dummy_param}{numeric; Dummy parameter (default: \code{0.5}; range: from \code{0} to
\code{1})}
}
\value{
A TI method wrapper to be used together with
\code{\link[dynwrap:infer_trajectories]{infer_trajectory}}
}
\description{
Will generate a trajectory using Control: shuffle.

This method was wrapped inside a
\href{https://github.com/dynverse/ti_shuffle}{container}.
}
