% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_slicer.R
\name{ti_slicer}
\alias{ti_slicer}
\title{Inferring a trajectory inference using SLICER}
\usage{
ti_slicer(kmin = 10L, m = 2L)
}
\arguments{
\item{kmin}{Parameter; Smallest value of k to try., Domain: U(2, 20), Default:
10, Format: integer.}

\item{m}{Parameter; Intrinsic dimension of the data. This parameter mainly
influences the visualisation of the results. The real intrinsic dimension will
be calculated automaticly., Domain: U(2, 20), Default: 2, Format: integer.}
}
\value{
A TI method wrapper to be used together with
\code{\link[dynwrap:infer_trajectories]{infer_trajectory}}
}
\description{
Will generate a trajectory using
\href{https://doi.org/10.1186/s13059-016-0975-3}{SLICER}.

This method was wrapped inside a
\href{https://github.com/dynverse/ti_slicer}{container}.
The original code of this method is available
\href{https://github.com/jw156605/SLICER}{here}.
}
\references{
Welch, J.D., Hartemink, A.J., Prins, J.F., 2016. SLICER: inferring
branched, nonlinear cellular trajectories from single cell RNA-seq data. Genome
Biology 17.
}
