% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_stemid.R
\name{ti_stemid}
\alias{ti_stemid}
\title{Inferring a trajectory inference using StemID}
\usage{
ti_stemid(clustnr = 30L, bootnr = 50L, metric = "pearson",
  num_cluster_method = "sat", SE.method = "Tibs2001SEmax",
  SE.factor = 0.25, B.gap = 50L, cln = 30L, FUNcluster = "kmedoids",
  dimred_method = "tsne", outminc = 0L, outlg = 2L, probthr = 0.001,
  thr_lower = -10, thr_upper = -5, outdistquant = 0.95, nmode = FALSE,
  pdishuf = 2000L, pthr = 0.01, pethr = 0.01, pvalue_cutoff = 0.05,
  linkscore_cutoff = 0.2, run_environment = NULL)
}
\arguments{
\item{clustnr}{integer; Maximum number of clusters for the computation of the
gap statistic or derivation of the cluster number by the saturation criterion.
Default is 30. If more major cell types are expected a higher number should be
chosen. (default: \code{30L}; range: from \code{10L} to \code{100L})}

\item{bootnr}{integer; Number of booststrapping runs for clusterboot. Default
is 50 (default: \code{50L}; range: from \code{20L} to \code{100L})}

\item{metric}{discrete; The input data are transformed to a distance object.
Distances can be computed based on different metrics. Possible values are
"pearson", "spearman", "logpearson", "euclidean", "kendall", "maximum",
"manhattan", "canberra", "binary" or "minkowski". Default is "pearson". In case
of the correlation based methods, the distance is computed as 1 - correlation.
K-medoids clustering is performed on this distance object. (default:
\code{"pearson"}; values: {\code{"pearson"}, \code{"spearman"}, \code{"kendall"}, \code{"euclidean"},
\code{"maximum"}, \code{"manhattan"}, \code{"canberra"}, \code{"binary"}, \code{"minkowski"}})}

\item{num_cluster_method}{discrete; The type of clustering method, can be sat,
gap or manual (default: \code{"sat"}; values: {\code{"sat"}, \code{"gap"}, \code{"manual"}})}

\item{SE.method}{discrete; The clustering routine calls a modified version of
the maxSE function from the cluster package to determine the first local
maximum of the gap statistic. By default, we use the method "Tibs2001SEmax" for
calling the first local maximum (see specification of maxSE). This method
requires that the maximum exceeds the values of its neighbors by a fraction of
their standard deviation. This fraction is defined by the parameter SE.factor.
All methods defined for the original maxSE function can also be used. (default:
\code{"Tibs2001SEmax"}; values: {\code{"firstSEmax"}, \code{"Tibs2001SEmax"}, \code{"globalSEmax"},
\code{"firstmax"}, \code{"globalmax"}})}

\item{SE.factor}{numeric; Fraction of the standard deviation by which the local
maximum is required to differ from the neighboring points it is compared to.
Default is 0.25. (default: \code{0.25}; range: from \code{0} to \code{1})}

\item{B.gap}{integer; Number of bootstrap runs for the calculation of the gap
statistic. Default is 50. (default: \code{50L}; range: from \code{20L} to \code{100L})}

\item{cln}{integer; The number of clusters for k-medoids clustering. Default is
0. In this case, the cluster number is determined based on the gap statistic6
and do.gap has to be TRUE. (default: \code{30L}; range: from \code{20L} to \code{100L})}

\item{FUNcluster}{discrete; The clustering method applied. One of the following
methods can be selected: kmedoids, kmeans, hclust. RaceID3 is designed for
k-medoids clustering and therefore it is recommended to use only the kmedoids
method. Default is kmedoids. (default: \code{"kmedoids"}; values: {\code{"kmedoids"},
\code{"kmeans"}, \code{"hclust"}})}

\item{dimred_method}{discrete; The dimensionality reduction method, can be
tsne, sammon or tsne_initcmd (default: \code{"tsne"}; values: {\code{"tsne"}, \code{"sammon"},
\code{"tsne_initcmd"}})}

\item{outminc}{integer; Expression cutoff for the identification of outlier
genes is defined. Default is 5. (default: \code{0L}; range: from \code{0L} to \code{100L})}

\item{outlg}{integer; Minimal number of outlier genes required to identify a
cell as an outlier. Default is 2. (default: \code{2L}; range: from \code{0L} to \code{100L})}

\item{probthr}{numeric; Defines the probability threshold for outlier calling.
If the probability of observing a given expression level for a gene in a cell
is lower than this cutoff (based on the negative binomial distribution for the
calibrated noise model), the cell is considered an outlier for this gene.
Default is 10-3. (default: \code{0.001}; range: from \code{1e-10} to \code{0.1})}

\item{thr_lower}{integer; Lower probability for which the number of outliers is
computed in order to plot the dependence of the number of outliers on the
probability threshold (default: \code{-10}; range: from \code{-100} to \code{-1})}

\item{thr_upper}{numeric; Upper probability for which the number of outliers is
computed in order to plot the dependence of the number of outliers on the
probability threshold (default: \code{-5}; range: from \code{-100} to \code{-1})}

\item{outdistquant}{numeric; Outlier cells are merged to outlier clusters if
their similarity exceeds the outdistquant-quantile of the similarity
distribution for all pairs of cells that are together in one of the original
clusters. Default is 0.95. (default: \code{0.95}; range: from \code{0} to \code{1})}

\item{nmode}{logical; Boolean argument. If nmode is set to TRUE the assignment
to inter-cluster links for each cell is not done based on the longest
projection, but based on identifying the cluster (other than the cluster the
cell belongs to) that contains the nearest neighbor of the cell, i. e. the cell
with the most similar transcriptome. The coordinate on the assigned link is
still derived based on the projection. Default is FALSE.}

\item{pdishuf}{integer; Positive integer. This is the number of randomizations
to be performed. As a rule of thumb this number should be at least one order of
magnitude larger than the desired p-value on the significance of the number of
cells on a connection. Default is 2000. (default: \code{2000L}; range: from \code{100L}
to \code{10000L})}

\item{pthr}{numeric; Positive number. This number corresponds to the p-value
threshold, which is used to determine, whether the magnitude of an observed
trajectory is significantly larger than observed for the randomized background
distribution. This criterion is not used to infer significance of a link, but
shown in a graphical representation of the tree (default: \code{0.01}; range: from
\code{0} to \code{1e-04})}

\item{pethr}{numeric; Positive number. This number corresponds to the p-value
threshold, which isused to determine for each link if it is populated by a
number of cellssignificantly larger than expected for the randomized background
distribution. This p-value threshold determines, which connections are
considered validdifferentiation trajectories in the derived lineage tree.
(default: \code{0.01}; range: from \code{0} to \code{1e-04})}

\item{pvalue_cutoff}{numeric; Upper cutoff of p-value for determining
significant edges (default: \code{0.05}; range: from \code{0} to \code{1})}

\item{linkscore_cutoff}{numeric; Lower cutoff of the linkscore for determining
significant edges (default: \code{0.2}; range: from \code{0} to \code{1})}

\item{run_environment}{In which environment to run the method, can be \code{"docker"} or \code{"singularity"}.}
}
\value{
A TI method wrapper to be used together with
\code{\link[dynwrap:infer_trajectories]{infer_trajectory}}
}
\description{
Will generate a trajectory using
\href{https://doi.org/10.1016/j.stem.2016.05.010}{StemID}.

This method was wrapped inside a
\href{https://github.com/dynverse/dynmethods/tree/master/containers/stemid}{container}.
The original code of this method is available
\href{https://github.com/dgrun/StemID}{here}.
}
\references{
Grün, D., Muraro, M.J., Boisset, J.-C., Wiebrands, K., Lyubimova,
A., Dharmadhikari, G., van den Born, M., van Es, J., Jansen, E., Clevers, H.,
de Koning, E.J.P., van Oudenaarden, A., 2016. De Novo Prediction of Stem Cell
Identity using Single-Cell Transcriptome Data. Cell Stem Cell 19, 266–277.
}
