% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_stemid2.R
\name{ti_stemid2}
\alias{ti_stemid2}
\title{Inferring trajectories with StemID2}
\usage{
ti_stemid2(clustnr = 30L, bootnr = 50L, metric = "pearson",
  num_cluster_method = "sat", SE.method = "Tibs2001SEmax",
  SE.factor = 0.25, B.gap = 50L, cln = 30L, FUNcluster = "kmedoids",
  dimred_method = "tsne", outminc = 0L, outlg = 2L, probthr = 1e-10,
  thr_lower = -40, thr_upper = -1, outdistquant = 0.95, nmode = FALSE,
  pdishuf = 500, pthr = 0.01, pethr = 0.01)
}
\description{
Arguments passed to this function will be used as default parameters for the method.
}
