% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_stemnet.R
\name{ti_stemnet}
\alias{ti_stemnet}
\title{Inferring a trajectory inference using STEMNET}
\usage{
ti_stemnet(alpha = 0.1, lambda_auto = TRUE, lambda = 0.1,
  run_environment = NULL)
}
\arguments{
\item{alpha}{numeric; The elastic net mixing parameter of the ‘glmnet’
classifier. (default: \code{0.1}; range: from \code{0.001} to \code{10L})}

\item{lambda_auto}{logical; Whether to select the lambda by cross-validation}

\item{lambda}{numeric; The lambda penalty of GLM. (default: \code{0.1}; range: from
\code{0.05} to \code{1L})}

\item{run_environment}{In which environment to run the method, can be \code{"docker"} or \code{"singularity"}.}
}
\value{
A TI method wrapper to be used together with
\code{\link[dynwrap:infer_trajectories]{infer_trajectory}}
}
\description{
Will generate a trajectory using \href{https://doi.org/10.1038/ncb3493}{STEMNET}.

This method was wrapped inside a
\href{https://github.com/dynverse/dynmethods/tree/master/containers/stemnet}{container}.
The original code of this method is available
\href{https://git.embl.de/velten/STEMNET}{here}.
}
\references{
Velten, L., Haas, S.F., Raffel, S., Blaszkiewicz, S., Islam, S.,
Hennig, B.P., Hirche, C., Lutz, C., Buss, E.C., Nowak, D., Boch, T., Hofmann,
W.-K., Ho, A.D., Huber, W., Trumpp, A., Essers, M.A.G., Steinmetz, L.M., 2017.
Human haematopoietic stem cell lineage commitment is a continuous process.
Nature Cell Biology 19, 271–281.
}
