% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_stemnet.R
\name{ti_stemnet}
\alias{ti_stemnet}
\title{STEMNET}
\usage{
ti_stemnet(alpha = 0.1, lambda_auto = TRUE, lambda = 0.1, force = FALSE)
}
\arguments{
\item{alpha}{The elastic net mixing parameter of the ‘glmnet’ classifier.
Domain: e^U(-6.91, 2.30). Default: 0.1. Format: numeric.}

\item{lambda_auto}{Whether to select the lambda by cross-validation. Default:
TRUE. Format: logical.}

\item{lambda}{The lambda penalty of GLM. Domain: e^U(-3.00, 0.00). Default:
0.1. Format: numeric.}

\item{force}{Do not use! This is a parameter to force STEMNET to run on
benchmark datasets where not enough end groups are present. Default: FALSE.
Format: logical.}
}
\value{
A TI method wrapper to be used together with
\code{\link[dynwrap:infer_trajectories]{infer_trajectory}}
}
\description{
Will generate a trajectory using \href{https://doi.org/10.1038/ncb3493}{STEMNET}.

This method was wrapped inside a
\href{https://github.com/dynverse/ti_stemnet}{container}.
The original code of this method is available
\href{https://git.embl.de/velten/STEMNET}{here}.
}
\references{
Velten, L., Haas, S.F., Raffel, S., Blaszkiewicz, S., Islam, S.,
Hennig, B.P., Hirche, C., Lutz, C., Buss, E.C., Nowak, D., Boch, T., Hofmann,
W.-K., Ho, A.D., Huber, W., Trumpp, A., Essers, M.A.G., Steinmetz, L.M., 2017.
Human haematopoietic stem cell lineage commitment is a continuous process.
Nature Cell Biology 19, 271–281.
}
\keyword{method}
