% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_topslam.R
\name{ti_topslam}
\alias{ti_topslam}
\title{Inferring a trajectory inference using topslam}
\usage{
ti_topslam(n_components = 2L, n_neighbors = 10L, linear_dims = 0L,
  max_iters = 1000L, dimreds = c(TRUE, TRUE, TRUE, TRUE, TRUE),
  config = dynwrap::container_config())
}
\arguments{
\item{n_components}{integer; The number of components (default: \code{2L}; range:
from \code{2L} to \code{10L})}

\item{n_neighbors}{integer; The number of neighbors (default: \code{10L}; range:
from \code{2L} to \code{100L})}

\item{linear_dims}{integer; (default: \code{0L}; range: from \code{0L} to \code{5L})}

\item{max_iters}{integer; The number of iterations to optimize over (default:
\code{1000L}; range: from \code{10L} to \code{10000L})}

\item{dimreds}{logical_vector; Which dimensionality reductions to use; tSNE,
PCA, Spectral, Isomap and/or ICA}

\item{config}{A container config. See \code{\link[=container_config]{container_config()}} for more information.}
}
\value{
A TI method wrapper to be used together with
\code{\link[dynwrap:infer_trajectories]{infer_trajectory}}
}
\description{
Will generate a trajectory using \href{https://doi.org/10.1101/057778}{topslam}.

This method was wrapped inside a
\href{https://github.com/dynverse/ti_topslam}{container}.
The original code of this method is available
\href{https://github.com/mzwiessele/topslam}{here}.
}
\references{
Zwiessele, M., Lawrence, N.D., 2016. Topslam: Waddington Landscape
Recovery for Single Cell Experiments.
}
