% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_topslam.R
\name{ti_topslam}
\alias{ti_topslam}
\title{Inferring a trajectory inference using topslam}
\usage{
ti_topslam(n_components = 2L, n_neighbors = 10L, linear_dims = 0L,
  max_iters = 1000L, dimreds = c("t-SNE", "PCA", "Spectral", "Isomap",
  "ICA"))
}
\arguments{
\item{n_components}{The number of components. Domain: U(2, 10). Default: 2.
Format: integer.}

\item{n_neighbors}{The number of neighbors. Domain: U(2, 100). Default: 10.
Format: integer.}

\item{linear_dims}{. Domain: U(0, 5). Default: 0. Format: integer.}

\item{max_iters}{The number of iterations to optimize over. Domain: e^U(2.30,
9.21). Default: 1000. Format: integer.}

\item{dimreds}{Which dimensionality reductions to use and/or ICA. Domain: {x |
x ⊆ {t-SNE, PCA, Spectral, Isomap, ICA}}. Default: {t-SNE, PCA, Spectral,
Isomap, ICA}. Format: subset.}
}
\value{
A TI method wrapper to be used together with
\code{\link[dynwrap:infer_trajectories]{infer_trajectory}}
}
\description{
Will generate a trajectory using \href{https://doi.org/10.1101/057778}{topslam}.

This method was wrapped inside a
\href{https://github.com/dynverse/ti_topslam}{container}.
The original code of this method is available
\href{https://github.com/mzwiessele/topslam}{here}.
}
\references{
Zwiessele, M., Lawrence, N.D., 2016. Topslam: Waddington Landscape
Recovery for Single Cell Experiments.
}
