% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_tscan.R, R/ti_container.R
\name{ti_tscan}
\alias{ti_tscan}
\alias{ti_tscan}
\title{Inferring trajectories with TSCAN}
\usage{

  ti_tscan(run_environment = ifelse(is.null(getOption("dynwrap_run_environment")),
  "local", getOption("dynwrap_run_environment")), minexpr_percent = 0,
  minexpr_value = 0, cvcutoff = 0, clusternum_lower = 2,
  clusternum_upper = 9, modelNames = "VVV")


  ti_tscan(run_environment = ifelse(is.null(getOption("dynwrap_run_environment")),
  "local", getOption("dynwrap_run_environment")), minexpr_percent = 0,
  minexpr_value = 0, cvcutoff = 0, clusternum_lower = 2,
  clusternum_upper = 9, modelNames = "VVV")
}
\arguments{
\item{run_environment}{In which environment to run the method, can be 'local', 'docker' or 'singularity'}

\item{minexpr_percent}{Numeric value specifying the lowest percentage of highly expressed cells (expression value bigger than minexpr_value) for the genes/features to be retained.}

\item{minexpr_value}{Numeric value specifying the minimum cutoff of log transformed (if takelog is TRUE) value}

\item{cvcutoff}{Numeric value specifying the minimum value of coefficient of variance for the genes/features to be retained.}

\item{clusternum_lower}{An integer specifying the minimal possible cluster numbers. The best cluster number will be picked using BIC.}

\item{clusternum_upper}{An integer specifying the maximal possible cluster numbers. The best cluster number will be picked using BIC.}

\item{modelNames}{model to be used in model-based clustering. By default "ellipsoidal, varying volume, shape, and orientation" is used.}
}
\value{
The trajectory model
}
\description{
Inferring trajectories with TSCAN

Will generate a trajectory using \href{https://doi.org/10.1093/nar/gkw430}{TSCAN}. This method was wrapped inside a \href{https://github.com/dynverse/dynmethods/tree/master/containers/tscan}{container}.
}
\details{
This methods was first wrapped inside R, see \link{ti_tscan}

The original code of this method is available \href{https://github.com/zji90/TSCAN}{here}.

The method is described in: \href{https://doi.org/10.1093/nar/gkw430}{Ji, Z., Ji, H., 2016. TSCAN: Pseudo-time reconstruction and evaluation in single-cell RNA-seq analysis. Nucleic Acids Research 44, e117–e117.}
}
