% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_urd.R
\name{ti_urd}
\alias{ti_urd}
\title{Inferring a trajectory inference using URD}
\usage{
ti_urd(knn = 0L, sigma.use = 0, distance = "euclidean",
  n_floods = 20L, stability.div = 10L, mp.factor = 1L,
  perplexity = 30L, theta = 0.5, max_iter = 1000L, num.nn = 30L,
  do.jaccard = TRUE, optimal.cells.forward = 20L,
  max.cells.back = 40L, n.per.tip = 25000L, root.visits = 1L,
  max.steps = 25000L, n.subsample = 10L, divergence.method = "ks",
  cells.per.pseudotime.bin = 80L, bins.per.pseudotime.window = 5L,
  p.thresh = 0.01)
}
\arguments{
\item{knn}{integer; Number of nearest neighbors to use. \code{0} takes a guess.
(default: \code{0L}; range: from \code{0L} to \code{50L})}

\item{sigma.use}{numeric; Kernel width to use for the diffusion map. \code{0} uses
destiny's global auto-detection procedure. (default: \code{0}; range: from \code{0} to
\code{1})}

\item{distance}{discrete; Distance metric to use for determining transition
probabilities. (default: \code{"euclidean"}; values: {\code{"euclidean"}, \code{"cosine"},
\code{"rankcor"}})}

\item{n_floods}{integer; Number of simulations to perform and average.
(default: \code{20L}; range: from \code{5L} to \code{50L})}

\item{stability.div}{numeric; Number of simulation subsamplings to calculate.
(default: \code{10L}; range: from \code{2L} to \code{50L})}

\item{mp.factor}{numeric; Retain PCs than are this factor more than the
estimated maximum singular value expected or random data. This is useful in
cases when there are many PCs that have standard deviations just above that
expected by random, which probably represent noise and should be excluded.
(default: \code{1L}; range: from \code{0L} to \code{10L})}

\item{perplexity}{numeric; Perplexity parameter for the tSNE. (default: \code{30L};
range: from \code{0L} to \code{100L})}

\item{theta}{numeric; Speed/accuracy trade-off for Barnes-Hut approximation of
tSNE. 0 is exact tSNE, higher is less accurate. (default: \code{0.5}; range: from
\code{0L} to \code{1L})}

\item{max_iter}{integer; Number of nearest neighbors to use. \code{0} takes a guess.
(default: \code{1000L}; range: from \code{100L} to \code{10000L})}

\item{num.nn}{integer; How many nearest-neighbors to use in the k-nn graph.
(default: \code{30L}; range: from \code{10L} to \code{100L})}

\item{do.jaccard}{logical; Weight edges in the k-nn graph according to their
Jaccard overlap?}

\item{optimal.cells.forward}{numeric; The number of cells in the direction
specified by pseudotime.direction at which the logistic should reach
1-asymptote. (default: \code{20L}; range: from \code{5L} to \code{100L})}

\item{max.cells.back}{numeric; The number of cells in the direction opposite
from that specified by pseudotime.direction at which the logistic should reach
asymptote. (default: \code{40L}; range: from \code{5L} to \code{200L})}

\item{n.per.tip}{integer; Number of walks to do per tip. (default: \code{25000L};
range: from \code{100L} to \code{1000000L})}

\item{root.visits}{integer; Number of steps to take that visit a root.cell
before stopping. (default: \code{1L}; range: from \code{1L} to \code{5L})}

\item{max.steps}{integer; Number of walks to do per tip. (default: \code{25000L};
range: from \code{100L} to \code{1000000L})}

\item{n.subsample}{integer; Number of subsamplings to perform for calculating
stability. (default: \code{10L}; range: from \code{2L} to \code{100L})}

\item{divergence.method}{discrete; Distance metric to use for determining
transition probabilities. (default: \code{"ks"}; values: {\code{"ks"}, \code{"preference"}})}

\item{cells.per.pseudotime.bin}{integer; Approximate number of cells to assign
to each pseudotime bin for branchpoint finding. (default: \code{80L}; range: from
\code{10L} to \code{1000L})}

\item{bins.per.pseudotime.window}{integer; Width of moving window in pseudotime
used for branchpoint finding, in terms of bins. (default: \code{5L}; range: from
\code{2L} to \code{20L})}

\item{p.thresh}{numeric; P-value threshold to use in determining whether
visitation is significantly different from pairs of tips (default: \code{0.01};
range: from \code{1e-05} to \code{1L})}
}
\value{
A TI method wrapper to be used together with
\code{\link[dynwrap:infer_trajectories]{infer_trajectory}}
}
\description{
Will generate a trajectory using
\href{https://doi.org/10.1126/science.aar3131}{URD}.

This method was wrapped inside a
\href{https://github.com/dynverse/ti_urd}{container}.
The original code of this method is available
\href{https://github.com/farrellja/URD}{here}.
}
\references{
Farrell, J.A., Wang, Y., Riesenfeld, S.J., Shekhar, K., Regev, A.,
Schier, A.F., 2018. Single-cell reconstruction of developmental trajectories
during zebrafish embryogenesis. Science 360, eaar3131.
}
