% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_wanderlust.R
\name{ti_wanderlust}
\alias{ti_wanderlust}
\title{Inferring a trajectory inference using Wanderlust}
\usage{
ti_wanderlust(normalise = TRUE, knn = 25L,
  n_diffusion_components = 3L, n_pca_components = 30L, k = 25L,
  num_waypoints = 250L, epsilon = 1L,
  config = dynwrap::container_config())
}
\arguments{
\item{normalise}{logical;}

\item{knn}{integer; K-nearest neighbours for diffusion (default: \code{25L}; range:
from \code{15L} to \code{100L})}

\item{n_diffusion_components}{integer; Number of diffusion components (default:
\code{3L}; range: from \code{3L} to \code{20L})}

\item{n_pca_components}{integer; Number of pca components (default: \code{30L};
range: from \code{15L} to \code{100L})}

\item{k}{integer; K parameter (default: \code{25L}; range: from \code{15L} to \code{100L})}

\item{num_waypoints}{integer; Number of waypoints (default: \code{250L}; range: from
\code{100L} to \code{500L})}

\item{epsilon}{numeric; Epsilon (default: \code{1L}; range: from \code{0.1} to \code{5L})}

\item{config}{A container config. See \code{\link[=container_config]{container_config()}} for more information.}
}
\value{
A TI method wrapper to be used together with
\code{\link[dynwrap:infer_trajectories]{infer_trajectory}}
}
\description{
Will generate a trajectory using
\href{https://doi.org/10.1016/j.cell.2014.04.005}{Wanderlust}.

This method was wrapped inside a
\href{https://github.com/dynverse/dynmethods/tree/master/containers/wishbone}{container}.
The original code of this method is available
\href{https://github.com/ManuSetty/wishbone}{here}.
}
\references{
Bendall, S.C., Davis, K.L., Amir, E.D., Tadmor, M.D., Simonds,
E.F., Chen, T.J., Shenfeld, D.K., Nolan, G.P., Pe’er, D., 2014. Single-Cell
Trajectory Detection Uncovers Progression and Regulatory Coordination in Human
B Cell Development. Cell 157, 714–725.
}
