% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_wanderlust.R
\name{ti_wanderlust}
\alias{ti_wanderlust}
\title{Inferring a trajectory inference using Wanderlust}
\usage{
ti_wanderlust(normalise = TRUE, knn = 25L, n_diffusion_components = 3L,
  n_pca_components = 30L, k = 25L, num_waypoints = 250L, epsilon = 1L,
  branch = FALSE, run_environment = NULL)
}
\arguments{
\item{normalise}{logical;}

\item{knn}{integer; K-nearest neighbours for diffusion (default: \code{25L}; range:
from \code{15L} to \code{100L})}

\item{n_diffusion_components}{integer; Number of diffusion components (default:
\code{3L}; range: from \code{3L} to \code{20L})}

\item{n_pca_components}{integer; Number of pca components (default: \code{30L};
range: from \code{15L} to \code{100L})}

\item{k}{integer; K parameter (default: \code{25L}; range: from \code{15L} to \code{100L})}

\item{num_waypoints}{integer; Number of waypoints (default: \code{250L}; range: from
\code{100L} to \code{500L})}

\item{epsilon}{numeric; Epsilon (default: \code{1L}; range: from \code{0.1} to \code{5L})}

\item{branch}{logical; Whether to allow a single bifurcation within the
trajectory (wishbone versus wanderlust)}

\item{run_environment}{In which environment to run the method, can be \code{"docker"} or \code{"singularity"}.}
}
\value{
A TI method wrapper to be used together with
\code{\link[dynwrap:infer_trajectories]{infer_trajectory}}
}
\description{
Will generate a trajectory using
\href{https://doi.org/10.1038/nbt.3569}{Wanderlust}.

This method was wrapped inside a
\href{https://github.com/dynverse/dynmethods/tree/master/containers/wishbone}{container}.
The original code of this method is available
\href{https://github.com/ManuSetty/wishbone}{here}.
}
\references{
Setty, M., Tadmor, M.D., Reich-Zeliger, S., Angel, O., Salame,
T.M., Kathail, P., Choi, K., Bendall, S., Friedman, N., Pe’er, D., 2016.
Wishbone identifies bifurcating developmental trajectories from single-cell
data. Nature Biotechnology 34, 637–645.
}
