% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_wanderlust.R
\name{ti_wanderlust}
\alias{ti_wanderlust}
\title{Inferring a trajectory inference using Wanderlust}
\usage{
ti_wanderlust(normalise = TRUE, knn = 25L,
  n_diffusion_components = 3L, n_pca_components = 30L, k = 25L,
  num_waypoints = 250L, epsilon = 1L)
}
\arguments{
\item{normalise}{Parameter; , Default: TRUE, Format: logical.}

\item{knn}{Parameter; K-nearest neighbours for diffusion, Domain: U(15, 100),
Default: 25, Format: integer.}

\item{n_diffusion_components}{Parameter; Number of diffusion components,
Domain: U(3, 20), Default: 3, Format: integer.}

\item{n_pca_components}{Parameter; Number of pca components, Domain: U(15,
100), Default: 30, Format: integer.}

\item{k}{Parameter; K parameter, Domain: U(15, 100), Default: 25, Format:
integer.}

\item{num_waypoints}{Parameter; Number of waypoints, Domain: U(100, 500),
Default: 250, Format: integer.}

\item{epsilon}{Parameter; Epsilon, Domain: U(0.1, 5), Default: 1, Format:
numeric.}
}
\value{
A TI method wrapper to be used together with
\code{\link[dynwrap:infer_trajectories]{infer_trajectory}}
}
\description{
Will generate a trajectory using
\href{https://doi.org/10.1016/j.cell.2014.04.005}{Wanderlust}.

This method was wrapped inside a
\href{https://github.com/dynverse/ti_wanderlust}{container}.
The original code of this method is available
\href{https://github.com/ManuSetty/wishbone}{here}.
}
\references{
Bendall, S.C., Davis, K.L., Amir, E.D., Tadmor, M.D., Simonds,
E.F., Chen, T.J., Shenfeld, D.K., Nolan, G.P., Pe’er, D., 2014. Single-Cell
Trajectory Detection Uncovers Progression and Regulatory Coordination in Human
B Cell Development. Cell 157, 714–725.
}
