% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_waterfall.R, R/ti_container.R
\name{ti_waterfall}
\alias{ti_waterfall}
\alias{ti_waterfall}
\title{Inferring trajectories with Waterfall}
\usage{
ti_waterfall(docker = "stevedore" \%in\% rownames(installed.packages()),
  num_clusters = 10)

ti_waterfall(docker = "stevedore" \%in\% rownames(installed.packages()),
  num_clusters = 10)
}
\arguments{
\item{docker}{Whether to use the docker container or the R wrapper}

\item{num_clusters}{Number of clusters}
}
\value{
The trajectory model
}
\description{
Inferring trajectories with Waterfall

Will generate a trajectory using \href{https://doi.org/10.1016/j.stem.2015.07.013}{Waterfall}. This method was wrapped inside a \href{https://github.com/dynverse/dynmethods/tree/master/containers/waterfall}{container}.
}
\details{
This methods was first wrapped inside R, see \link{ti_waterfall}

The original code of this method is available \href{http://www.cell.com/cms/attachment/2038326541/2052521637/mmc9.zip}{here}.

The method is described in: \href{https://doi.org/10.1016/j.stem.2015.07.013}{Shin, J., Berg, D.A., Zhu, Y., Shin, J.Y., Song, J., Bonaguidi, M.A., Enikolopov, G., Nauen, D.W., Christian, K.M., Ming, G., Song, H., 2015. Single-Cell RNA-Seq with Waterfall Reveals Molecular Cascades underlying Adult Neurogenesis. Cell Stem Cell 17, 360–372.}
}
