% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_waterfall.R
\name{ti_waterfall}
\alias{ti_waterfall}
\title{Waterfall}
\usage{
ti_waterfall(num_clusters = 10L)
}
\arguments{
\item{num_clusters}{. Domain: U(2, 20). Default: 10. Format: integer.}
}
\value{
A TI method wrapper to be used together with
\code{\link[dynwrap:infer_trajectories]{infer_trajectory}}
}
\description{
Will generate a trajectory using
\href{https://doi.org/10.1016/j.stem.2015.07.013}{Waterfall}.

This method was wrapped inside a
\href{https://github.com/dynverse/ti_waterfall}{container}.
The original code of this method is available
\href{http://www.cell.com/cms/attachment/2038326541/2052521637/mmc9.zip}{here}.
}
\references{
Shin, J., Berg, D.A., Zhu, Y., Shin, J.Y., Song, J., Bonaguidi,
M.A., Enikolopov, G., Nauen, D.W., Christian, K.M., Ming, G., Song, H., 2015.
Single-Cell RNA-Seq with Waterfall Reveals Molecular Cascades underlying Adult
Neurogenesis. Cell Stem Cell 17, 360–372.
}
\keyword{method}
