% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_wishbone.R
\name{ti_wishbone}
\alias{ti_wishbone}
\title{Wishbone}
\usage{
ti_wishbone(normalise = TRUE, knn = 25L, n_diffusion_components = 3L,
  n_pca_components = 30L, k = 25L, num_waypoints = 250L,
  epsilon = 1L)
}
\arguments{
\item{normalise}{. Default: TRUE. Format: logical.}

\item{knn}{K-nearest neighbours for diffusion. Domain: U(15, 100). Default: 25.
Format: integer.}

\item{n_diffusion_components}{Number of diffusion components. Domain: U(3, 20).
Default: 3. Format: integer.}

\item{n_pca_components}{Number of pca components. Domain: U(15, 100). Default:
30. Format: integer.}

\item{k}{K parameter. Domain: U(15, 100). Default: 25. Format: integer.}

\item{num_waypoints}{Number of waypoints. Domain: U(100, 500). Default: 250.
Format: integer.}

\item{epsilon}{Epsilon. Domain: U(0.1, 5). Default: 1. Format: numeric.}
}
\value{
A TI method wrapper to be used together with
\code{\link[dynwrap:infer_trajectories]{infer_trajectory}}
}
\description{
Will generate a trajectory using \href{https://doi.org/10.1038/nbt.3569}{Wishbone}.

This method was wrapped inside a
\href{https://github.com/dynverse/ti_wishbone}{container}.
The original code of this method is available
\href{https://github.com/ManuSetty/wishbone}{here}.
}
\references{
Setty, M., Tadmor, M.D., Reich-Zeliger, S., Angel, O., Salame,
T.M., Kathail, P., Choi, K., Bendall, S., Friedman, N., Pe’er, D., 2016.
Wishbone identifies bifurcating developmental trajectories from single-cell
data. Nature Biotechnology 34, 637–645.
}
\keyword{method}
