% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_wishbone.R
\name{ti_wishbone}
\alias{ti_wishbone}
\title{Inferring a trajectory inference using \href{https://doi.org/10.1038/nbt.3569}{wishbone}}
\usage{
ti_wishbone(normalise = TRUE, knn = 15L, n_diffusion_components = 2L,
  n_pca_components = 15L, k = 15L, num_waypoints = 250L, epsilon = 1L,
  branch = TRUE)
}
\arguments{
\item{normalise}{\cr}

\item{knn}{K-nearest neighbours for diffusion \cr
integer; default: 15L; possible values between 2 and 100}

\item{n_diffusion_components}{Number of diffusion components \cr
integer; default: 2L; possible values between 2 and 20}

\item{n_pca_components}{Number of pca components \cr
integer; default: 15L; possible values between 2 and 30}

\item{k}{K parameter \cr
integer; default: 15L; possible values between 2 and 100}

\item{num_waypoints}{Number of waypoints \cr
integer; default: 250L; possible values between 2 and 500}

\item{epsilon}{Epsilon \cr
numeric; default: 1L; possible values between 0.1 and 10}

\item{branch}{Whether to allow a single bifurcation within the trajectory (wishbone versus wanderlust) \cr}
}
\value{
The trajectory model
}
\description{
Will generate a trajectory using \href{https://doi.org/10.1038/nbt.3569}{wishbone}. This method was wrapped inside a \href{https://github.com/dynverse/dynmethods/tree/master/containers/wishbone}{container}.
}
\details{
The original code of this method is available \href{https://github.com/ManuSetty/wishbone}{here}.

The method is described in: \href{https://doi.org/10.1038/nbt.3569}{Setty, M., Tadmor, M.D., Reich-Zeliger, S., Angel, O., Salame, T.M., Kathail, P., Choi, K., Bendall, S., Friedman, N., Pe’er, D., 2016. Wishbone identifies bifurcating developmental trajectories from single-cell data. Nature Biotechnology 34, 637–645.}
}
